addWGSR

Function to add the WHO Growth Reference z-scores to a data frame

addWGSR(data, sex, firstPart, secondPart, thirdPart = NA, index = NA,
  standing = NULL, output = paste(index, "z", sep = ""), digits = 2)

Arguments

data
A survey dataset as a data.frame object

sex
Name of variable specifying the sex of the subject. This must be coded as 1 = male; 2 = female.

Give a quoted variable name as in (e.g.) "sex".

firstPart
Name of variable specifying:

Weight (kg) for BMI/A, W/A, W/H, or W/L

Head circumference (cm) for HC/A

Height (cm) for BMI/A for H/A

Length (cm) for L/A

MUAC (cm) for MUAC/A

Sub-scapular skinfold (mm) for SSF/A

Triceps skinfold (mm) for TSF/A

Give a quoted variable name as in (e.g.) "weight".

Be careful with units (weight in kg; height, length, head circumference, and MUAC in cm, skinfolds in mm).

secondPart
Name of variable specifying:

Age (days) for H/A, HC/A, L/A, MUAC/A, SSF/A, or TSF/A

Height (cm) BMI/A or W/H

Lenght (cm) for W/L

Give a quoted variable name as in (e.g.) "age".

Be careful with units (age in days; height and length in cm).

thirdPart
Name of varible specifying age (in days) for BMI/A

Give a quoted variable name as in (e.g.) "age".

Be careful with units (age in days).

index
The index to be calculated and added to data. One of:

bfa

BMI for age

hca

Head circumgerence for age

hfa

Height for age

lfa

Length for age

mfa

MUAC for age

ssa

Sub-scapular skinfold for age

tsa

Triceps skinfold for age

wfa

Weight for age

wfh

Weight for height

wfl

Weight for length

Give a quoted index name as in (e.g.) "wfh".

standing
Variable specifying how stature was measured. If NULL then age (for "hfa" or "lfa") or height rules (for "wfh" or "wfl") will be applied.

This must be coded as 1 = Standing; 2 = Supine; 3 = Unknown. All other values will be recoded to 3 = Unknown.

Give a quoted variable name as in (e.g.) "measured" or a single value (e.g. "measured = 1"). If no value (or NULL) is specified then height and age rules will be applied.

output
The name of the column containing the specified index to be added to the dataset. This ia an optional parameter. If you do not specify a value for output then the added column will take the name of the specified index with a "z" appended.

digits
The number of decimal places for output. Defaults to 2 d.p.

Value

A data.frame of the survey dataset with the calculated z-scores added.

Examples

# Calculate weight-for-height (wfh) for the anthro3 dataset
addWGSR(data = anthro3,
        sex = "sex",
        firstPart = "weight",
        secondPart = "height",
        index = "wfh")

#> ================================================================================

#>     psu age sex weight height muac oedema  wfhz
#> 1     1  10   1    5.7   64.2  125      2 -2.73
#> 2     1  10   2    5.8   64.4  121      2 -2.04
#> 3     1   9   2    6.5   62.2  139      2  0.13
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2 -3.44
#> 6     1  12   2    6.6   69.4  126      2 -2.26
#> 7     1   9   2    7.0   66.7  136      2 -0.71
#> 8     1   7   1    7.1   63.5  139      2  0.34
#> 9     1   9   2    7.1   66.2  144      2 -0.39
#> 10    1  16   2    7.2   69.0  131      2 -1.12
#> 11    1  13   2    7.4   68.2  137      2 -0.57
#> 12    1  11   2    7.4   70.0  132      2 -1.10
#> 13    1  14   2    7.5   64.8  125      2  0.69
#> 14    2   8   2    7.5   65.7  151      2  0.38
#> 15    2   8   1    7.5   69.1  140      2 -1.13
#> 16    2  13   1    7.6   69.2  131      2 -1.00
#> 17    2  13   2    7.6   70.0  125      2 -0.80
#> 18    2  22   2    7.6   76.2  121      2 -2.43
#> 19    2  16   1    7.7   69.6  136      2 -0.97
#> 20    2   8   2    7.8   63.3  144      2  1.64
#> 21    2  11   2    7.8   71.5  134      2 -0.92
#> 22    2  14   2    7.8   76.6  124      2 -2.21
#> 23    2   7   1    8.0   66.8  139      2  0.48
#> 24    2   8   1    8.0   69.3  147      2 -0.40
#> 25    2  24   2    8.2   70.7  125      2 -0.14
#> 26    2  10   1    8.2   71.1  145      2 -0.68
#> 27    3  15   1    8.3   70.4  133      2 -0.31
#> 28    3  14   2    8.3   71.5  137      2 -0.22
#> 29    3  14   2    8.3   72.9  141      2 -0.59
#> 30    3  17   2    8.3   73.5  136      2 -0.74
#> 31    3  15   1    8.3   76.1  133      2 -1.96
#> 32    3  13   2    8.4   71.8  143      2 -0.17
#> 33    3  20   1    8.4   77.8  132      2 -2.22
#> 34    3  29   2    8.5   75.6  127      2 -0.97
#> 35    3  21   2    8.5   75.6  133      2 -0.97
#> 36    3  21   1    8.5   76.7  123      2 -1.81
#> 37    3  17   1    8.5   77.6  137      2 -2.02
#> 38    3  15   2    8.6   74.9  134      2 -0.67
#> 39    3  24   1    8.6   77.8  137      2 -1.92
#> 40    3  12   1    8.7   72.1  156      2 -0.26
#> 41    4  27   2    8.7   75.8  127      2 -0.75
#> 42    4  25   2    8.7   80.8  129      2 -1.90
#> 43    4  25   1    8.8   81.4  125      2 -2.46
#> 44    4  24   2    8.9   74.3  142      2 -0.15
#> 45    4  30   1    9.0   77.6  132      2 -1.30
#> 46    4  22   1    9.0   82.2  135      2 -2.36
#> 47    4  16   1    9.1   76.4  145      2 -0.88
#> 48    4  24   1    9.1   80.0  126      2 -1.70
#> 49    4  17   2    9.2   74.8  144      2  0.11
#> 50    4  15   2    9.2   75.3  145      2 -0.01
#> 51    4  33   1    9.2   80.1  136      2 -1.59
#> 52    4  25   2    9.2   80.9  142      2 -1.26
#> 53    5  16   1    9.4   75.0  148      2 -0.13
#> 54    5  29   2    9.4   81.9  138      2 -1.26
#> 55    5  16   1    9.5   73.2  151      2  0.48
#> 56    5  17   1    9.6   73.5  146      2  0.52
#> 57    5  29   2    9.6   84.6  134      2 -1.70
#> 58    5  16   1    9.7   74.2  148      2  0.45
#> 59    5  20   1    9.7   80.4  142      2 -0.99
#> 60    5  34   1    9.8   80.2  138      2 -0.82
#> 61    5  21   2    9.8   80.9  145      2 -0.54
#> 62    5  20   1    9.8   82.2  139      2 -1.28
#> 63    5  29   1    9.9   78.6  142      2 -0.35
#> 64    5  25   2    9.9   82.9  138      2 -0.90
#> 65    5  15   1   10.0   75.3  157      2  0.53
#> 66    6  39   1   10.0   81.4  133      2 -0.84
#> 67    6   9   1   10.1   72.6  155      2  1.37
#> 68    6  21   1   10.1   76.7  153      2  0.31
#> 69    6  17   2   10.1   79.1  143      2  0.21
#> 70    6  25   1   10.1   79.5  140      2 -0.30
#> 71    6  13   2   10.2   74.6  157      2  1.27
#> 72    6  25   2   10.3   81.3  141      2 -0.07
#> 73    6  31   2   10.3   85.1  130      2 -1.00
#> 74    6  25   2   10.4   81.3  138      2  0.04
#> 75    6  37   1   10.6   76.2  153      2  1.00
#> 76    6  25   2   10.6   80.7  138      2  0.39
#> 77    6  26   1   10.6   83.1  142      2 -0.52
#> 78    6  36   2   10.6   84.2  135      2 -0.44
#> 79    6  37   2   10.7   84.2  150      2 -0.34
#> 80    7  34   2   10.7   85.7  146      2 -0.71
#> 81    7  47   1   10.7   89.8  131      2 -2.34
#> 82    7  12   1   10.8   82.0  147      2 -0.04
#> 83    7  25   2   10.8   83.6  141      2 -0.08
#> 84    7  31   1   10.8   83.8  158      2 -0.47
#> 85    7  34   1   10.8   84.4  144      2 -0.62
#> 86    7  27   2   10.9   79.0  153      2  1.06
#> 87    7  23   2   10.9   85.9  145      2 -0.55
#> 88    7  19   1   11.0   75.4  168      2  1.62
#> 89    7  26   2   11.0   83.2  145      2  0.22
#> 90    7  54   1   11.0   85.9  139      2 -0.79
#> 91    7  20   1   11.1   79.9  156      2  0.74
#> 92    7  23   1   11.1   84.7  146      2 -0.36
#> 93    7  43   2   11.1   86.0  142      2 -0.37
#> 94    8  31   1   11.1   87.1  143      2 -1.18
#> 95    8  40   2   11.2   95.2  137      2 -2.54
#> 96    8  18   2   11.3   75.5  161      2  2.15
#> 97    8  41   1   11.3   89.4  142      2 -1.53
#> 98    8  40   1   11.3   90.7  148      2 -1.84
#> 99    8  33   2   11.3   91.2  139      2 -1.56
#> 100   8  11   1   11.4   78.4  171      2  1.36
#> 101   8  25   1   11.4   82.2  163      2  0.57
#> 102   8  27   2   11.4   86.8  155      2 -0.26
#> 103   8  38   1   11.4   89.4  141      2 -1.41
#> 104   8  17   2   11.5   80.6  158      2  1.30
#> 105   8  31   2   11.5   85.4  153      2  0.18
#> 106   9  45   1   11.5   86.2  143      2 -0.31
#> 107   9  53   1   11.6   81.4  153      2  0.95
#> 108   9  23   2   11.6   86.3  143      2  0.05
#> 109   9  37   1   11.6   86.3  149      2 -0.23
#> 110   9  38   1   11.6   89.9  146      2 -1.31
#> 111   9  40   1   11.7   93.0  140      2 -1.91
#> 112   9  36   1   11.8   87.1  146      2 -0.40
#> 113   9  29   1   11.9   84.3  166      2  0.59
#> 114   9  26   2   12.0   81.2  164      2  1.62
#> 115   9  33   1   12.0   86.0  154      2  0.26
#> 116   9  36   1   12.0   87.5  150      2 -0.29
#> 117   9  55   2   12.0   96.2  144      2 -1.93
#> 118  10  16   1   12.1   82.3  157      2  1.25
#> 119  10  35   2   12.1   86.6  147      2  0.45
#> 120  10  43   1   12.1   91.7  151      2 -1.18
#> 121  10  38   2   12.2   80.9  175      2  1.86
#> 122  10  37   1   12.2   85.5  162      2  0.59
#> 123  10  41   2   12.2   87.6  155      2  0.14
#> 124  10  45   2   12.2   90.1  149      2 -0.43
#> 125  10  46   2   12.2   99.6  146      2 -2.46
#> 126  10  31   1   12.3   88.0  167      2 -0.11
#> 127  10  44   1   12.3   88.1  151      2 -0.13
#> 128  10  47   2   12.3   94.9  133      2 -1.37
#> 129  10  29   1   12.4   91.8  141      2 -0.89
#> 130  10  46   2   12.4   99.6  140      2 -2.26
#> 131  10  34   1   12.6   87.0  156      2  0.43
#> 132  10  38   2   12.7   85.7  152      2  1.20
#> 133  11  32   1   12.8   87.6  161      2  0.48
#> 134  11  36   1   12.8   90.4  150      2 -0.17
#> 135  11  39   1   12.8   92.7  152      2 -0.69
#> 136  11  42   1   12.8   94.0  146      2 -0.97
#> 137  11  38   2   12.9   83.4  174      2  1.90
#> 138  11  26   2   12.9   86.2  159      2  1.25
#> 139  11  35   1   12.9   90.9  154      2 -0.19
#> 140  11  41   1   12.9   91.2  160      2 -0.26
#> 141  11  32   1   12.9   91.4  148      2 -0.30
#> 142  11  49   1   13.0   93.1  152      2 -0.58
#> 143  11  30   1   13.1   84.3  161      2  1.72
#> 144  11  32   1   13.1   85.1  136      2  1.53
#> 145  11  35   1   13.1   87.5  158      2  0.78
#> 146  11  34   1   13.1   88.6  162      2  0.52
#> 147  11  45   2   13.1   91.6  150      2  0.05
#> 148  12  29   1   13.2   90.8  158      2  0.12
#> 149  12  47   2   13.2   91.4  146      2  0.18
#> 150  12  37   2   13.2   93.8  155      2 -0.33
#> 151  12  35   1   13.3   96.5  142      2 -1.02
#> 152  12  15   2   13.4   82.4  157      2  2.51
#> 153  12  38   2   13.4   93.0  152      2  0.01
#> 154  12  37   2   13.4   94.1  149      2 -0.22
#> 155  12  53   2   13.4  100.3  145      2 -1.50
#> 156  12  37   2   13.5   84.4  168      2  2.14
#> 157  12  43   1   13.5   96.1  151      2 -0.74
#> 158  12  52   1   13.5   97.0  144      2 -0.94
#> 159  12  45   2   13.5   98.0  148      2 -0.94
#> 160  12  39   2   13.5   99.1  145      2 -1.16
#> 161  12  47   2   13.6   97.0  146      2 -0.65
#> 162  13  26   2   13.7   85.7  171      2  2.00
#> 163  13  24   1   13.8   80.9  172      2  3.02
#> 164  13  38   2   13.8   93.3  153      2  0.27
#> 165  13  57   1   13.8   94.0  149      2 -0.02
#> 166  13  42   2   13.8  103.4  144      2 -1.84
#> 167  13  42   2   13.9   92.3  175      2  0.56
#> 168  13  52   1   13.9   95.5  151      2 -0.25
#> 169  13  52   1   13.9   95.7  156      2 -0.29
#> 170  13  57   2   13.9   97.2  153      2 -0.44
#> 171  13  52   2   13.9   97.8  141      2 -0.56
#> 172  13  52   2   13.9   97.9  155      2 -0.58
#> 173  13  49   2   13.9   98.2  150      2 -0.64
#> 174  13  58   2   14.0   96.6  154      2 -0.24
#> 175  13  50   1   14.0   97.9  158      2 -0.68
#> 176  13  39   2   14.1   93.7  155      2  0.43
#> 177  14  49   2   14.1   96.4  140      2 -0.12
#> 178  14  50   2   14.1   97.1  148      2 -0.26
#> 179  14  59   2   14.1   97.3  152      2 -0.30
#> 180  14  52   2   14.1  102.4  156      2 -1.37
#> 181  14  41   1   14.2   96.7  146      2 -0.24
#> 182  14  53   1   14.2   97.7  151      2 -0.46
#> 183  14  49   2   14.2   98.1  156      2 -0.38
#> 184  14  35   2   14.3   89.8  165      2  1.39
#> 185  14  50   1   14.3   93.7  152      2  0.48
#> 186  14  41   1   14.4   94.8  158      2  0.34
#> 187  14  42   1   14.5   95.1  161      2  0.36
#> 188  14  58   1   14.5  104.9  130      2 -1.79
#> 189  14  49   1   14.7   96.6  152      2  0.21
#> 190  14  52   2   14.7   96.7  154      2  0.28
#> 191  14  51   2   14.7   99.2  155      2 -0.22
#> 192  14  49   1   14.7   99.6  164      2 -0.45
#> 193  15  42   2   15.0   93.5  172      2  1.14
#> 194  15  28   2   15.1   91.5  164      2  1.62
#> 195  15  41   2   15.2   95.9  165      2  0.81
#> 196  15  53   1   15.3   95.0  166      2  1.03
#> 197  15  58   1   15.3  100.7  155      2 -0.20
#> 198  15  48   1   15.3  103.5  143      2 -0.82
#> 199  15  54   1   15.5   97.9  159      2  0.57
#> 200  15  58   1   15.5  101.3  166      2 -0.18
#> 201  15  52   2   15.6  100.0  164      2  0.27
#> 202  15  50   1   15.7   99.6  153      2  0.36
#> 203  15  53   2   15.7  103.1  158      2 -0.32
#> 204  15  43   1   15.8   91.1  179      2  2.22
#> 205  15  57   1   15.9  103.8  161      2 -0.42
#> 206  15  55   2   16.3   95.4  169      2  1.65
#> 207  16  54   2   15.3  102.0  156      2 -0.37
#> 208  16  44   1   16.3   96.2  173      2  1.54
#> 209  16  47   1   16.3  102.0  166      2  0.27
#> 210  16  57   1   16.3  108.4  138      2 -1.13
#> 211  16  52   1   16.4  103.9  152      2 -0.07
#> 212  16  52   2   16.6   97.8  144      2  1.37
#> 213  16  56   1   16.6  103.9  148      2  0.07
#> 214  16  55   1   16.7  106.3  154      2 -0.39
#> 215  16  52   1   17.0  101.3  163      2  0.93
#> 216  16  50   1   17.3  101.8  168      2  1.02
#> 217  16  53   1   17.5  102.2  168      2  1.06
#> 218  16  42   1   17.7  100.9  145      2  1.48
#> 219  16  48   2   17.8  111.3  176      2 -0.77
#> 220  16  53   1   17.9   98.7  171      2  2.10
#> 221  16  50   1   18.1  106.4  166      2  0.53

# Calculate weight-for-age (wfa) for the anthro3 dataset
addWGSR(data = anthro3,
        sex = "sex",
        firstPart = "weight",
        secondPart = "age",
        index = "wfa")

#> ================================================================================

#>     psu age sex weight height muac oedema  wfaz
#> 1     1  10   1    5.7   64.2  125      2  3.45
#> 2     1  10   2    5.8   64.4  121      2  3.95
#> 3     1   9   2    6.5   62.2  139      2  5.12
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2  3.82
#> 6     1  12   2    6.6   69.4  126      2  5.01
#> 7     1   9   2    7.0   66.7  136      2  5.91
#> 8     1   7   1    7.1   63.5  139      2  5.84
#> 9     1   9   2    7.1   66.2  144      2  6.06
#> 10    1  16   2    7.2   69.0  131      2  5.53
#> 11    1  13   2    7.4   68.2  137      2  6.13
#> 12    1  11   2    7.4   70.0  132      2  6.33
#> 13    1  14   2    7.5   64.8  125      2  6.18
#> 14    2   8   2    7.5   65.7  151      2  6.79
#> 15    2   8   1    7.5   69.1  140      2  6.36
#> 16    2  13   1    7.6   69.2  131      2  6.02
#> 17    2  13   2    7.6   70.0  125      2  6.43
#> 18    2  22   2    7.6   76.2  121      2  5.52
#> 19    2  16   1    7.7   69.6  136      2  5.84
#> 20    2   8   2    7.8   63.3  144      2  7.26
#> 21    2  11   2    7.8   71.5  134      2  6.95
#> 22    2  14   2    7.8   76.6  124      2  6.63
#> 23    2   7   1    8.0   66.8  139      2  7.23
#> 24    2   8   1    8.0   69.3  147      2  7.13
#> 25    2  24   2    8.2   70.7  125      2  6.17
#> 26    2  10   1    8.2   71.1  145      2  7.23
#> 27    3  15   1    8.3   70.4  133      2  6.82
#> 28    3  14   2    8.3   71.5  137      2  7.38
#> 29    3  14   2    8.3   72.9  141      2  7.38
#> 30    3  17   2    8.3   73.5  136      2  7.04
#> 31    3  15   1    8.3   76.1  133      2  6.82
#> 32    3  13   2    8.4   71.8  143      2  7.64
#> 33    3  20   1    8.4   77.8  132      2  6.41
#> 34    3  29   2    8.5   75.6  127      2  6.12
#> 35    3  21   2    8.5   75.6  133      2  6.89
#> 36    3  21   1    8.5   76.7  123      2  6.44
#> 37    3  17   1    8.5   77.6  137      2  6.88
#> 38    3  15   2    8.6   74.9  134      2  7.71
#> 39    3  24   1    8.6   77.8  137      2  6.26
#> 40    3  12   1    8.7   72.1  156      2  7.76
#> 41    4  27   2    8.7   75.8  127      2  6.57
#> 42    4  25   2    8.7   80.8  129      2  6.76
#> 43    4  25   1    8.8   81.4  125      2  6.43
#> 44    4  24   2    8.9   74.3  142      2  7.14
#> 45    4  30   1    9.0   77.6  132      2  6.21
#> 46    4  22   1    9.0   82.2  135      2  7.02
#> 47    4  16   1    9.1   76.4  145      2  7.86
#> 48    4  24   1    9.1   80.0  126      2  6.94
#> 49    4  17   2    9.2   74.8  144      2  8.35
#> 50    4  15   2    9.2   75.3  145      2  8.60
#> 51    4  33   1    9.2   80.1  136      2  6.20
#> 52    4  25   2    9.2   80.9  142      2  7.45
#> 53    5  16   1    9.4   75.0  148      2  8.30
#> 54    5  29   2    9.4   81.9  138      2  7.32
#> 55    5  16   1    9.5   73.2  151      2  8.44
#> 56    5  17   1    9.6   73.5  146      2  8.46
#> 57    5  29   2    9.6   84.6  134      2  7.58
#> 58    5  16   1    9.7   74.2  148      2  8.73
#> 59    5  20   1    9.7   80.4  142      2  8.22
#> 60    5  34   1    9.8   80.2  138      2  6.88
#> 61    5  21   2    9.8   80.9  145      2  8.73
#> 62    5  20   1    9.8   82.2  139      2  8.36
#> 63    5  29   1    9.9   78.6  142      2  7.48
#> 64    5  25   2    9.9   82.9  138      2  8.41
#> 65    5  15   1   10.0   75.3  157      2  9.30
#> 66    6  39   1   10.0   81.4  133      2  6.71
#> 67    6   9   1   10.1   72.6  155      2 10.22
#> 68    6  21   1   10.1   76.7  153      2  8.66
#> 69    6  17   2   10.1   79.1  143      2  9.67
#> 70    6  25   1   10.1   79.5  140      2  8.18
#> 71    6  13   2   10.2   74.6  157      2 10.36
#> 72    6  25   2   10.3   81.3  141      2  8.95
#> 73    6  31   2   10.3   85.1  130      2  8.31
#> 74    6  25   2   10.4   81.3  138      2  9.09
#> 75    6  37   1   10.6   76.2  153      2  7.62
#> 76    6  25   2   10.6   80.7  138      2  9.37
#> 77    6  26   1   10.6   83.1  142      2  8.74
#> 78    6  36   2   10.6   84.2  135      2  8.22
#> 79    6  37   2   10.7   84.2  150      2  8.26
#> 80    7  34   2   10.7   85.7  146      2  8.54
#> 81    7  47   1   10.7   89.8  131      2  6.96
#> 82    7  12   1   10.8   82.0  147      2 10.88
#> 83    7  25   2   10.8   83.6  141      2  9.64
#> 84    7  31   1   10.8   83.8  158      2  8.45
#> 85    7  34   1   10.8   84.4  144      2  8.15
#> 86    7  27   2   10.9   79.0  153      2  9.54
#> 87    7  23   2   10.9   85.9  145      2 10.02
#> 88    7  19   1   11.0   75.4  168      2 10.18
#> 89    7  26   2   11.0   83.2  145      2  9.79
#> 90    7  54   1   11.0   85.9  139      2  6.85
#> 91    7  20   1   11.1   79.9  156      2 10.18
#> 92    7  23   1   11.1   84.7  146      2  9.78
#> 93    7  43   2   11.1   86.0  142      2  8.24
#> 94    8  31   1   11.1   87.1  143      2  8.84
#> 95    8  40   2   11.2   95.2  137      2  8.62
#> 96    8  18   2   11.3   75.5  161      2 11.27
#> 97    8  41   1   11.3   89.4  142      2  8.15
#> 98    8  40   1   11.3   90.7  148      2  8.23
#> 99    8  33   2   11.3   91.2  139      2  9.41
#> 100   8  11   1   11.4   78.4  171      2 11.92
#> 101   8  25   1   11.4   82.2  163      2  9.94
#> 102   8  27   2   11.4   86.8  155      2 10.22
#> 103   8  38   1   11.4   89.4  141      2  8.53
#> 104   8  17   2   11.5   80.6  158      2 11.71
#> 105   8  31   2   11.5   85.4  153      2  9.89
#> 106   9  45   1   11.5   86.2  143      2  8.07
#> 107   9  53   1   11.6   81.4  153      2  7.61
#> 108   9  23   2   11.6   86.3  143      2 11.00
#> 109   9  37   1   11.6   86.3  149      2  8.88
#> 110   9  38   1   11.6   89.9  146      2  8.78
#> 111   9  40   1   11.7   93.0  140      2  8.73
#> 112   9  36   1   11.8   87.1  146      2  9.22
#> 113   9  29   1   11.9   84.3  166      2 10.11
#> 114   9  26   2   12.0   81.2  164      2 11.15
#> 115   9  33   1   12.0   86.0  154      2  9.79
#> 116   9  36   1   12.0   87.5  150      2  9.48
#> 117   9  55   2   12.0   96.2  144      2  8.41
#> 118  10  16   1   12.1   82.3  157      2 12.19
#> 119  10  35   2   12.1   86.6  147      2 10.24
#> 120  10  43   1   12.1   91.7  151      2  8.96
#> 121  10  38   2   12.2   80.9  175      2 10.06
#> 122  10  37   1   12.2   85.5  162      2  9.63
#> 123  10  41   2   12.2   87.6  155      2  9.77
#> 124  10  45   2   12.2   90.1  149      2  9.42
#> 125  10  46   2   12.2   99.6  146      2  9.33
#> 126  10  31   1   12.3   88.0  167      2 10.40
#> 127  10  44   1   12.3   88.1  151      2  9.12
#> 128  10  47   2   12.3   94.9  133      2  9.37
#> 129  10  29   1   12.4   91.8  141      2 10.77
#> 130  10  46   2   12.4   99.6  140      2  9.57
#> 131  10  34   1   12.6   87.0  156      2 10.45
#> 132  10  38   2   12.7   85.7  152      2 10.69
#> 133  11  32   1   12.8   87.6  161      2 10.93
#> 134  11  36   1   12.8   90.4  150      2 10.49
#> 135  11  39   1   12.8   92.7  152      2 10.18
#> 136  11  42   1   12.8   94.0  146      2  9.90
#> 137  11  38   2   12.9   83.4  174      2 10.94
#> 138  11  26   2   12.9   86.2  159      2 12.38
#> 139  11  35   1   12.9   90.9  154      2 10.72
#> 140  11  41   1   12.9   91.2  160      2 10.12
#> 141  11  32   1   12.9   91.4  148      2 11.06
#> 142  11  49   1   13.0   93.1  152      2  9.55
#> 143  11  30   1   13.1   84.3  161      2 11.56
#> 144  11  32   1   13.1   85.1  136      2 11.32
#> 145  11  35   1   13.1   87.5  158      2 10.97
#> 146  11  34   1   13.1   88.6  162      2 11.08
#> 147  11  45   2   13.1   91.6  150      2 10.51
#> 148  12  29   1   13.2   90.8  158      2 11.82
#> 149  12  47   2   13.2   91.4  146      2 10.45
#> 150  12  37   2   13.2   93.8  155      2 11.42
#> 151  12  35   1   13.3   96.5  142      2 11.23
#> 152  12  15   2   13.4   82.4  157      2 14.84
#> 153  12  38   2   13.4   93.0  152      2 11.57
#> 154  12  37   2   13.4   94.1  149      2 11.68
#> 155  12  53   2   13.4  100.3  145      2 10.18
#> 156  12  37   2   13.5   84.4  168      2 11.80
#> 157  12  43   1   13.5   96.1  151      2 10.66
#> 158  12  52   1   13.5   97.0  144      2  9.91
#> 159  12  45   2   13.5   98.0  148      2 10.99
#> 160  12  39   2   13.5   99.1  145      2 11.59
#> 161  12  47   2   13.6   97.0  146      2 10.93
#> 162  13  26   2   13.7   85.7  171      2 13.47
#> 163  13  24   1   13.8   80.9  172      2 13.32
#> 164  13  38   2   13.8   93.3  153      2 12.07
#> 165  13  57   1   13.8   94.0  149      2  9.90
#> 166  13  42   2   13.8  103.4  144      2 11.65
#> 167  13  42   2   13.9   92.3  175      2 11.77
#> 168  13  52   1   13.9   95.5  151      2 10.38
#> 169  13  52   1   13.9   95.7  156      2 10.38
#> 170  13  57   2   13.9   97.2  153      2 10.45
#> 171  13  52   2   13.9   97.8  141      2 10.85
#> 172  13  52   2   13.9   97.9  155      2 10.85
#> 173  13  49   2   13.9   98.2  150      2 11.10
#> 174  13  58   2   14.0   96.6  154      2 10.49
#> 175  13  50   1   14.0   97.9  158      2 10.66
#> 176  13  39   2   14.1   93.7  155      2 12.34
#> 177  14  49   2   14.1   96.4  140      2 11.34
#> 178  14  50   2   14.1   97.1  148      2 11.25
#> 179  14  59   2   14.1   97.3  152      2 10.52
#> 180  14  52   2   14.1  102.4  156      2 11.08
#> 181  14  41   1   14.2   96.7  146      2 11.71
#> 182  14  53   1   14.2   97.7  151      2 10.66
#> 183  14  49   2   14.2   98.1  156      2 11.46
#> 184  14  35   2   14.3   89.8  165      2 13.05
#> 185  14  50   1   14.3   93.7  152      2 11.01
#> 186  14  41   1   14.4   94.8  158      2 11.96
#> 187  14  42   1   14.5   95.1  161      2 11.98
#> 188  14  58   1   14.5  104.9  130      2 10.64
#> 189  14  49   1   14.7   96.6  152      2 11.57
#> 190  14  52   2   14.7   96.7  154      2 11.78
#> 191  14  51   2   14.7   99.2  155      2 11.87
#> 192  14  49   1   14.7   99.6  164      2 11.57
#> 193  15  42   2   15.0   93.5  172      2 13.12
#> 194  15  28   2   15.1   91.5  164      2 15.06
#> 195  15  41   2   15.2   95.9  165      2 13.48
#> 196  15  53   1   15.3   95.0  166      2 11.94
#> 197  15  58   1   15.3  100.7  155      2 11.56
#> 198  15  48   1   15.3  103.5  143      2 12.37
#> 199  15  54   1   15.5   97.9  159      2 12.10
#> 200  15  58   1   15.5  101.3  166      2 11.79
#> 201  15  52   2   15.6  100.0  164      2 12.83
#> 202  15  50   1   15.7   99.6  153      2 12.67
#> 203  15  53   2   15.7  103.1  158      2 12.86
#> 204  15  43   1   15.8   91.1  179      2 13.47
#> 205  15  57   1   15.9  103.8  161      2 12.32
#> 206  15  55   2   16.3   95.4  169      2 13.37
#> 207  16  54   2   15.3  102.0  156      2 12.30
#> 208  16  44   1   16.3   96.2  173      2 13.97
#> 209  16  47   1   16.3  102.0  166      2 13.66
#> 210  16  57   1   16.3  108.4  138      2 12.78
#> 211  16  52   1   16.4  103.9  152      2 13.32
#> 212  16  52   2   16.6   97.8  144      2 14.00
#> 213  16  56   1   16.6  103.9  148      2 13.21
#> 214  16  55   1   16.7  106.3  154      2 13.41
#> 215  16  52   1   17.0  101.3  163      2 14.02
#> 216  16  50   1   17.3  101.8  168      2 14.56
#> 217  16  53   1   17.5  102.2  168      2 14.52
#> 218  16  42   1   17.7  100.9  145      2 15.90
#> 219  16  48   2   17.8  111.3  176      2 15.84
#> 220  16  53   1   17.9   98.7  171      2 14.98
#> 221  16  50   1   18.1  106.4  166      2 15.51

# Calculate height-for-age (hfa) for the anthro3 dataset
addWGSR(data = anthro3,
        sex = "sex",
        firstPart = "height",
        secondPart = "age",
        index = "hfa")

#> ================================================================================

#>     psu age sex weight height muac oedema  hfaz
#> 1     1  10   1    5.7   64.2  125      2  6.58
#> 2     1  10   2    5.8   64.4  121      2  7.17
#> 3     1   9   2    6.5   62.2  139      2  6.11
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2 10.35
#> 6     1  12   2    6.6   69.4  126      2  9.60
#> 7     1   9   2    7.0   66.7  136      2  8.49
#> 8     1   7   1    7.1   63.5  139      2  6.51
#> 9     1   9   2    7.1   66.2  144      2  8.22
#> 10    1  16   2    7.2   69.0  131      2  9.01
#> 11    1  13   2    7.4   68.2  137      2  8.87
#> 12    1  11   2    7.4   70.0  132      2 10.02
#> 13    1  14   2    7.5   64.8  125      2  6.98
#> 14    2   8   2    7.5   65.7  151      2  8.06
#> 15    2   8   1    7.5   69.1  140      2  9.35
#> 16    2  13   1    7.6   69.2  131      2  8.91
#> 17    2  13   2    7.6   70.0  125      2  9.82
#> 18    2  22   2    7.6   76.2  121      2 12.24
#> 19    2  16   1    7.7   69.6  136      2  8.85
#> 20    2   8   2    7.8   63.3  144      2  6.79
#> 21    2  11   2    7.8   71.5  134      2 10.82
#> 22    2  14   2    7.8   76.6  124      2 13.18
#> 23    2   7   1    8.0   66.8  139      2  8.24
#> 24    2   8   1    8.0   69.3  147      2  9.46
#> 25    2  24   2    8.2   70.7  125      2  9.22
#> 26    2  10   1    8.2   71.1  145      2 10.20
#> 27    3  15   1    8.3   70.4  133      2  9.36
#> 28    3  14   2    8.3   71.5  137      2 10.50
#> 29    3  14   2    8.3   72.9  141      2 11.24
#> 30    3  17   2    8.3   73.5  136      2 11.28
#> 31    3  15   1    8.3   76.1  133      2 12.33
#> 32    3  13   2    8.4   71.8  143      2 10.77
#> 33    3  20   1    8.4   77.8  132      2 12.76
#> 34    3  29   2    8.5   75.6  127      2 11.33
#> 35    3  21   2    8.5   75.6  133      2 12.02
#> 36    3  21   1    8.5   76.7  123      2 12.10
#> 37    3  17   1    8.5   77.6  137      2 12.93
#> 38    3  15   2    8.6   74.9  134      2 12.20
#> 39    3  24   1    8.6   77.8  137      2 12.40
#> 40    3  12   1    8.7   72.1  156      2 10.53
#> 41    4  27   2    8.7   75.8  127      2 11.60
#> 42    4  25   2    8.7   80.8  129      2 14.35
#> 43    4  25   1    8.8   81.4  125      2 14.18
#> 44    4  24   2    8.9   74.3  142      2 11.08
#> 45    4  30   1    9.0   77.6  132      2 11.79
#> 46    4  22   1    9.0   82.2  135      2 14.86
#> 47    4  16   1    9.1   76.4  145      2 12.40
#> 48    4  24   1    9.1   80.0  126      2 13.54
#> 49    4  17   2    9.2   74.8  144      2 11.96
#> 50    4  15   2    9.2   75.3  145      2 12.41
#> 51    4  33   1    9.2   80.1  136      2 12.82
#> 52    4  25   2    9.2   80.9  142      2 14.40
#> 53    5  16   1    9.4   75.0  148      2 11.67
#> 54    5  29   2    9.4   81.9  138      2 14.56
#> 55    5  16   1    9.5   73.2  151      2 10.73
#> 56    5  17   1    9.6   73.5  146      2 10.79
#> 57    5  29   2    9.6   84.6  134      2 15.95
#> 58    5  16   1    9.7   74.2  148      2 11.25
#> 59    5  20   1    9.7   80.4  142      2 14.11
#> 60    5  34   1    9.8   80.2  138      2 12.79
#> 61    5  21   2    9.8   80.9  145      2 14.77
#> 62    5  20   1    9.8   82.2  139      2 15.04
#> 63    5  29   1    9.9   78.6  142      2 12.38
#> 64    5  25   2    9.9   82.9  138      2 15.43
#> 65    5  15   1   10.0   75.3  157      2 11.92
#> 66    6  39   1   10.0   81.4  133      2 13.01
#> 67    6   9   1   10.1   72.6  155      2 11.09
#> 68    6  21   1   10.1   76.7  153      2 12.10
#> 69    6  17   2   10.1   79.1  143      2 14.21
#> 70    6  25   1   10.1   79.5  140      2 13.19
#> 71    6  13   2   10.2   74.6  157      2 12.24
#> 72    6  25   2   10.3   81.3  141      2 14.61
#> 73    6  31   2   10.3   85.1  130      2 16.03
#> 74    6  25   2   10.4   81.3  138      2 14.61
#> 75    6  37   1   10.6   76.2  153      2 10.51
#> 76    6  25   2   10.6   80.7  138      2 14.30
#> 77    6  26   1   10.6   83.1  142      2 14.96
#> 78    6  36   2   10.6   84.2  135      2 15.14
#> 79    6  37   2   10.7   84.2  150      2 15.06
#> 80    7  34   2   10.7   85.7  146      2 16.07
#> 81    7  47   1   10.7   89.8  131      2 16.66
#> 82    7  12   1   10.8   82.0  147      2 15.72
#> 83    7  25   2   10.8   83.6  141      2 15.80
#> 84    7  31   1   10.8   83.8  158      2 14.89
#> 85    7  34   1   10.8   84.4  144      2 14.94
#> 86    7  27   2   10.9   79.0  153      2 13.25
#> 87    7  23   2   10.9   85.9  145      2 17.17
#> 88    7  19   1   11.0   75.4  168      2 11.60
#> 89    7  26   2   11.0   83.2  145      2 15.50
#> 90    7  54   1   11.0   85.9  139      2 14.19
#> 91    7  20   1   11.1   79.9  156      2 13.85
#> 92    7  23   1   11.1   84.7  146      2 16.06
#> 93    7  43   2   11.1   86.0  142      2 15.49
#> 94    8  31   1   11.1   87.1  143      2 16.58
#> 95    8  40   2   11.2   95.2  137      2 20.37
#> 96    8  18   2   11.3   75.5  161      2 12.23
#> 97    8  41   1   11.3   89.4  142      2 16.92
#> 98    8  40   1   11.3   90.7  148      2 17.67
#> 99    8  33   2   11.3   91.2  139      2 18.96
#> 100   8  11   1   11.4   78.4  171      2 13.93
#> 101   8  25   1   11.4   82.2  163      2 14.59
#> 102   8  27   2   11.4   86.8  155      2 17.26
#> 103   8  38   1   11.4   89.4  141      2 17.16
#> 104   8  17   2   11.5   80.6  158      2 14.99
#> 105   8  31   2   11.5   85.4  153      2 16.18
#> 106   9  45   1   11.5   86.2  143      2 14.99
#> 107   9  53   1   11.6   81.4  153      2 12.00
#> 108   9  23   2   11.6   86.3  143      2 17.38
#> 109   9  37   1   11.6   86.3  149      2 15.67
#> 110   9  38   1   11.6   89.9  146      2 17.42
#> 111   9  40   1   11.7   93.0  140      2 18.84
#> 112   9  36   1   11.8   87.1  146      2 16.16
#> 113   9  29   1   11.9   84.3  166      2 15.32
#> 114   9  26   2   12.0   81.2  164      2 14.47
#> 115   9  33   1   12.0   86.0  154      2 15.85
#> 116   9  36   1   12.0   87.5  150      2 16.36
#> 117   9  55   2   12.0   96.2  144      2 19.65
#> 118  10  16   1   12.1   82.3  157      2 15.48
#> 119  10  35   2   12.1   86.6  147      2 16.44
#> 120  10  43   1   12.1   91.7  151      2 17.94
#> 121  10  38   2   12.2   80.9  175      2 13.31
#> 122  10  37   1   12.2   85.5  162      2 15.26
#> 123  10  41   2   12.2   87.6  155      2 16.45
#> 124  10  45   2   12.2   90.1  149      2 17.39
#> 125  10  46   2   12.2   99.6  146      2 22.06
#> 126  10  31   1   12.3   88.0  167      2 17.05
#> 127  10  44   1   12.3   88.1  151      2 16.03
#> 128  10  47   2   12.3   94.9  133      2 19.63
#> 129  10  29   1   12.4   91.8  141      2 19.17
#> 130  10  46   2   12.4   99.6  140      2 22.06
#> 131  10  34   1   12.6   87.0  156      2 16.27
#> 132  10  38   2   12.7   85.7  152      2 15.74
#> 133  11  32   1   12.8   87.6  161      2 16.75
#> 134  11  36   1   12.8   90.4  150      2 17.84
#> 135  11  39   1   12.8   92.7  152      2 18.77
#> 136  11  42   1   12.8   94.0  146      2 19.18
#> 137  11  38   2   12.9   83.4  174      2 14.57
#> 138  11  26   2   12.9   86.2  159      2 17.04
#> 139  11  35   1   12.9   90.9  154      2 18.18
#> 140  11  41   1   12.9   91.2  160      2 17.84
#> 141  11  32   1   12.9   91.4  148      2 18.70
#> 142  11  49   1   13.0   93.1  152      2 18.18
#> 143  11  30   1   13.1   84.3  161      2 15.23
#> 144  11  32   1   13.1   85.1  136      2 15.47
#> 145  11  35   1   13.1   87.5  158      2 16.45
#> 146  11  34   1   13.1   88.6  162      2 17.09
#> 147  11  45   2   13.1   91.6  150      2 18.14
#> 148  12  29   1   13.2   90.8  158      2 18.66
#> 149  12  47   2   13.2   91.4  146      2 17.88
#> 150  12  37   2   13.2   93.8  155      2 19.92
#> 151  12  35   1   13.3   96.5  142      2 21.05
#> 152  12  15   2   13.4   82.4  157      2 16.13
#> 153  12  38   2   13.4   93.0  152      2 19.43
#> 154  12  37   2   13.4   94.1  149      2 20.07
#> 155  12  53   2   13.4  100.3  145      2 21.84
#> 156  12  37   2   13.5   84.4  168      2 15.16
#> 157  12  43   1   13.5   96.1  151      2 20.17
#> 158  12  52   1   13.5   97.0  144      2 19.92
#> 159  12  45   2   13.5   98.0  148      2 21.35
#> 160  12  39   2   13.5   99.1  145      2 22.42
#> 161  12  47   2   13.6   97.0  146      2 20.68
#> 162  13  26   2   13.7   85.7  171      2 16.79
#> 163  13  24   1   13.8   80.9  172      2 14.00
#> 164  13  38   2   13.8   93.3  153      2 19.58
#> 165  13  57   1   13.8   94.0  149      2 18.05
#> 166  13  42   2   13.8  103.4  144      2 24.32
#> 167  13  42   2   13.9   92.3  175      2 18.74
#> 168  13  52   1   13.9   95.5  151      2 19.17
#> 169  13  52   1   13.9   95.7  156      2 19.27
#> 170  13  57   2   13.9   97.2  153      2 19.99
#> 171  13  52   2   13.9   97.8  141      2 20.68
#> 172  13  52   2   13.9   97.9  155      2 20.73
#> 173  13  49   2   13.9   98.2  150      2 21.11
#> 174  13  58   2   14.0   96.6  154      2 19.62
#> 175  13  50   1   14.0   97.9  158      2 20.53
#> 176  13  39   2   14.1   93.7  155      2 19.70
#> 177  14  49   2   14.1   96.4  140      2 20.22
#> 178  14  50   2   14.1   97.1  148      2 20.48
#> 179  14  59   2   14.1   97.3  152      2 19.89
#> 180  14  52   2   14.1  102.4  156      2 22.96
#> 181  14  41   1   14.2   96.7  146      2 20.63
#> 182  14  53   1   14.2   97.7  151      2 20.20
#> 183  14  49   2   14.2   98.1  156      2 21.06
#> 184  14  35   2   14.3   89.8  165      2 18.07
#> 185  14  50   1   14.3   93.7  152      2 18.41
#> 186  14  41   1   14.4   94.8  158      2 19.67
#> 187  14  42   1   14.5   95.1  161      2 19.74
#> 188  14  58   1   14.5  104.9  130      2 23.45
#> 189  14  49   1   14.7   96.6  152      2 19.95
#> 190  14  52   2   14.7   96.7  154      2 20.13
#> 191  14  51   2   14.7   99.2  155      2 21.45
#> 192  14  49   1   14.7   99.6  164      2 21.46
#> 193  15  42   2   15.0   93.5  172      2 19.34
#> 194  15  28   2   15.1   91.5  164      2 19.58
#> 195  15  41   2   15.2   95.9  165      2 20.63
#> 196  15  53   1   15.3   95.0  166      2 18.84
#> 197  15  58   1   15.3  100.7  155      2 21.34
#> 198  15  48   1   15.3  103.5  143      2 23.51
#> 199  15  54   1   15.5   97.9  159      2 20.23
#> 200  15  58   1   15.5  101.3  166      2 21.64
#> 201  15  52   2   15.6  100.0  164      2 21.77
#> 202  15  50   1   15.7   99.6  153      2 21.38
#> 203  15  53   2   15.7  103.1  158      2 23.22
#> 204  15  43   1   15.8   91.1  179      2 17.63
#> 205  15  57   1   15.9  103.8  161      2 22.97
#> 206  15  55   2   16.3   95.4  169      2 19.25
#> 207  16  54   2   15.3  102.0  156      2 22.60
#> 208  16  44   1   16.3   96.2  173      2 20.14
#> 209  16  47   1   16.3  102.0  166      2 22.83
#> 210  16  57   1   16.3  108.4  138      2 25.28
#> 211  16  52   1   16.4  103.9  152      2 23.40
#> 212  16  52   2   16.6   97.8  144      2 20.68
#> 213  16  56   1   16.6  103.9  148      2 23.09
#> 214  16  55   1   16.7  106.3  154      2 24.37
#> 215  16  52   1   17.0  101.3  163      2 22.09
#> 216  16  50   1   17.3  101.8  168      2 22.49
#> 217  16  53   1   17.5  102.2  168      2 22.46
#> 218  16  42   1   17.7  100.9  145      2 22.69
#> 219  16  48   2   17.8  111.3  176      2 27.73
#> 220  16  53   1   17.9   98.7  171      2 20.70
#> 221  16  50   1   18.1  106.4  166      2 24.81