Calculate CDC Growth Reference z-score for given anthropometric measurement/s.

The first function, getCDC(), is usually called by the addCDC() function but could be used as a stand-alone calculator for getting z-score for a given anthropometric measurement.

getCDC(sex, firstPart, secondPart, thirdPart = NA, index = NA, standing = NA)

addCDC(
  data,
  sex,
  firstPart,
  secondPart,
  thirdPart = NA,
  index = NA,
  standing = NULL,
  output = paste(index, "z", sep = ""),
  digits = 2
)

Arguments

sex
A vector specifying the sex of the subject or a character value for the name of the variable in data specifying the sex of the subject. This must be coded as 1 = male and 2 = female. Give a quoted variable name as in (e.g.) "sex".

firstPart
A vector or a character value for the name of the variable in data specifying:

Weight (kg) for BMI/A, W/A, W/H, or W/L

Head circumference (cm) for HC/A

Height (cm) for BMI/A for H/A

Length (cm) for L/A

If name of variable in data, give a quoted variable name as in (e.g.) "weight". Be careful with units (weight in kg; height, length, head circumference in cm).

secondPart
A vector or a character value for the name of variable in data specifying:

Age (days) for H/A, HC/A, or L/A

Height (cm) BMI/A or W/H

Length (cm) for W/L

If name of variable in data, give a quoted variable name as in (e.g.) "age". Be careful with units (age in months; height and length in cm).

thirdPart
A vector or a character value for the name of variable in data specifying age (in months) for BMI/A. If name of variable in data, give a quoted variable name as in (e.g.) "age". Be careful with units (age in months).

index
The index to be calculated. One of:

Index	Definition

bfa	BMI for age

hca	Head circumference for age

hfa	Height for age

lfa	Length for age

wfa	Weight for age

wfh	Weight for height

wfl	Weight for length

Give a quoted index name as in (e.g.) "wfh".

standing
A vector or a character value for name of variable in data specifying how stature was measured. If NULL then age (for "hfa" or "lfa") or height rules (for "wfh" or "wfl") will be applied. This must be coded as 1 = Standing; 2 = Supine; 3 = Unknown. All other values will be recoded to 3 = Unknown. If name of variable in data, give a quoted variable name as in (e.g.) "measured" or a single value (e.g. "measured = 1). If no value (or NULL) is specified then age rules will be applied.

data
A survey dataset as a data.frame object

output
The name of the column containing the specified index to be added to the dataset. This is an optional parameter. If you do not specify a value for output then the added column will take the name of the specified index with a "z" appended.

digits
The number of decimal places for output. Defaults to 2 d.p.

Value

A data.frame of the survey dataset with the calculated z-scores added.

Details

addCDC() adds the CDC Growth Reference z-scores to a data frame of anthropometric data for weight, height or length, head circumference, and body mass index (BMI)

Examples

# Given a male child 10 months old with a weight of 5.7 kgs, height of 64.2
# cms, and MUAC of 125 mm:
#
# Calculate weight-for-height
getCDC(sex = 1,
       firstPart = 5.7,
       secondPart = 64.2,
       index = "wfh",
       standing = 3)

#> [1] NA

# calculate weight-for-age
getCDC(sex = 1,
       firstPart = 5.7,
       secondPart = 10,
       index = "wfa",
       standing = 3)

#> [1] -4.929997

# calculate height-for-age
getCDC(sex = 1,
       firstPart = 64.2,
       secondPart = 10,
       index = "hfa",
       standing = 3)

#> [1] -3.58536

# Calculate weight-for-height (wfh) for the anthro3 dataset
addCDC(data = anthro3,
       sex = "sex",
       firstPart = "weight",
       secondPart = "height",
       index = "wfh")

#> ================================================================================

#>     psu age sex weight height muac oedema  wfhz
#> 1     1  10   1    5.7   64.2  125      2    NA
#> 2     1  10   2    5.8   64.4  121      2    NA
#> 3     1   9   2    6.5   62.2  139      2    NA
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2    NA
#> 6     1  12   2    6.6   69.4  126      2    NA
#> 7     1   9   2    7.0   66.7  136      2    NA
#> 8     1   7   1    7.1   63.5  139      2    NA
#> 9     1   9   2    7.1   66.2  144      2    NA
#> 10    1  16   2    7.2   69.0  131      2    NA
#> 11    1  13   2    7.4   68.2  137      2    NA
#> 12    1  11   2    7.4   70.0  132      2    NA
#> 13    1  14   2    7.5   64.8  125      2    NA
#> 14    2   8   2    7.5   65.7  151      2    NA
#> 15    2   8   1    7.5   69.1  140      2    NA
#> 16    2  13   1    7.6   69.2  131      2    NA
#> 17    2  13   2    7.6   70.0  125      2    NA
#> 18    2  22   2    7.6   76.2  121      2    NA
#> 19    2  16   1    7.7   69.6  136      2    NA
#> 20    2   8   2    7.8   63.3  144      2    NA
#> 21    2  11   2    7.8   71.5  134      2    NA
#> 22    2  14   2    7.8   76.6  124      2    NA
#> 23    2   7   1    8.0   66.8  139      2    NA
#> 24    2   8   1    8.0   69.3  147      2    NA
#> 25    2  24   2    8.2   70.7  125      2    NA
#> 26    2  10   1    8.2   71.1  145      2    NA
#> 27    3  15   1    8.3   70.4  133      2    NA
#> 28    3  14   2    8.3   71.5  137      2    NA
#> 29    3  14   2    8.3   72.9  141      2    NA
#> 30    3  17   2    8.3   73.5  136      2    NA
#> 31    3  15   1    8.3   76.1  133      2    NA
#> 32    3  13   2    8.4   71.8  143      2    NA
#> 33    3  20   1    8.4   77.8  132      2 -3.17
#> 34    3  29   2    8.5   75.6  127      2    NA
#> 35    3  21   2    8.5   75.6  133      2    NA
#> 36    3  21   1    8.5   76.7  123      2    NA
#> 37    3  17   1    8.5   77.6  137      2 -2.91
#> 38    3  15   2    8.6   74.9  134      2    NA
#> 39    3  24   1    8.6   77.8  137      2 -2.80
#> 40    3  12   1    8.7   72.1  156      2    NA
#> 41    4  27   2    8.7   75.8  127      2    NA
#> 42    4  25   2    8.7   80.8  129      2 -3.16
#> 43    4  25   1    8.8   81.4  125      2 -3.62
#> 44    4  24   2    8.9   74.3  142      2    NA
#> 45    4  30   1    9.0   77.6  132      2 -2.03
#> 46    4  22   1    9.0   82.2  135      2 -3.51
#> 47    4  16   1    9.1   76.4  145      2    NA
#> 48    4  24   1    9.1   80.0  126      2 -2.64
#> 49    4  17   2    9.2   74.8  144      2    NA
#> 50    4  15   2    9.2   75.3  145      2    NA
#> 51    4  33   1    9.2   80.1  136      2 -2.51
#> 52    4  25   2    9.2   80.9  142      2 -2.36
#> 53    5  16   1    9.4   75.0  148      2    NA
#> 54    5  29   2    9.4   81.9  138      2 -2.35
#> 55    5  16   1    9.5   73.2  151      2    NA
#> 56    5  17   1    9.6   73.5  146      2    NA
#> 57    5  29   2    9.6   84.6  134      2 -2.85
#> 58    5  16   1    9.7   74.2  148      2    NA
#> 59    5  20   1    9.7   80.4  142      2 -1.80
#> 60    5  34   1    9.8   80.2  138      2 -1.60
#> 61    5  21   2    9.8   80.9  145      2 -1.47
#> 62    5  20   1    9.8   82.2  139      2 -2.19
#> 63    5  29   1    9.9   78.6  142      2 -0.97
#> 64    5  25   2    9.9   82.9  138      2 -1.90
#> 65    5  15   1   10.0   75.3  157      2    NA
#> 66    6  39   1   10.0   81.4  133      2 -1.66
#> 67    6   9   1   10.1   72.6  155      2    NA
#> 68    6  21   1   10.1   76.7  153      2    NA
#> 69    6  17   2   10.1   79.1  143      2 -0.56
#> 70    6  25   1   10.1   79.5  140      2 -0.96
#> 71    6  13   2   10.2   74.6  157      2    NA
#> 72    6  25   2   10.3   81.3  141      2 -0.90
#> 73    6  31   2   10.3   85.1  130      2 -1.95
#> 74    6  25   2   10.4   81.3  138      2 -0.78
#> 75    6  37   1   10.6   76.2  153      2    NA
#> 76    6  25   2   10.6   80.7  138      2 -0.37
#> 77    6  26   1   10.6   83.1  142      2 -1.30
#> 78    6  36   2   10.6   84.2  135      2 -1.30
#> 79    6  37   2   10.7   84.2  150      2 -1.17
#> 80    7  34   2   10.7   85.7  146      2 -1.57
#> 81    7  47   1   10.7   89.8  131      2 -3.01
#> 82    7  12   1   10.8   82.0  147      2 -0.74
#> 83    7  25   2   10.8   83.6  141      2 -0.88
#> 84    7  31   1   10.8   83.8  158      2 -1.23
#> 85    7  34   1   10.8   84.4  144      2 -1.39
#> 86    7  27   2   10.9   79.0  153      2  0.42
#> 87    7  23   2   10.9   85.9  145      2 -1.36
#> 88    7  19   1   11.0   75.4  168      2    NA
#> 89    7  26   2   11.0   83.2  145      2 -0.53
#> 90    7  54   1   11.0   85.9  139      2 -1.53
#> 91    7  20   1   11.1   79.9  156      2  0.19
#> 92    7  23   1   11.1   84.7  146      2 -1.08
#> 93    7  43   2   11.1   86.0  142      2 -1.13
#> 94    8  31   1   11.1   87.1  143      2 -1.71
#> 95    8  40   2   11.2   95.2  137      2 -3.74
#> 96    8  18   2   11.3   75.5  161      2    NA
#> 97    8  41   1   11.3   89.4  142      2 -2.06
#> 98    8  40   1   11.3   90.7  148      2 -2.41
#> 99    8  33   2   11.3   91.2  139      2 -2.27
#> 100   8  11   1   11.4   78.4  171      2  0.92
#> 101   8  25   1   11.4   82.2  163      2 -0.06
#> 102   8  27   2   11.4   86.8  155      2 -0.97
#> 103   8  38   1   11.4   89.4  141      2 -1.93
#> 104   8  17   2   11.5   80.6  158      2  0.67
#> 105   8  31   2   11.5   85.4  153      2 -0.50
#> 106   9  45   1   11.5   86.2  143      2 -0.97
#> 107   9  53   1   11.6   81.4  153      2  0.37
#> 108   9  23   2   11.6   86.3  143      2 -0.61
#> 109   9  37   1   11.6   86.3  149      2 -0.88
#> 110   9  38   1   11.6   89.9  146      2 -1.80
#> 111   9  40   1   11.7   93.0  140      2 -2.50
#> 112   9  36   1   11.8   87.1  146      2 -0.84
#> 113   9  29   1   11.9   84.3  166      2 -0.03
#> 114   9  26   2   12.0   81.2  164      2  1.03
#> 115   9  33   1   12.0   86.0  154      2 -0.34
#> 116   9  36   1   12.0   87.5  150      2 -0.71
#> 117   9  55   2   12.0   96.2  144      2 -2.78
#> 118  10  16   1   12.1   82.3  157      2  0.68
#> 119  10  35   2   12.1   86.6  147      2 -0.12
#> 120  10  43   1   12.1   91.7  151      2 -1.65
#> 121  10  38   2   12.2   80.9  175      2  1.29
#> 122  10  37   1   12.2   85.5  162      2  0.00
#> 123  10  41   2   12.2   87.6  155      2 -0.25
#> 124  10  45   2   12.2   90.1  149      2 -0.85
#> 125  10  46   2   12.2   99.6  146      2 -3.64
#> 126  10  31   1   12.3   88.0  167      2 -0.50
#> 127  10  44   1   12.3   88.1  151      2 -0.52
#> 128  10  47   2   12.3   94.9  133      2 -1.97
#> 129  10  29   1   12.4   91.8  141      2 -1.32
#> 130  10  46   2   12.4   99.6  140      2 -3.31
#> 131  10  34   1   12.6   87.0  156      2  0.07
#> 132  10  38   2   12.7   85.7  152      2  0.69
#> 133  11  32   1   12.8   87.6  161      2  0.13
#> 134  11  36   1   12.8   90.4  150      2 -0.53
#> 135  11  39   1   12.8   92.7  152      2 -1.09
#> 136  11  42   1   12.8   94.0  146      2 -1.41
#> 137  11  38   2   12.9   83.4  174      2  1.39
#> 138  11  26   2   12.9   86.2  159      2  0.77
#> 139  11  35   1   12.9   90.9  154      2 -0.54
#> 140  11  41   1   12.9   91.2  160      2 -0.62
#> 141  11  32   1   12.9   91.4  148      2 -0.66
#> 142  11  49   1   13.0   93.1  152      2 -0.96
#> 143  11  30   1   13.1   84.3  161      2  1.19
#> 144  11  32   1   13.1   85.1  136      2  1.01
#> 145  11  35   1   13.1   87.5  158      2  0.45
#> 146  11  34   1   13.1   88.6  162      2  0.20
#> 147  11  45   2   13.1   91.6  150      2 -0.24
#> 148  12  29   1   13.2   90.8  158      2 -0.21
#> 149  12  47   2   13.2   91.4  146      2 -0.10
#> 150  12  37   2   13.2   93.8  155      2 -0.64
#> 151  12  35   1   13.3   96.5  142      2 -1.47
#> 152  12  15   2   13.4   82.4  157      2  2.03
#> 153  12  38   2   13.4   93.0  152      2 -0.26
#> 154  12  37   2   13.4   94.1  149      2 -0.51
#> 155  12  53   2   13.4  100.3  145      2 -2.09
#> 156  12  37   2   13.5   84.4  168      2  1.69
#> 157  12  43   1   13.5   96.1  151      2 -1.15
#> 158  12  52   1   13.5   97.0  144      2 -1.38
#> 159  12  45   2   13.5   98.0  148      2 -1.34
#> 160  12  39   2   13.5   99.1  145      2 -1.63
#> 161  12  47   2   13.6   97.0  146      2 -0.99
#> 162  13  26   2   13.7   85.7  171      2  1.58
#> 163  13  24   1   13.8   80.9  172      2  2.54
#> 164  13  38   2   13.8   93.3  153      2  0.05
#> 165  13  57   1   13.8   94.0  149      2 -0.35
#> 166  13  42   2   13.8  103.4  144      2 -2.50
#> 167  13  42   2   13.9   92.3  175      2  0.35
#> 168  13  52   1   13.9   95.5  151      2 -0.60
#> 169  13  52   1   13.9   95.7  156      2 -0.65
#> 170  13  57   2   13.9   97.2  153      2 -0.73
#> 171  13  52   2   13.9   97.8  141      2 -0.87
#> 172  13  52   2   13.9   97.9  155      2 -0.89
#> 173  13  49   2   13.9   98.2  150      2 -0.96
#> 174  13  58   2   14.0   96.6  154      2 -0.49
#> 175  13  50   1   14.0   97.9  158      2 -1.07
#> 176  13  39   2   14.1   93.7  155      2  0.22
#> 177  14  49   2   14.1   96.4  140      2 -0.36
#> 178  14  50   2   14.1   97.1  148      2 -0.51
#> 179  14  59   2   14.1   97.3  152      2 -0.56
#> 180  14  52   2   14.1  102.4  156      2 -1.83
#> 181  14  41   1   14.2   96.7  146      2 -0.59
#> 182  14  53   1   14.2   97.7  151      2 -0.82
#> 183  14  49   2   14.2   98.1  156      2 -0.64
#> 184  14  35   2   14.3   89.8  165      2  1.20
#> 185  14  50   1   14.3   93.7  152      2  0.18
#> 186  14  41   1   14.4   94.8  158      2  0.03
#> 187  14  42   1   14.5   95.1  161      2  0.05
#> 188  14  58   1   14.5  104.9  130      2 -2.45
#> 189  14  49   1   14.7   96.6  152      2 -0.10
#> 190  14  52   2   14.7   96.7  154      2  0.10
#> 191  14  51   2   14.7   99.2  155      2 -0.43
#> 192  14  49   1   14.7   99.6  164      2 -0.79
#> 193  15  42   2   15.0   93.5  172      2  0.97
#> 194  15  28   2   15.1   91.5  164      2  1.44
#> 195  15  41   2   15.2   95.9  165      2  0.64
#> 196  15  53   1   15.3   95.0  166      2  0.75
#> 197  15  58   1   15.3  100.7  155      2 -0.49
#> 198  15  48   1   15.3  103.5  143      2 -1.17
#> 199  15  54   1   15.5   97.9  159      2  0.29
#> 200  15  58   1   15.5  101.3  166      2 -0.46
#> 201  15  52   2   15.6  100.0  164      2  0.12
#> 202  15  50   1   15.7   99.6  153      2  0.09
#> 203  15  53   2   15.7  103.1  158      2 -0.46
#> 204  15  43   1   15.8   91.1  179      2  1.94
#> 205  15  57   1   15.9  103.8  161      2 -0.69
#> 206  15  55   2   16.3   95.4  169      2  1.45
#> 207  16  54   2   15.3  102.0  156      2 -0.54
#> 208  16  44   1   16.3   96.2  173      2  1.25
#> 209  16  47   1   16.3  102.0  166      2  0.05
#> 210  16  57   1   16.3  108.4  138      2 -1.51
#> 211  16  52   1   16.4  103.9  152      2 -0.29
#> 212  16  52   2   16.6   97.8  144      2  1.18
#> 213  16  56   1   16.6  103.9  148      2 -0.13
#> 214  16  55   1   16.7  106.3  154      2 -0.60
#> 215  16  52   1   17.0  101.3  163      2  0.70
#> 216  16  50   1   17.3  101.8  168      2  0.80
#> 217  16  53   1   17.5  102.2  168      2  0.84
#> 218  16  42   1   17.7  100.9  145      2  1.22
#> 219  16  48   2   17.8  111.3  176      2 -0.71
#> 220  16  53   1   17.9   98.7  171      2  1.77
#> 221  16  50   1   18.1  106.4  166      2  0.40

# Calculate weight-for-age (wfa) for the anthro3 dataset
addCDC(data = anthro3,
       sex = "sex",
       firstPart = "weight",
       secondPart = "age",
       index = "wfa")

#> ================================================================================

#>     psu age sex weight height muac oedema  wfaz
#> 1     1  10   1    5.7   64.2  125      2 -4.93
#> 2     1  10   2    5.8   64.4  121      2 -3.86
#> 3     1   9   2    6.5   62.2  139      2 -2.39
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2 -7.28
#> 6     1  12   2    6.6   69.4  126      2 -3.49
#> 7     1   9   2    7.0   66.7  136      2 -1.74
#> 8     1   7   1    7.1   63.5  139      2 -1.47
#> 9     1   9   2    7.1   66.2  144      2 -1.61
#> 10    1  16   2    7.2   69.0  131      2 -3.91
#> 11    1  13   2    7.4   68.2  137      2 -2.71
#> 12    1  11   2    7.4   70.0  132      2 -2.03
#> 13    1  14   2    7.5   64.8  125      2 -2.89
#> 14    2   8   2    7.5   65.7  151      2 -0.69
#> 15    2   8   1    7.5   69.1  140      2 -1.48
#> 16    2  13   1    7.6   69.2  131      2 -3.11
#> 17    2  13   2    7.6   70.0  125      2 -2.45
#> 18    2  22   2    7.6   76.2  121      2 -4.73
#> 19    2  16   1    7.7   69.6  136      2 -3.65
#> 20    2   8   2    7.8   63.3  144      2 -0.34
#> 21    2  11   2    7.8   71.5  134      2 -1.54
#> 22    2  14   2    7.8   76.6  124      2 -2.50
#> 23    2   7   1    8.0   66.8  139      2 -0.43
#> 24    2   8   1    8.0   69.3  147      2 -0.90
#> 25    2  24   2    8.2   70.7  125      2 -4.15
#> 26    2  10   1    8.2   71.1  145      2 -1.48
#> 27    3  15   1    8.3   70.4  133      2 -2.72
#> 28    3  14   2    8.3   71.5  137      2 -1.87
#> 29    3  14   2    8.3   72.9  141      2 -1.87
#> 30    3  17   2    8.3   73.5  136      2 -2.63
#> 31    3  15   1    8.3   76.1  133      2 -2.72
#> 32    3  13   2    8.4   71.8  143      2 -1.47
#> 33    3  20   1    8.4   77.8  132      2 -3.44
#> 34    3  29   2    8.5   75.6  127      2 -4.44
#> 35    3  21   2    8.5   75.6  133      2 -3.19
#> 36    3  21   1    8.5   76.7  123      2 -3.46
#> 37    3  17   1    8.5   77.6  137      2 -2.87
#> 38    3  15   2    8.6   74.9  134      2 -1.78
#> 39    3  24   1    8.6   77.8  137      2 -3.73
#> 40    3  12   1    8.7   72.1  156      2 -1.56
#> 41    4  27   2    8.7   75.8  127      2 -3.87
#> 42    4  25   2    8.7   80.8  129      2 -3.58
#> 43    4  25   1    8.8   81.4  125      2 -3.62
#> 44    4  24   2    8.9   74.3  142      2 -3.16
#> 45    4  30   1    9.0   77.6  132      2 -4.01
#> 46    4  22   1    9.0   82.2  135      2 -3.03
#> 47    4  16   1    9.1   76.4  145      2 -2.04
#> 48    4  24   1    9.1   80.0  126      2 -3.17
#> 49    4  17   2    9.2   74.8  144      2 -1.56
#> 50    4  15   2    9.2   75.3  145      2 -1.11
#> 51    4  33   1    9.2   80.1  136      2 -4.15
#> 52    4  25   2    9.2   80.9  142      2 -2.93
#> 53    5  16   1    9.4   75.0  148      2 -1.73
#> 54    5  29   2    9.4   81.9  138      2 -3.20
#> 55    5  16   1    9.5   73.2  151      2 -1.63
#> 56    5  17   1    9.6   73.5  146      2 -1.70
#> 57    5  29   2    9.6   84.6  134      2 -2.95
#> 58    5  16   1    9.7   74.2  148      2 -1.43
#> 59    5  20   1    9.7   80.4  142      2 -2.05
#> 60    5  34   1    9.8   80.2  138      2 -3.58
#> 61    5  21   2    9.8   80.9  145      2 -1.64
#> 62    5  20   1    9.8   82.2  139      2 -1.95
#> 63    5  29   1    9.9   78.6  142      2 -2.90
#> 64    5  25   2    9.9   82.9  138      2 -2.11
#> 65    5  15   1   10.0   75.3  157      2 -0.96
#> 66    6  39   1   10.0   81.4  133      2 -3.97
#> 67    6   9   1   10.1   72.6  155      2  0.75
#> 68    6  21   1   10.1   76.7  153      2 -1.79
#> 69    6  17   2   10.1   79.1  143      2 -0.63
#> 70    6  25   1   10.1   79.5  140      2 -2.26
#> 71    6  13   2   10.2   74.6  157      2  0.37
#> 72    6  25   2   10.3   81.3  141      2 -1.68
#> 73    6  31   2   10.3   85.1  130      2 -2.39
#> 74    6  25   2   10.4   81.3  138      2 -1.58
#> 75    6  37   1   10.6   76.2  153      2 -3.08
#> 76    6  25   2   10.6   80.7  138      2 -1.38
#> 77    6  26   1   10.6   83.1  142      2 -1.90
#> 78    6  36   2   10.6   84.2  135      2 -2.60
#> 79    6  37   2   10.7   84.2  150      2 -2.59
#> 80    7  34   2   10.7   85.7  146      2 -2.29
#> 81    7  47   1   10.7   89.8  131      2 -4.12
#> 82    7  12   1   10.8   82.0  147      2  0.42
#> 83    7  25   2   10.8   83.6  141      2 -1.19
#> 84    7  31   1   10.8   83.8  158      2 -2.23
#> 85    7  34   1   10.8   84.4  144      2 -2.55
#> 86    7  27   2   10.9   79.0  153      2 -1.34
#> 87    7  23   2   10.9   85.9  145      2 -0.84
#> 88    7  19   1   11.0   75.4  168      2 -0.72
#> 89    7  26   2   11.0   83.2  145      2 -1.13
#> 90    7  54   1   11.0   85.9  139      2 -4.56
#> 91    7  20   1   11.1   79.9  156      2 -0.77
#> 92    7  23   1   11.1   84.7  146      2 -1.13
#> 93    7  43   2   11.1   86.0  142      2 -2.80
#> 94    8  31   1   11.1   87.1  143      2 -1.96
#> 95    8  40   2   11.2   95.2  137      2 -2.40
#> 96    8  18   2   11.3   75.5  161      2  0.26
#> 97    8  41   1   11.3   89.4  142      2 -2.82
#> 98    8  40   1   11.3   90.7  148      2 -2.71
#> 99    8  33   2   11.3   91.2  139      2 -1.61
#> 100   8  11   1   11.4   78.4  171      2  1.17
#> 101   8  25   1   11.4   82.2  163      2 -1.09
#> 102   8  27   2   11.4   86.8  155      2 -0.89
#> 103   8  38   1   11.4   89.4  141      2 -2.41
#> 104   8  17   2   11.5   80.6  158      2  0.61
#> 105   8  31   2   11.5   85.4  153      2 -1.23
#> 106   9  45   1   11.5   86.2  143      2 -3.06
#> 107   9  53   1   11.6   81.4  153      2 -3.81
#> 108   9  23   2   11.6   86.3  143      2 -0.24
#> 109   9  37   1   11.6   86.3  149      2 -2.12
#> 110   9  38   1   11.6   89.9  146      2 -2.22
#> 111   9  40   1   11.7   93.0  140      2 -2.34
#> 112   9  36   1   11.8   87.1  146      2 -1.85
#> 113   9  29   1   11.9   84.3  166      2 -1.08
#> 114   9  26   2   12.0   81.2  164      2 -0.28
#> 115   9  33   1   12.0   86.0  154      2 -1.39
#> 116   9  36   1   12.0   87.5  150      2 -1.68
#> 117   9  55   2   12.0   96.2  144      2 -3.19
#> 118  10  16   1   12.1   82.3  157      2  0.61
#> 119  10  35   2   12.1   86.6  147      2 -1.13
#> 120  10  43   1   12.1   91.7  151      2 -2.30
#> 121  10  38   2   12.2   80.9  175      2 -1.34
#> 122  10  37   1   12.2   85.5  162      2 -1.61
#> 123  10  41   2   12.2   87.6  155      2 -1.62
#> 124  10  45   2   12.2   90.1  149      2 -2.00
#> 125  10  46   2   12.2   99.6  146      2 -2.10
#> 126  10  31   1   12.3   88.0  167      2 -0.96
#> 127  10  44   1   12.3   88.1  151      2 -2.23
#> 128  10  47   2   12.3   94.9  133      2 -2.11
#> 129  10  29   1   12.4   91.8  141      2 -0.69
#> 130  10  46   2   12.4   99.6  140      2 -1.94
#> 131  10  34   1   12.6   87.0  156      2 -1.01
#> 132  10  38   2   12.7   85.7  152      2 -0.96
#> 133  11  32   1   12.8   87.6  161      2 -0.67
#> 134  11  36   1   12.8   90.4  150      2 -1.05
#> 135  11  39   1   12.8   92.7  152      2 -1.33
#> 136  11  42   1   12.8   94.0  146      2 -1.62
#> 137  11  38   2   12.9   83.4  174      2 -0.82
#> 138  11  26   2   12.9   86.2  159      2  0.38
#> 139  11  35   1   12.9   90.9  154      2 -0.88
#> 140  11  41   1   12.9   91.2  160      2 -1.45
#> 141  11  32   1   12.9   91.4  148      2 -0.60
#> 142  11  49   1   13.0   93.1  152      2 -2.15
#> 143  11  30   1   13.1   84.3  161      2 -0.27
#> 144  11  32   1   13.1   85.1  136      2 -0.46
#> 145  11  35   1   13.1   87.5  158      2 -0.73
#> 146  11  34   1   13.1   88.6  162      2 -0.64
#> 147  11  45   2   13.1   91.6  150      2 -1.32
#> 148  12  29   1   13.2   90.8  158      2 -0.11
#> 149  12  47   2   13.2   91.4  146      2 -1.43
#> 150  12  37   2   13.2   93.8  155      2 -0.52
#> 151  12  35   1   13.3   96.5  142      2 -0.59
#> 152  12  15   2   13.4   82.4  157      2  2.37
#> 153  12  38   2   13.4   93.0  152      2 -0.48
#> 154  12  37   2   13.4   94.1  149      2 -0.38
#> 155  12  53   2   13.4  100.3  145      2 -1.84
#> 156  12  37   2   13.5   84.4  168      2 -0.32
#> 157  12  43   1   13.5   96.1  151      2 -1.20
#> 158  12  52   1   13.5   97.0  144      2 -2.06
#> 159  12  45   2   13.5   98.0  148      2 -1.04
#> 160  12  39   2   13.5   99.1  145      2 -0.50
#> 161  12  47   2   13.6   97.0  146      2 -1.15
#> 162  13  26   2   13.7   85.7  171      2  0.90
#> 163  13  24   1   13.8   80.9  172      2  0.78
#> 164  13  38   2   13.8   93.3  153      2 -0.22
#> 165  13  57   1   13.8   94.0  149      2 -2.30
#> 166  13  42   2   13.8  103.4  144      2 -0.58
#> 167  13  42   2   13.9   92.3  175      2 -0.52
#> 168  13  52   1   13.9   95.5  151      2 -1.77
#> 169  13  52   1   13.9   95.7  156      2 -1.77
#> 170  13  57   2   13.9   97.2  153      2 -1.86
#> 171  13  52   2   13.9   97.8  141      2 -1.41
#> 172  13  52   2   13.9   97.9  155      2 -1.41
#> 173  13  49   2   13.9   98.2  150      2 -1.14
#> 174  13  58   2   14.0   96.6  154      2 -1.88
#> 175  13  50   1   14.0   97.9  158      2 -1.51
#> 176  13  39   2   14.1   93.7  155      2 -0.13
#> 177  14  49   2   14.1   96.4  140      2 -1.01
#> 178  14  50   2   14.1   97.1  148      2 -1.10
#> 179  14  59   2   14.1   97.3  152      2 -1.91
#> 180  14  52   2   14.1  102.4  156      2 -1.28
#> 181  14  41   1   14.2   96.7  146      2 -0.54
#> 182  14  53   1   14.2   97.7  151      2 -1.65
#> 183  14  49   2   14.2   98.1  156      2 -0.95
#> 184  14  35   2   14.3   89.8  165      2  0.35
#> 185  14  50   1   14.3   93.7  152      2 -1.31
#> 186  14  41   1   14.4   94.8  158      2 -0.41
#> 187  14  42   1   14.5   95.1  161      2 -0.44
#> 188  14  58   1   14.5  104.9  130      2 -1.90
#> 189  14  49   1   14.7   96.6  152      2 -0.96
#> 190  14  52   2   14.7   96.7  154      2 -0.91
#> 191  14  51   2   14.7   99.2  155      2 -0.83
#> 192  14  49   1   14.7   99.6  164      2 -0.96
#> 193  15  42   2   15.0   93.5  172      2  0.11
#> 194  15  28   2   15.1   91.5  164      2  1.47
#> 195  15  41   2   15.2   95.9  165      2  0.30
#> 196  15  53   1   15.3   95.0  166      2 -0.95
#> 197  15  58   1   15.3  100.7  155      2 -1.39
#> 198  15  48   1   15.3  103.5  143      2 -0.51
#> 199  15  54   1   15.5   97.9  159      2 -0.93
#> 200  15  58   1   15.5  101.3  166      2 -1.27
#> 201  15  52   2   15.6  100.0  164      2 -0.42
#> 202  15  50   1   15.7   99.6  153      2 -0.46
#> 203  15  53   2   15.7  103.1  158      2 -0.45
#> 204  15  43   1   15.8   91.1  179      2  0.22
#> 205  15  57   1   15.9  103.8  161      2 -0.96
#> 206  15  55   2   16.3   95.4  169      2 -0.32
#> 207  16  54   2   15.3  102.0  156      2 -0.75
#> 208  16  44   1   16.3   96.2  173      2  0.39
#> 209  16  47   1   16.3  102.0  166      2  0.12
#> 210  16  57   1   16.3  108.4  138      2 -0.74
#> 211  16  52   1   16.4  103.9  152      2 -0.26
#> 212  16  52   2   16.6   97.8  144      2  0.05
#> 213  16  56   1   16.6  103.9  148      2 -0.50
#> 214  16  55   1   16.7  106.3  154      2 -0.37
#> 215  16  52   1   17.0  101.3  163      2  0.03
#> 216  16  50   1   17.3  101.8  168      2  0.35
#> 217  16  53   1   17.5  102.2  168      2  0.18
#> 218  16  42   1   17.7  100.9  145      2  1.23
#> 219  16  48   2   17.8  111.3  176      2  0.85
#> 220  16  53   1   17.9   98.7  171      2  0.36
#> 221  16  50   1   18.1  106.4  166      2  0.70

# Calculate height-for-age (hfa) for the anthro3 dataset
addCDC(data = anthro3,
       sex = "sex",
       firstPart = "height",
       secondPart = "age",
       index = "hfa")

#> ================================================================================

#>     psu age sex weight height muac oedema  hfaz
#> 1     1  10   1    5.7   64.2  125      2 -3.59
#> 2     1  10   2    5.8   64.4  121      2 -2.39
#> 3     1   9   2    6.5   62.2  139      2 -2.72
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2 -3.46
#> 6     1  12   2    6.6   69.4  126      2 -1.49
#> 7     1   9   2    7.0   66.7  136      2 -1.15
#> 8     1   7   1    7.1   63.5  139      2 -2.13
#> 9     1   9   2    7.1   66.2  144      2 -1.33
#> 10    1  16   2    7.2   69.0  131      2 -2.91
#> 11    1  13   2    7.4   68.2  137      2 -2.24
#> 12    1  11   2    7.4   70.0  132      2 -0.89
#> 13    1  14   2    7.5   64.8  125      2 -3.63
#> 14    2   8   2    7.5   65.7  151      2 -1.01
#> 15    2   8   1    7.5   69.1  140      2 -0.41
#> 16    2  13   1    7.6   69.2  131      2 -2.75
#> 17    2  13   2    7.6   70.0  125      2 -1.66
#> 18    2  22   2    7.6   76.2  121      2 -2.28
#> 19    2  16   1    7.7   69.6  136      2 -3.59
#> 20    2   8   2    7.8   63.3  144      2 -1.88
#> 21    2  11   2    7.8   71.5  134      2 -0.37
#> 22    2  14   2    7.8   76.6  124      2  0.16
#> 23    2   7   1    8.0   66.8  139      2 -0.72
#> 24    2   8   1    8.0   69.3  147      2 -0.34
#> 25    2  24   2    8.2   70.7  125      2 -4.09
#> 26    2  10   1    8.2   71.1  145      2 -0.70
#> 27    3  15   1    8.3   70.4  133      2 -2.98
#> 28    3  14   2    8.3   71.5  137      2 -1.51
#> 29    3  14   2    8.3   72.9  141      2 -1.06
#> 30    3  17   2    8.3   73.5  136      2 -1.81
#> 31    3  15   1    8.3   76.1  133      2 -0.93
#> 32    3  13   2    8.4   71.8  143      2 -1.06
#> 33    3  20   1    8.4   77.8  132      2 -1.84
#> 34    3  29   2    8.5   75.6  127      2 -3.69
#> 35    3  21   2    8.5   75.6  133      2 -2.23
#> 36    3  21   1    8.5   76.7  123      2 -2.42
#> 37    3  17   1    8.5   77.6  137      2 -1.08
#> 38    3  15   2    8.6   74.9  134      2 -0.75
#> 39    3  24   1    8.6   77.8  137      2 -2.49
#> 40    3  12   1    8.7   72.1  156      2 -1.22
#> 41    4  27   2    8.7   75.8  127      2 -3.25
#> 42    4  25   2    8.7   80.8  129      2 -1.43
#> 43    4  25   1    8.8   81.4  125      2 -1.66
#> 44    4  24   2    8.9   74.3  142      2 -3.07
#> 45    4  30   1    9.0   77.6  132      2 -3.81
#> 46    4  22   1    9.0   82.2  135      2 -0.99
#> 47    4  16   1    9.1   76.4  145      2 -1.16
#> 48    4  24   1    9.1   80.0  126      2 -1.86
#> 49    4  17   2    9.2   74.8  144      2 -1.41
#> 50    4  15   2    9.2   75.3  145      2 -0.62
#> 51    4  33   1    9.2   80.1  136      2 -3.68
#> 52    4  25   2    9.2   80.9  142      2 -1.40
#> 53    5  16   1    9.4   75.0  148      2 -1.64
#> 54    5  29   2    9.4   81.9  138      2 -1.97
#> 55    5  16   1    9.5   73.2  151      2 -2.27
#> 56    5  17   1    9.6   73.5  146      2 -2.46
#> 57    5  29   2    9.6   84.6  134      2 -1.24
#> 58    5  16   1    9.7   74.2  148      2 -1.92
#> 59    5  20   1    9.7   80.4  142      2 -1.03
#> 60    5  34   1    9.8   80.2  138      2 -3.86
#> 61    5  21   2    9.8   80.9  145      2 -0.67
#> 62    5  20   1    9.8   82.2  139      2 -0.48
#> 63    5  29   1    9.9   78.6  142      2 -3.30
#> 64    5  25   2    9.9   82.9  138      2 -0.84
#> 65    5  15   1   10.0   75.3  157      2 -1.21
#> 66    6  39   1   10.0   81.4  133      2 -4.30
#> 67    6   9   1   10.1   72.6  155      2  0.34
#> 68    6  21   1   10.1   76.7  153      2 -2.42
#> 69    6  17   2   10.1   79.1  143      2 -0.07
#> 70    6  25   1   10.1   79.5  140      2 -2.20
#> 71    6  13   2   10.2   74.6  157      2 -0.13
#> 72    6  25   2   10.3   81.3  141      2 -1.29
#> 73    6  31   2   10.3   85.1  130      2 -1.49
#> 74    6  25   2   10.4   81.3  138      2 -1.29
#> 75    6  37   1   10.6   76.2  153      2 -5.80
#> 76    6  25   2   10.6   80.7  138      2 -1.46
#> 77    6  26   1   10.6   83.1  142      2 -1.39
#> 78    6  36   2   10.6   84.2  135      2 -2.52
#> 79    6  37   2   10.7   84.2  150      2 -2.66
#> 80    7  34   2   10.7   85.7  146      2 -1.83
#> 81    7  47   1   10.7   89.8  131      2 -2.88
#> 82    7  12   1   10.8   82.0  147      2  2.06
#> 83    7  25   2   10.8   83.6  141      2 -0.64
#> 84    7  31   1   10.8   83.8  158      2 -2.18
#> 85    7  34   1   10.8   84.4  144      2 -2.58
#> 86    7  27   2   10.9   79.0  153      2 -2.37
#> 87    7  23   2   10.9   85.9  145      2  0.29
#> 88    7  19   1   11.0   75.4  168      2 -2.36
#> 89    7  26   2   11.0   83.2  145      2 -0.98
#> 90    7  54   1   11.0   85.9  139      2 -4.40
#> 91    7  20   1   11.1   79.9  156      2 -1.19
#> 92    7  23   1   11.1   84.7  146      2 -0.50
#> 93    7  43   2   11.1   86.0  142      2 -2.97
#> 94    8  31   1   11.1   87.1  143      2 -1.25
#> 95    8  40   2   11.2   95.2  137      2 -0.25
#> 96    8  18   2   11.3   75.5  161      2 -1.48
#> 97    8  41   1   11.3   89.4  142      2 -2.26
#> 98    8  40   1   11.3   90.7  148      2 -1.77
#> 99    8  33   2   11.3   91.2  139      2 -0.23
#> 100   8  11   1   11.4   78.4  171      2  1.36
#> 101   8  25   1   11.4   82.2  163      2 -1.43
#> 102   8  27   2   11.4   86.8  155      2 -0.21
#> 103   8  38   1   11.4   89.4  141      2 -1.83
#> 104   8  17   2   11.5   80.6  158      2  0.41
#> 105   8  31   2   11.5   85.4  153      2 -1.41
#> 106   9  45   1   11.5   86.2  143      2 -3.58
#> 107   9  53   1   11.6   81.4  153      2 -5.34
#> 108   9  23   2   11.6   86.3  143      2  0.40
#> 109   9  37   1   11.6   86.3  149      2 -2.56
#> 110   9  38   1   11.6   89.9  146      2 -1.69
#> 111   9  40   1   11.7   93.0  140      2 -1.15
#> 112   9  36   1   11.8   87.1  146      2 -2.16
#> 113   9  29   1   11.9   84.3  166      2 -1.65
#> 114   9  26   2   12.0   81.2  164      2 -1.54
#> 115   9  33   1   12.0   86.0  154      2 -1.93
#> 116   9  36   1   12.0   87.5  150      2 -2.04
#> 117   9  55   2   12.0   96.2  144      2 -1.94
#> 118  10  16   1   12.1   82.3  157      2  0.74
#> 119  10  35   2   12.1   86.6  147      2 -1.75
#> 120  10  43   1   12.1   91.7  151      2 -1.92
#> 121  10  38   2   12.2   80.9  175      2 -3.68
#> 122  10  37   1   12.2   85.5  162      2 -2.80
#> 123  10  41   2   12.2   87.6  155      2 -2.30
#> 124  10  45   2   12.2   90.1  149      2 -2.18
#> 125  10  46   2   12.2   99.6  146      2  0.00
#> 126  10  31   1   12.3   88.0  167      2 -1.00
#> 127  10  44   1   12.3   88.1  151      2 -2.98
#> 128  10  47   2   12.3   94.9  133      2 -1.25
#> 129  10  29   1   12.4   91.8  141      2  0.41
#> 130  10  46   2   12.4   99.6  140      2  0.00
#> 131  10  34   1   12.6   87.0  156      2 -1.83
#> 132  10  38   2   12.7   85.7  152      2 -2.40
#> 133  11  32   1   12.8   87.6  161      2 -1.29
#> 134  11  36   1   12.8   90.4  150      2 -1.22
#> 135  11  39   1   12.8   92.7  152      2 -1.08
#> 136  11  42   1   12.8   94.0  146      2 -1.19
#> 137  11  38   2   12.9   83.4  174      2 -3.01
#> 138  11  26   2   12.9   86.2  159      2 -0.15
#> 139  11  35   1   12.9   90.9  154      2 -0.91
#> 140  11  41   1   12.9   91.2  160      2 -1.78
#> 141  11  32   1   12.9   91.4  148      2 -0.25
#> 142  11  49   1   13.0   93.1  152      2 -2.29
#> 143  11  30   1   13.1   84.3  161      2 -1.85
#> 144  11  32   1   13.1   85.1  136      2 -2.00
#> 145  11  35   1   13.1   87.5  158      2 -1.86
#> 146  11  34   1   13.1   88.6  162      2 -1.37
#> 147  11  45   2   13.1   91.6  150      2 -1.80
#> 148  12  29   1   13.2   90.8  158      2  0.14
#> 149  12  47   2   13.2   91.4  146      2 -2.10
#> 150  12  37   2   13.2   93.8  155      2 -0.18
#> 151  12  35   1   13.3   96.5  142      2  0.56
#> 152  12  15   2   13.4   82.4  157      2  1.72
#> 153  12  38   2   13.4   93.0  152      2 -0.52
#> 154  12  37   2   13.4   94.1  149      2 -0.10
#> 155  12  53   2   13.4  100.3  145      2 -0.74
#> 156  12  37   2   13.5   84.4  168      2 -2.61
#> 157  12  43   1   13.5   96.1  151      2 -0.80
#> 158  12  52   1   13.5   97.0  144      2 -1.72
#> 159  12  45   2   13.5   98.0  148      2 -0.25
#> 160  12  39   2   13.5   99.1  145      2  0.84
#> 161  12  47   2   13.6   97.0  146      2 -0.75
#> 162  13  26   2   13.7   85.7  171      2 -0.29
#> 163  13  24   1   13.8   80.9  172      2 -1.60
#> 164  13  38   2   13.8   93.3  153      2 -0.45
#> 165  13  57   1   13.8   94.0  149      2 -2.89
#> 166  13  42   2   13.8  103.4  144      2  1.43
#> 167  13  42   2   13.9   92.3  175      2 -1.24
#> 168  13  52   1   13.9   95.5  151      2 -2.06
#> 169  13  52   1   13.9   95.7  156      2 -2.01
#> 170  13  57   2   13.9   97.2  153      2 -1.95
#> 171  13  52   2   13.9   97.8  141      2 -1.19
#> 172  13  52   2   13.9   97.9  155      2 -1.17
#> 173  13  49   2   13.9   98.2  150      2 -0.72
#> 174  13  58   2   14.0   96.6  154      2 -2.21
#> 175  13  50   1   14.0   97.9  158      2 -1.28
#> 176  13  39   2   14.1   93.7  155      2 -0.49
#> 177  14  49   2   14.1   96.4  140      2 -1.14
#> 178  14  50   2   14.1   97.1  148      2 -1.10
#> 179  14  59   2   14.1   97.3  152      2 -2.17
#> 180  14  52   2   14.1  102.4  156      2 -0.14
#> 181  14  41   1   14.2   96.7  146      2 -0.35
#> 182  14  53   1   14.2   97.7  151      2 -1.67
#> 183  14  49   2   14.2   98.1  156      2 -0.74
#> 184  14  35   2   14.3   89.8  165      2 -0.91
#> 185  14  50   1   14.3   93.7  152      2 -2.26
#> 186  14  41   1   14.4   94.8  158      2 -0.84
#> 187  14  42   1   14.5   95.1  161      2 -0.91
#> 188  14  58   1   14.5  104.9  130      2 -0.64
#> 189  14  49   1   14.7   96.6  152      2 -1.46
#> 190  14  52   2   14.7   96.7  154      2 -1.45
#> 191  14  51   2   14.7   99.2  155      2 -0.74
#> 192  14  49   1   14.7   99.6  164      2 -0.75
#> 193  15  42   2   15.0   93.5  172      2 -0.94
#> 194  15  28   2   15.1   91.5  164      2  0.84
#> 195  15  41   2   15.2   95.9  165      2 -0.22
#> 196  15  53   1   15.3   95.0  166      2 -2.28
#> 197  15  58   1   15.3  100.7  155      2 -1.55
#> 198  15  48   1   15.3  103.5  143      2  0.30
#> 199  15  54   1   15.5   97.9  159      2 -1.73
#> 200  15  58   1   15.5  101.3  166      2 -1.42
#> 201  15  52   2   15.6  100.0  164      2 -0.68
#> 202  15  50   1   15.7   99.6  153      2 -0.88
#> 203  15  53   2   15.7  103.1  158      2 -0.11
#> 204  15  43   1   15.8   91.1  179      2 -2.07
#> 205  15  57   1   15.9  103.8  161      2 -0.76
#> 206  15  55   2   16.3   95.4  169      2 -2.13
#> 207  16  54   2   15.3  102.0  156      2 -0.48
#> 208  16  44   1   16.3   96.2  173      2 -0.91
#> 209  16  47   1   16.3  102.0  166      2  0.08
#> 210  16  57   1   16.3  108.4  138      2  0.25
#> 211  16  52   1   16.4  103.9  152      2 -0.14
#> 212  16  52   2   16.6   97.8  144      2 -1.19
#> 213  16  56   1   16.6  103.9  148      2 -0.62
#> 214  16  55   1   16.7  106.3  154      2  0.03
#> 215  16  52   1   17.0  101.3  163      2 -0.73
#> 216  16  50   1   17.3  101.8  168      2 -0.36
#> 217  16  53   1   17.5  102.2  168      2 -0.65
#> 218  16  42   1   17.7  100.9  145      2  0.54
#> 219  16  48   2   17.8  111.3  176      2  2.33
#> 220  16  53   1   17.9   98.7  171      2 -1.44
#> 221  16  50   1   18.1  106.4  166      2  0.71