Calculate z-scores for WHZ, HAZ, WAZ using the WHO Growth Reference (2006) for a cohort or sample of children.

Calculate z-scores for WHZ, HAZ, WAZ using the WHO Growth Reference (2006) for a cohort or sample of children.

getCohortWGS(data, FUN = getWGS, sexObserved, firstPart, secondPart, index)

Arguments

data
Data frame containing the variables needed for calculation

FUN
Function to apply; default to getWGS()

sexObserved
Sex of child (1 = Male; 2 = Female)

firstPart
Weight (kg; for WHZ and WAZ) or height (cm; for HAZ)

secondPart
Age (months; for HAZ and WAZ) or height (cm; for WHZ)

index
One of "wfh", "hfa", "wfa" (specifies the required index)

Value

Numeric vector of z-scores of the anthropometric index selected

The function fails messily when secondPart is outside of the range given in the WGS reference (i.e. 45 to 120 cm for height and 0 to 60 months for age). It is up to you to check the ranges of your data.

The reference data for W/H assumes supine length is used for children with a standing height below 85cm

Heights should be specified in cm to the nearest mm (i.e. to 1 d.p.)

Ages should be specified in whole months

Weights should be specified in kg to available precision

The function requires reference data wgsData included in this package

Examples

  # apply getWGS to first child in sample data anthro1
  wazAll <- getCohortWGS(data = anthro1,
                         sexObserved = "sex",
                         firstPart = "weight",
                         secondPart = "age",
                         index = "wfa")
  wazAll

#>   [1] -0.756055486 -1.390215032 -1.055978534  1.415750955 -2.677572422
#>   [6]  1.492380502 -0.129877040 -0.023481593 -1.506473442 -1.543816296
#>  [11] -2.874957122 -0.434972402 -1.038995403 -1.692818549 -1.312458980
#>  [16] -2.210032596 -0.011892256 -0.909177616 -0.678398553 -0.947466947
#>  [21] -2.499604250 -0.956596436 -1.654426862 -1.250527597  0.673357507
#>  [26]  0.301563007  0.242613457 -2.786707088 -1.158206506 -1.154771825
#>  [31] -1.355408197 -0.591349587 -4.149672176 -0.457487519 -0.743316691
#>  [36] -1.697258365 -1.057450666 -0.188695078 -0.420957702 -2.210304143
#>  [41] -1.305367153 -3.637781428 -0.606625256 -0.543604699 -1.591717797
#>  [46] -1.747457376 -0.348033375  0.698961487 -0.744671297  0.189245725
#>  [51] -1.481088556 -1.957531446 -1.311912614  0.301554844 -0.757890327
#>  [56] -1.226363368 -1.274787811 -0.145532096 -1.304808185 -1.962091877
#>  [61] -1.928346854 -1.045541559 -1.397289784 -1.481088556 -0.749372496
#>  [66] -1.125037571 -2.776300420 -0.198826369  0.042195760 -0.521945326
#>  [71] -1.465031581  0.854087182 -1.637194752  0.314098599 -0.812227665
#>  [76] -1.002263141 -0.447220866 -1.172281706 -2.153122123 -1.818201583
#>  [81] -1.131968739 -2.868805513 -0.735551227 -2.430221048 -2.872771868
#>  [86] -0.943465717 -2.748662500 -2.550612731 -1.967110644 -3.365311234
#>  [91]  0.741080763  0.233970396 -1.169615107 -1.543684754 -0.501936967
#>  [96] -2.804976960 -1.895827835 -1.713427237 -1.872694403 -1.648348298
#> [101] -1.653569907 -0.664600505 -1.811316644  0.143249353 -3.229518785
#> [106] -2.583688246 -2.218952825 -0.286176123 -0.531412722 -1.673624449
#> [111] -1.909606074 -2.776300420  0.120104700  0.546563755 -0.879654589
#> [116]  0.124839647 -2.852646118 -3.083157784 -2.823920666  0.315261149
#> [121]  0.209088574 -2.476374011 -2.122715091 -0.318150340 -1.312458980
#> [126] -1.695358739 -0.902766193 -0.373410410 -1.112382697 -1.879607319
#> [131] -2.892609089 -1.513067916 -1.332513353 -1.551030431 -2.041633539
#> [136] -0.275693702 -0.944461306 -0.209682330 -2.238226801 -3.592510318
#> [141] -2.013488349 -0.084605129 -1.105513344 -1.531545257  0.617931505
#> [146] -2.743872367 -3.493445475 -1.153931285 -1.846156645  0.225663608
#> [151] -1.532200095  0.359795481 -2.952944021 -2.161430916 -0.910802060
#> [156]  0.552954267 -1.363669512 -0.878309363 -2.196964263 -1.859606282
#> [161] -0.844611315  0.131834686 -1.633907207 -0.141055654 -2.297881902
#> [166] -1.744077852 -2.404642192 -0.640953054 -3.805458627 -2.175400988
#> [171] -3.175229907 -4.951617316  1.060809620 -1.411766933 -0.086520173
#> [176] -1.059711075 -1.595718548 -1.304808185  1.683150739  0.287619430
#> [181] -2.466697289 -3.877673946 -3.715955123 -5.069451389 -1.292488999
#> [186] -2.132021716  0.097411137 -5.327514429 -2.024963023 -0.717288982
#> [191] -0.404520268 -1.577966855 -1.950532218 -3.048756422 -1.489593749
#> [196] -0.343636789  0.824983186 -1.399353926 -1.648172643 -4.367371436
#> [201] -3.433431026 -0.598778294 -2.601971260 -0.231187486 -1.957531446
#> [206] -3.681708972 -0.546363468 -3.068549949 -0.730357826 -1.144613582
#> [211] -1.997551142 -0.767387647 -1.074102109 -1.141639829 -2.245087848
#> [216] -1.751902365 -0.624691967  0.484916295 -0.154383628 -2.412799886
#> [221]  0.484102095  0.113248367 -1.819784699 -0.524863586 -2.241480963
#> [226]  0.232574778 -0.592830223 -4.408006457 -0.916470904 -0.270281822
#> [231] -3.362407730  1.659060555 -2.330062628 -0.025358589 -2.266557118
#> [236] -1.906475790 -0.162723565 -0.909685477 -3.015559997 -1.487433025
#> [241] -2.638696253 -0.328848959 -0.982682844  0.360114336 -2.351076974
#> [246] -0.878212777 -2.837649328 -2.201048684 -0.660124491 -1.367404192
#> [251] -1.464122373 -2.175835476 -1.775624791 -2.984881693 -1.713427237
#> [256]  1.830183855 -0.657047325 -2.102158382 -2.462376472 -1.896835294
#> [261] -0.406636636 -2.867595482 -0.449522123 -1.570105496 -2.671399103
#> [266] -1.246443716 -0.851668189  1.196571231 -0.359280565 -3.434281901
#> [271] -2.019348230 -2.311109808  0.715963423 -0.907679960  0.546544175
#> [276] -1.937404501 -0.697417462 -1.141639829 -0.549192808 -2.154154451
#> [281] -2.685855791 -1.855047344 -3.094781320 -0.337836304 -0.351805641
#> [286] -3.177111729 -2.629294375 -1.081084723 -1.937227979 -0.245582426
#> [291] -1.627670643  0.354050543 -2.798447133 -0.689401520  0.182498736
#> [296] -0.112542159  1.594478333  0.508973986 -3.108289096 -0.663817695
#> [301] -0.221651631 -0.034977312 -3.183148429  1.853282581 -0.792791017
#> [306] -1.576880204 -0.123516659 -2.845299102 -1.604658179  0.205987272
#> [311] -2.051523837 -2.301013361 -0.384022524 -0.230183057 -3.026927179
#> [316] -0.835240203 -2.462999394 -0.109811373 -1.960514990 -0.878309363
#> [321] -1.394810748 -2.351455627 -1.543684754 -0.715547505 -2.945522502
#> [326] -0.617988079 -0.272684214 -0.342539895 -2.852646118 -3.084644744
#> [331] -1.170014829 -0.093618819 -0.781064479 -0.498923459 -3.163469027
#> [336] -3.646604750 -1.049788802  1.624849713 -2.748662500 -2.972835153
#> [341] -0.434972402  0.062925270 -2.284855185 -2.747861331 -1.643124128
#> [346] -0.546904918 -1.241599699 -2.697575527  1.122813676 -2.466697289
#> [351] -2.126793106 -0.863317478 -1.551030431 -0.525671169 -1.620356652
#> [356] -1.480633991  0.659527471 -2.148186825 -0.626673702 -1.172186556
#> [361] -0.989785269 -2.852646118 -1.038912856 -1.850928121 -2.982414705
#> [366] -1.817759528 -0.231187486 -1.906475790  0.114824375 -1.000060046
#> [371]  0.554692600 -0.042255588 -3.268576579 -1.470900381 -2.748662500
#> [376] -1.788749384 -3.226884329 -0.743534869 -0.572020880 -1.670026320
#> [381] -3.415159500 -0.714024159 -0.817410344 -1.755385241 -0.200526598
#> [386] -4.477574185 -0.645911959 -0.815842738 -3.096726406  1.029333951
#> [391] -2.930995215 -1.389912367 -2.609795520 -1.456936597 -0.916470904
#> [396] -2.529748930 -0.750465925 -3.326437506  1.656554460 -1.692179889
#> [401] -1.104944738 -0.956596436 -1.582545376 -1.142387124 -0.584079302
#> [406] -3.410344932 -0.595649505 -0.817341033 -0.498532338 -5.039788691
#> [411] -3.083157784 -2.485977391 -2.350118962 -2.689754701  0.833784797
#> [416]  0.758391899 -0.518548924 -1.879607319 -0.121723944 -1.504595211
#> [421]  0.003316075  0.599922687 -3.337108252  2.324555793 -2.875030801
#> [426] -1.152732408 -1.569441473 -1.243250254 -1.014305170  1.604764700
#> [431] -0.195662157 -1.930618814  0.268531118 -0.214333202 -1.342903149
#> [436] -0.216382295 -1.521747332 -2.919187060  0.567307198 -4.651884618
#> [441] -1.349340494 -2.114330307 -0.946376876 -0.663455149 -0.268177085
#> [446] -0.128654219 -1.944055028 -1.544890736 -0.014281488 -1.650517990
#> [451] -1.410010752 -3.524498776 -1.226363368 -0.452656139  0.090162278
#> [456] -1.480633991 -1.262175350 -0.927424651  0.501466357 -0.249922102
#> [461] -4.511094693  0.374835201 -2.603203677 -2.399370924 -2.666782401
#> [466] -2.193814705 -1.523353689 -3.342403222 -0.933738370 -0.844176406
#> [471] -2.134256394 -0.871521751 -1.216546155  0.058795927 -1.361065260
#> [476] -0.014281488 -2.813667619 -1.731457716 -0.687324323 -0.305595522
#> [481] -4.039396324  1.005822670 -1.343177590  0.101291935 -2.408153427
#> [486]  1.075956965 -0.449522123 -1.454843573 -2.350118962 -0.447220866
#> [491]  0.421537602 -1.049986755 -2.052932471 -0.649019916 -2.051958705
#> [496] -1.951960311 -4.254778035 -2.534408271  0.531976227 -1.478207396
#> [501] -2.550612731  1.766831596  1.025450917 -0.323438320 -3.083459957
#> [506] -1.961505342 -1.100241679 -0.452489616 -3.030742894 -4.066108952
#> [511]  0.498851361 -2.318231031 -1.475287430  0.803568582 -0.804359848
#> [516] -2.568949038 -0.685565429  1.097742915 -0.076376870 -1.885063827
#> [521] -0.948512096 -1.936713057 -1.701211743  0.130129784 -3.845337433
#> [526] -0.331889357  0.724186061 -0.271255044  0.645253177 -0.809001348
#> [531] -1.042698295 -0.701776155 -0.605678014 -1.241599699 -4.014604456
#> [536] -0.743660212 -0.858444864 -3.428359547 -0.344626732 -1.448416374
#> [541] -1.085946409 -1.549164179 -1.180456124  0.274568696 -0.762104740
#> [546] -1.657673616  0.153570942 -0.638731864 -1.725515858 -2.913894869
#> [551] -2.132464038 -2.888030513  0.436084121 -0.409948871 -0.858652176
#> [556] -2.840267647 -0.129877040 -1.388124454 -1.464762653 -0.308801294
#> [561] -1.393951704 -0.404520268 -4.664513469 -2.860611693  2.966593355
#> [566] -2.474342401 -1.951960311 -1.411579782 -0.333674773 -2.158028743
#> [571] -4.011807265 -0.123516659 -2.016781977 -2.313538179 -1.685406984
#> [576] -1.125037571 -0.257199864 -3.030742894 -2.042630562 -2.596165900
#> [581] -1.353288267  2.266619108 -2.550530853  0.762734511  0.557075564
#> [586] -1.364228189 -1.074102109 -0.511081531 -1.087875713 -1.642167585
#> [591] -0.756778223 -1.775358593 -1.387207574 -0.172291102 -2.514542483
#> [596] -1.071869289 -0.634877120 -2.350118962 -3.835376421 -1.668819462
#> [601] -1.525451697 -1.889753425 -2.445798052 -2.062251974 -0.804359848
#> [606] -1.353288267 -3.287335255 -0.562606796 -2.584900012 -0.137833590
#> [611] -3.046645181 -2.408153427  0.778753435 -1.389912367 -1.696435815
#> [616] -1.850999818 -1.411766933 -0.522322077  0.778384280 -1.964497833
#> [621] -1.988019429 -1.763834261 -0.845180918 -1.473675968 -0.470367764
#> [626] -0.289366093 -4.087888560 -0.270281822 -2.303637370 -0.482029856
#> [631]  0.464160455 -2.435950945 -2.621198857 -2.126793106 -0.209682330
#> [636] -1.509367003  1.289057432 -0.767052573 -1.700482921 -2.269642865
#> [641] -1.002263141 -2.215960625 -1.524536607 -1.475287430 -1.627670643
#> [646] -2.418634789 -0.193240433 -1.027979181 -0.040013672  0.027710306
#> [651]  1.205587671 -0.613928267  0.495278185 -2.456080880  0.436084121
#> [656] -1.376517027 -3.526586034 -1.624976321 -2.866375030 -1.332737687
#> [661] -0.030921521 -2.031149118 -0.132988301 -1.799535611 -2.603203677
#> [666] -0.589188541 -0.437628860 -2.162872522 -0.238066632 -0.988275103
#> [671]  2.103491750 -1.147769848 -2.823920666 -2.187585929 -1.509367003
#> [676] -0.335766390 -0.813109640  2.551604501 -0.152565280 -1.760010328
#> [681]  0.351233604 -2.025953359 -1.692818549 -1.648348298  0.675125807
#> [686]  1.641129715 -0.997128278 -1.080035704 -0.827786980 -0.503220678
#> [691] -2.174841454 -0.262529902 -1.657673616 -1.172281706 -1.707753151
#> [696] -2.428866741 -2.272017605 -0.280036813 -0.838269831 -3.168505051
#> [701] -0.160582375 -0.316093558 -3.328803635 -0.315868701  0.068344284
#> [706]  0.436084121 -1.285488843 -2.529748930 -0.324737326  0.075253298
#> [711] -1.782714555 -2.891236422 -0.878309363 -1.903092625 -0.860248820
#> [716] -0.702754409 -1.775720121 -1.849894742 -0.628198870 -2.184054769
#> [721] -3.001646010  0.267970601 -1.106291572 -3.274035863  2.132214614
#> [726] -1.898739963 -2.197815642 -1.389751490 -1.333609157 -0.376548425
#> [731] -1.725515858 -1.706310228 -1.412103599 -3.068806776 -1.470900381
#> [736]  1.327696179 -1.576880204 -0.651362633 -0.755316902 -2.741776036
#> [741] -1.059711075  1.207201995 -2.071206423 -0.008165454 -0.422695504
#> [746] -1.805834528  0.075253298  0.028631512 -1.026902741 -1.250906814
#> [751]  0.328478004 -1.370781875 -1.642348564 -2.952944021 -2.115011933
#> [756] -1.692818549 -0.947018947 -0.819916798 -2.419833091 -1.511242282
#> [761] -1.353288267 -1.153931285  1.656554460 -1.169318856 -1.103874820
#> [766]  0.734507043 -0.943273158 -0.211721260 -2.901500553 -2.360617971
#> [771]  0.488710373 -2.188135215 -2.826634719  0.903004850 -1.421667610
#> [776]  0.116233552 -0.025358589 -2.674200233 -0.027168645  0.003316075
#> [781] -2.747861331 -2.435605996 -1.459565720 -2.832711598 -1.659875006
#> [786] -1.515805936 -0.708608966 -0.727915306 -0.450772962 -0.177468244
#> [791] -2.167216091 -2.089107721  0.977002274 -1.037272298 -1.515805936
#> [796] -1.798663137 -2.210304143 -2.692434016 -0.216382295 -1.906475790
#> [801] -1.944771848 -2.592449469 -0.430535627 -1.051552300 -1.840658484
#> [806] -3.383950213 -0.974877565 -1.485446939 -1.574010185 -1.772335638
#> [811] -1.325962622 -2.360617971  0.299402915 -3.334475623  0.973319452
#> [816] -2.098858543 -0.930134981 -1.596543076  0.107549191 -2.307175486
#> [821] -0.183040002 -0.547336342 -1.990288709 -0.188695078 -1.554983693
#> [826] -2.210304143 -0.141055654 -0.753262437 -0.643845767 -0.155222438
#> [831] -0.661878973 -2.528107466 -2.135988044 -1.659875006 -1.172186556
#> [836] -0.637561917 -1.538117281 -2.364641588 -2.985367166 -0.136270930
#> [841] -0.521945326 -1.431643354 -1.602651852 -1.294456317 -2.421870681
#> [846] -0.952420756 -2.599948845 -2.653390053 -0.269574295 -4.207402273
#> [851] -0.663694269 -0.724066677 -1.165794147 -0.256948978 -2.098324404
#> [856]  0.286117864  0.975075622 -2.748662500 -0.295178066  0.242613457
#> [861] -2.016633344 -2.200605501 -2.491486698 -0.678143030  0.000000000
#> [866] -1.509367003  0.003316075 -1.835853172 -2.059231383 -0.025358589
#> [871] -3.428359547 -0.731155305 -1.015223977

  hazAll <- getCohortWGS(data = anthro1,
                         sexObserved = "sex",
                         firstPart = "height",
                         secondPart = "age",
                         index = "hfa")
  hazAll

#>   [1] -1.225816946 -2.347588600 -2.951804110 -0.281285163 -4.205666336
#>   [6] -0.538767762 -2.402071937 -1.031769943 -2.741088448 -4.703757107
#>  [11] -2.567054999 -2.114496047 -2.232350548 -2.315545790 -2.751616526
#>  [16] -2.793069374  0.112134888 -1.900179710 -2.954372959 -1.967104178
#>  [21] -3.871652199  0.866720638 -2.825206927 -2.141228500 -2.799464293
#>  [26]  0.549645927 -1.437200160 -3.797940968 -2.566175205 -1.830118281
#>  [31] -1.654858885 -2.711033318 -3.639964208 -1.795506877 -1.677509989
#>  [36] -1.031769943 -0.435688124 -1.266015205  0.499032644 -4.608566028
#>  [41] -3.166235115 -1.069592983 -1.847793611 -2.550231358 -1.830118281
#>  [46] -2.275549336 -3.281653219  0.487677381 -2.439641036 -0.479474373
#>  [51] -2.450429333 -3.205342947 -3.769430576 -0.277522775 -0.365386792
#>  [56] -2.581582674 -1.777577458 -0.954753624 -2.959961331 -3.140270937
#>  [61] -2.377870949 -2.187306414 -2.558538134 -2.180662247 -3.237652935
#>  [66] -2.263484502 -4.200005278 -0.280901088 -3.518043240 -0.243778110
#>  [71] -1.054376364  0.181249949 -2.485976566 -1.614151364 -2.992821011
#>  [76] -2.455903841 -1.968311162 -1.208150734 -2.346087378 -2.191971354
#>  [81] -2.626496085 -4.662923799 -2.211062752 -3.960598636 -3.100749795
#>  [86] -4.523449084 -1.475548558 -0.355471426 -2.574438064 -1.839393339
#>  [91] -1.461962885  2.288305751 -2.700799038 -2.447733338 -1.400156261
#>  [96] -2.867988675 -2.982303108 -3.346511053 -4.011175831 -3.284974582
#> [101] -5.909968736 -1.477951047 -1.947586543 -2.166784631 -4.972619408
#> [106] -4.099971316 -3.817940382  1.316886126 -2.036630209 -3.530280966
#> [111] -5.713938494 -5.670452344 -1.172586314 -3.145183047 -1.394329114
#> [116] -0.204123481 -3.631160538 -4.199736816 -4.118235233 -1.484444456
#> [121]  0.736843720 -4.680507899 -3.342792864 -0.149568850 -2.403374310
#> [126] -2.061088282 -0.769638885 -0.104810144 -4.426120485 -3.696953777
#> [131] -3.802341050 -3.555441559 -1.788208214 -3.529497681 -3.283522671
#> [136] -1.487378089 -2.509692349 -3.547868182 -2.597272665 -4.026114213
#> [141] -3.869729765 -1.064010524 -2.917504307 -2.597273135 -1.954313914
#> [146] -3.952845711 -4.995728974 -1.515679486 -3.784901354 -0.782529788
#> [151] -1.722510242  0.178478600 -3.487344614 -2.508428118 -3.765114839
#> [156] -1.151558283 -2.951104099 -1.101593899 -2.822404382 -2.799210947
#> [161] -2.562246061 -3.061881876 -1.949376906 -2.045778703 -4.088313893
#> [166] -2.263484502 -2.350855289 -5.166415805 -5.013216658 -1.908751318
#> [171] -4.827985991 -4.376194162 -2.299446566 -1.910895160 -0.805011230
#> [176] -3.618584552 -1.304576945 -3.154163809 -0.119085656 -0.369787668
#> [181] -3.153821221 -5.015569333 -3.418404266 -4.812793388 -2.746638937
#> [186] -3.858564391 -1.545743793 -5.305518401 -5.745638443 -3.348297345
#> [191] -0.634054049 -1.397058098 -2.524106907 -3.874696548 -2.523000908
#> [196] -2.236128801 -0.119506727 -2.653112493 -0.602577963 -4.957933429
#> [201] -3.849328151 -4.322133305 -2.153685538 -3.437551244 -3.618584552
#> [206] -3.479468305 -3.900658758 -3.696103188 -1.609724444 -2.240223489
#> [211] -1.885412674 -2.212357359 -2.585312877 -3.946294272 -2.419055594
#> [216] -2.811997656 -4.203915398  0.615538319 -1.160642255 -3.178260461
#> [221]  0.692112058 -0.536370564 -3.096787409 -2.436515279 -3.952845711
#> [226]  1.223163435 -2.789660454 -3.352205284 -2.069096641  0.464336645
#> [231] -5.418686878 -1.129834986 -4.447470873 -2.625901124 -2.939231766
#> [236] -2.384871747 -0.594772365 -3.187728089 -1.297610401 -2.808676999
#> [241] -3.941110820 -1.156638026 -2.873721651  0.357571363 -4.192248991
#> [246] -2.332416824 -2.119828036 -5.776581228 -0.840970311 -1.735442831
#> [251] -2.505794683 -3.386518625 -3.059933017 -0.499055368 -2.243951242
#> [256] -2.267048884 -1.133131395 -3.558479658 -5.482457697 -3.514836767
#> [261] -3.769391412 -4.676505593  2.465527134 -3.008387093 -5.619044290
#> [266] -1.333785180 -2.191781078  1.035812182 -1.643583490 -5.157947310
#> [271] -3.475860784 -3.827147417 -0.947773992 -2.463695689  1.346964832
#> [276] -1.457573764 -0.793617982 -1.585128972 -0.263517977 -4.129339190
#> [281] -3.064734071 -3.239968708 -5.490461954 -1.603027974 -1.995302370
#> [286] -4.057894264 -3.343090149 -2.777643198 -2.008377936 -2.631980882
#> [291] -1.668039699 -2.177511620 -4.754353088 -0.867080933  0.168475728
#> [296]  3.383466557  1.609985523  0.541449785 -3.654045789 -1.056218557
#> [301] -1.727547492 -1.445194749 -5.548698264  1.400281271 -2.221420946
#> [306] -3.558479658 -0.090409302 -3.471394771 -3.083504216  0.239652502
#> [311] -3.296402290 -3.462160326 -1.748200069 -0.854272833 -5.898272500
#> [316] -3.038996926 -5.547532409 -1.707307382 -1.875763170 -1.560197947
#> [321] -1.980985088 -4.561090718 -2.895411743 -2.230541873 -4.213942890
#> [326] -1.623306741 -2.440607333 -1.550441594 -3.662155566 -2.750086563
#> [331] -1.504701208 -1.221065797 -2.788941887 -3.162045695 -3.949812883
#> [336] -3.016921002 -1.263131018  0.905102767 -3.879244080 -4.443113935
#> [341]  0.038999275 -1.285847308 -3.674967206 -2.964285344 -0.756876817
#> [346] -1.532803223 -3.032074218 -3.875092689 -3.125157706 -3.058448130
#> [351] -2.645768545 -2.393253354 -1.290430860 -1.042360473 -3.583651358
#> [356] -1.325633695 -1.064010524 -4.441639132 -2.784611661 -1.500976882
#> [361] -1.329633349 -4.313051158 -0.617686359 -3.821535021 -5.735281129
#> [366] -2.676696337 -4.554001794 -3.239123564  0.817895718 -2.087440024
#> [371]  0.456863436 -0.009249374 -3.195236957 -1.523562559 -3.110061513
#> [376] -3.899174953 -4.277700054 -3.322554398 -3.569170482 -2.857068359
#> [381] -4.947725144 -1.216334994 -2.243533354 -1.425647779 -0.547141757
#> [386] -5.522432406 -1.670648194 -2.921711910 -4.277700054  0.290975720
#> [391] -3.802122905 -2.008014419 -4.383703735 -1.578896963 -1.142773715
#> [396] -3.388911027 -1.560197947 -3.951000078 -0.770519401 -3.883119375
#> [401] -2.372030618 -2.373403461 -2.630562844 -4.867138457 -1.198914499
#> [406] -5.090002501 -3.592372241 -3.065463463 -1.030958590 -5.593698426
#> [411] -3.398504976 -1.623306741 -3.093133903 -5.420221803  0.719729653
#> [416] -0.722602729 -3.151478524 -2.742698905 -1.426691981 -2.977083019
#> [421] -1.641814404  0.274046409 -4.398083349 -1.347507607 -3.594231919
#> [426] -1.509748644 -3.879120764 -2.508843726 -3.044175143  1.571777383
#> [431] -4.690771704 -2.530114921 -0.721724381 -1.362668856 -2.795804266
#> [436] -4.075999921 -0.668456287 -5.971469448 -3.030740764 -5.506217762
#> [441] -4.719932513 -2.117112302 -2.916932137 -1.608906916 -0.817180488
#> [446] -1.301749281 -2.547621194 -2.008377936 -0.492444117 -2.762359467
#> [451] -1.980150294 -4.102616146 -0.554904488 -2.402783639  1.995261668
#> [456] -3.195236957 -2.335117806 -1.578892148  1.304141498 -1.394182429
#> [461] -5.397877145 -0.426413909 -3.969650427 -4.679790680 -3.979524337
#> [466] -1.698624326  0.015972070 -4.396011437 -2.854046358 -1.876924367
#> [471] -2.504382751 -2.272068875 -1.838825686 -0.832860956 -4.544866790
#> [476] -0.903092257 -3.022471583 -2.887668074 -1.020663773 -1.099259682
#> [481] -4.513085573 -3.218905489 -2.087487917 -0.520981838 -4.463183790
#> [486] -0.603990696 -0.337426106 -1.715458090 -3.563005280 -1.783985243
#> [491] -0.215706396  0.117961198 -3.197103817  0.222137230 -1.908751318
#> [496] -4.507792167 -4.951896784 -4.554464906 -2.801126546 -4.844507931
#> [501] -2.839499811  1.603511517  0.273011683 -0.311941562 -4.158365910
#> [506] -5.077341314 -2.107934183 -1.711426633 -3.867585890 -3.558330481
#> [511]  0.117548106 -4.392032458 -1.613664241  0.752240413 -2.373744615
#> [516] -4.298918517  1.449773160  0.805200630 -3.487344614 -3.147493054
#> [521] -1.892180138 -3.265028084 -2.744787360 -2.627143560 -3.257164756
#> [526] -2.046369071  0.415752119  0.522173240 -2.630562844 -2.602471186
#> [531]  0.154531837 -0.261146883 -1.828404014 -2.626814225 -5.029593372
#> [536] -2.048795681 -1.163622714 -3.952845711 -1.918792036 -1.674575890
#> [541]  0.188598205 -2.722446795 -1.446586698 -1.763978786 -1.508232166
#> [546] -2.751478076 -1.469174957 -2.301507714 -3.863157399 -4.811355396
#> [551] -2.726167249 -5.067580027  0.527216210 -1.776814223  2.395525974
#> [556] -2.055376100 -3.308546246 -2.828831254 -1.493017135 -0.517977956
#> [561] -0.197917388 -2.328765003 -3.933048875 -4.199736816  2.512838669
#> [566] -4.596842270 -3.803374333 -3.306970551 -0.490310404 -2.231011263
#> [571] -5.591585222  1.632536281 -5.178188136 -2.125785995 -2.140448950
#> [576] -2.442585982 -3.401340812 -4.718604696 -3.585242941 -2.918416438
#> [581] -2.646620597 -2.578526404 -4.496864004  0.844389299 -0.756930769
#> [586] -3.445467524 -3.421590846 -0.291261951 -2.499529238 -3.264449556
#> [591] -3.156813868 -3.474904397 -2.309227482 -2.011499806 -2.136288566
#> [596] -1.940398326 -2.419284394 -2.182758112 -5.765527356 -2.353078590
#> [601] -2.157071263 -3.280218923 -1.990498565 -2.901911337 -1.042360473
#> [606] -3.384489875 -3.278800165  1.563091429 -2.958328830 -1.732701803
#> [611] -3.757649596 -4.205657751 -0.702558245 -2.190229884 -3.621739202
#> [616] -4.139259210 -1.991825286 -1.979348827 -1.147547774 -1.684383106
#> [621] -0.701066649 -2.839499811 -2.656367186 -0.794809100 -1.179870941
#> [626] -1.713908222 -6.027864397 -0.948493276 -2.534739175 -3.464781745
#> [631] -0.660617021 -4.158078833 -3.957596636 -2.190229884 -3.840129646
#> [636] -2.720168352 -1.625991598 -2.142515019 -3.444545833 -4.271573934
#> [641] -2.613476670 -3.001279819 -2.136011301 -2.626814225 -4.057894264
#> [646] -2.783284139 -0.742758510 -2.324543879 -1.195041599 -1.037748386
#> [651] -0.441369585 -2.411693818 -0.378986542 -3.566844073 -1.019919395
#> [656] -1.529089473 -4.744280994 -1.155055822 -5.330910332 -1.480382954
#> [661] -1.870731147 -3.881193568 -1.635519416 -3.923557457 -4.410495276
#> [666] -1.684091197 -1.579335273 -3.809460924 -0.146786529 -2.564774812
#> [671]  0.314585996 -1.175784960 -4.314727974 -3.508557174 -3.218915757
#> [676] -1.156785580 -0.900908964 -0.719684966 -1.603889678 -2.266761619
#> [681] -0.791017273 -4.975392391 -1.722058199 -3.313883723 -2.931428264
#> [686] -2.869970221 -1.956287662 -2.373403461 -1.686202439 -1.443499284
#> [691] -3.557605835 -2.150499975 -3.051967127 -3.247254990 -3.507539472
#> [696] -2.724291047 -3.198008284  0.565039106 -2.390836492 -2.903088279
#> [701] -3.104485728 -3.936080567 -4.455079933 -3.874059297 -1.122958640
#> [706] -3.082766867 -1.817837439 -2.820854601 -1.559234291 -0.504207527
#> [711] -2.424551764 -3.846765367 -2.369499207 -2.250374042 -3.754493239
#> [716] -0.570741618 -3.163841916 -0.803801920 -1.875493945 -1.098824528
#> [721] -2.842835944 -2.085976455 -4.398833456 -3.228225172  2.571505776
#> [726] -1.820868625 -3.425667438 -1.941978888 -4.297820424 -2.140502233
#> [731] -2.648973484 -3.635946558 -2.560952394 -4.334114065 -1.873686094
#> [736] -0.981931443 -2.977083019 -0.680484486 -1.670959425 -4.070063116
#> [741]  0.858199504 -2.646620597 -3.100749795 -1.710467040 -1.698168024
#> [746] -2.373744615  0.685896785 -1.924238389 -0.226516521 -2.572296249
#> [751] -3.681783208 -2.503450982 -3.002006341 -6.502219525 -2.361420578
#> [756] -2.747173129 -0.745650670 -1.326222940 -5.282432448 -2.878039273
#> [761] -2.064092219 -1.096788606  0.090953391 -2.077042993 -1.828964751
#> [766]  1.353407598 -3.047554873 -0.740782791 -1.674764052 -2.890588105
#> [771] -0.697247406 -2.510539385 -3.462160326  0.635041470 -2.501879733
#> [776] -3.779964092  0.188968844 -4.295538260 -1.273370850 -0.674551044
#> [781] -3.274216548 -3.093133903 -2.046369071 -5.247298467 -4.715986525
#> [786] -2.924175085 -1.416620832 -2.174394213 -2.995211297 -1.401273176
#> [791] -2.503450982 -3.033776836 -0.888525830 -2.913910396 -2.267048884
#> [796] -2.943644779 -4.230892107 -4.000260738 -0.314829271 -3.894049957
#> [801] -1.294045944 -2.519799963 -0.706154917 -2.732304721 -2.464710227
#> [806] -3.944366530 -2.009020684 -2.366409460 -3.077067404 -1.563495802
#> [811]  2.167073997 -1.744749254 -3.611437916 -2.839499811 -2.052435483
#> [816]  0.107864848 -2.830331611 -3.157306384 -0.114885165 -3.290485599
#> [821] -1.744801334 -2.324536232 -2.544352375 -0.351837955 -3.164981569
#> [826] -4.069031854 -1.829965034 -1.655523665 -2.835330263 -0.105623986
#> [831] -1.931489671 -3.847103931 -2.352253104 -1.678473762 -0.739875302
#> [836] -2.219406984 -1.396196587 -4.311822233 -3.957956760 -0.200001410
#> [841] -1.931181331 -4.278200710 -1.097449405 -1.990634335 -4.238427485
#> [846] -1.687066833 -2.565109192 -3.094725233 -1.049294945 -4.515130092
#> [851] -1.860139841 -2.197199919 -0.858791684  0.060193215 -2.834693526
#> [856] -1.965618132  0.033009941 -2.116534031 -0.595288188 -3.696103188
#> [861] -2.934280794 -3.222556228 -3.375454301 -1.495970174 -0.244538684
#> [866] -3.128234411 -0.029708803 -2.428728078 -3.018692131 -2.211949658
#> [871] -4.084156402 -1.913971036 -1.073637542

  whzAll <- getCohortWGS(data = anthro1,
                         sexObserved = "sex",
                         firstPart = "weight",
                         secondPart = "height",
                         index = "wfh")
  whzAll

#>   [1]  0.0557234714 -0.0197490262  0.5746911171  2.0623174908 -0.1408004373
#>   [6]  2.4904724554  1.8331519688  0.9361489054  0.1854194251  2.1159928654
#>  [11] -1.9694388656  1.0635104657  0.3531583035 -0.6115100340 -0.0104944116
#>  [16] -0.7503899336 -0.0800032190  0.3127757304  1.5645617536  0.2215208732
#>  [21] -0.0879875727 -2.1419787686 -0.3080482318  0.0077822704  3.2104141296
#>  [26]  0.0743446827  1.4096698641 -0.8148504970  0.6381664692 -0.3354039181
#>  [31] -0.6195553255  1.3571695186 -2.7736467092  1.0083109512  0.3284206278
#>  [36] -1.6670528060 -1.2115770176  0.8902447187 -0.8986503680  0.8216639340
#>  [41]  0.6444213746 -4.3984785014  0.3841113996  1.4829984704 -0.9306849536
#>  [46] -0.8855822827  1.6955141048  0.6514364924  0.6126939736  0.5981389091
#>  [51] -0.1687010067 -0.4677322410  1.1362773521  0.5423233261 -0.8000913452
#>  [56]  0.4770736434 -0.2522323706  0.6662672044  0.6647050527 -0.5654052264
#>  [61] -0.8934924962  0.3767994172 -0.2418318259 -0.4179779915  1.9717922269
#>  [66]  0.3477428706 -0.1269675642 -0.0295286635  2.2988976140 -0.5611723075
#>  [71] -1.3170243507  1.1419563020 -0.0745071984  1.7408584768  1.5338173017
#>  [76]  0.5911011325  0.9523246292 -0.8148504970 -1.0439986774 -0.9146392685
#>  [81]  0.1619047455  0.1222311200  0.9027398430  0.0187872624 -1.8700083738
#>  [86]  2.1214519835 -2.7141326484 -3.4816199297 -0.7728203841 -3.1692342709
#>  [91]  2.5244154302 -1.3862961957  0.6708359067 -0.4927296707  0.5046813500
#>  [96] -1.8487529196 -0.0776821004  0.5857013731  1.4000419414  0.4476587911
#> [101]  2.9392397389  0.1616179413 -1.1209104498  1.6063645227 -0.5266637906
#> [106] -0.7151880933  0.4219413373 -1.4452563798  0.8457972032  0.9037155616
#> [111]  2.4276204293  1.0949999730  0.9506906465  3.5315050843 -0.2808583569
#> [116]  0.2280058668 -0.9998945037 -1.0693768842 -0.5242874161  1.7635116261
#> [121] -0.1582127295  0.2462147510 -0.1765200161 -0.3318872853 -0.2600707679
#> [126] -0.9053390621 -0.7494647018 -0.4668374176  1.8509227201 -0.0554791216
#> [131] -1.4089620957  1.0753989132 -0.3553361505  0.8961807837 -0.2042242961
#> [136]  0.7252603064  0.7544251144  2.7307873983 -1.0293118814 -1.3819124352
#> [141]  0.5760190326  0.6719972911  1.0452368997 -0.2704051303  2.5245548420
#> [146] -0.2124122237 -0.2124122237 -0.3194036553  0.6810923121  0.7764962805
#> [151] -0.7593775130  0.3990645053 -1.3183446236 -0.9493764882  1.2917634086
#> [156]  1.7074149101  0.3487890954 -0.4337409801 -1.0489729127 -0.1992867590
#> [161]  1.0043322673  2.0740667248 -0.6217139892  1.4358657037  0.4803744399
#> [166] -0.5308504311 -1.4376449959  3.1064933992 -1.2783677921 -1.6426742332
#> [171] -0.2359565995 -3.9121885900  3.2642632858 -0.5563866064  0.6003924164
#> [176]  1.0037542703 -1.2750355943  0.8438738324  2.3202182021  0.7918635611
#> [181] -0.9218821708 -1.0474785516 -2.2325122392 -3.3581616213  0.5317218079
#> [186]  0.4032608608  1.3604720668 -3.3561282159  1.9712539677  1.8716345953
#> [191] -0.0177994509 -1.0825927385 -0.6512815324 -0.8640657221 -0.3907428860
#> [196]  1.3349210528  1.2023984847  0.2742414286 -1.9119168261 -2.8286308103
#> [201] -1.4499710126  3.0364137878 -2.1598133571  2.5631166774 -0.1409462062
#> [206] -2.2954423262  2.6363375508 -0.9022372333  0.2774096686 -0.1068586328
#> [211] -1.2985197779  0.9437939436  0.5553428033  1.9851586621 -0.9904119889
#> [216] -0.0563924910  2.6897222006  0.2827752836  0.7627492159 -1.1967581282
#> [221]  0.2940454484  0.6377448649  0.1738037793  1.3052423211  0.4431293832
#> [226] -0.7814864816  1.7558719627 -3.5902347411  0.3107088172 -0.6051540927
#> [231]  0.1216756078  2.9684934500  0.1032372440  1.9743701526 -0.8142999004
#> [236] -0.8478049140  0.2901042218  1.4854935159 -3.2237845201  0.4183461065
#> [241] -0.4762595149  0.3930660464  0.6724240523  0.1233645684  0.6019008347
#> [246]  0.7359501397 -2.3055647268  2.1787001405 -0.3142205239 -0.5654871026
#> [251]  0.0768118227 -0.2889320645  0.2707525516 -3.7466909463 -0.5770779677
#> [256]  3.8249414300  0.0416559358  0.0979314351  0.8721986351  0.4831892942
#> [261]  2.1769242212 -0.5725909753 -1.9823929095 -0.1152542033  1.4729383662
#> [266] -0.6401410111  0.2896455151  1.0067290441  0.7353210213 -0.7830648277
#> [271]  0.1551828563 -0.0178930846  1.6351776272  0.9148198152 -0.3810707170
#> [276] -1.4064142030 -0.3158492450 -0.3353149295 -0.6539066002  0.0439606448
#> [281] -1.3911229198  0.0947694419  0.6592033453  0.8961117851  0.8453430296
#> [286] -0.7102999243 -1.2330867027  0.9444936090 -0.9321847029  1.7862387792
#> [291] -0.9897039153  2.1677676947  0.2952774532 -0.3043018708  0.0864741626
#> [296] -2.1633361748  0.9917750394  0.3364862040 -1.7488567017 -0.0591098165
#> [301]  1.0430022139  1.0491253280  0.8859224026  1.4156353369  0.7341331305
#> [306]  0.8397992334 -0.0188529163 -0.9851695735 -0.1178183489  0.1311747434
#> [311] -0.0189629160 -0.4838599499  0.9040132938  0.4608594118  0.9130828542
#> [316]  1.5882432832  1.9428446741  1.2370853295 -1.0867388645 -0.0548513735
#> [321] -0.2667661609  0.8963476471 -0.1654257834  0.4482842014 -0.2124122237
#> [326]  0.6047418962  1.7674089434  0.9645510247 -0.9752071713 -1.8836060049
#> [331] -0.4366155153  0.9586801115  1.3191883010  1.8165099927 -1.1636981911
#> [336] -2.5522143656 -0.4324828988  1.6265647913 -0.9004147511 -0.0186145187
#> [341] -0.7022828878  1.1994972409  0.1723880152 -1.6943578704 -1.8560354864
#> [346]  0.2233555849  0.6886026162 -0.3792683035  4.3753533624 -1.0149718074
#> [351] -0.7867108303  0.6983558389 -1.3069999256  0.0033064785  0.7498183133
#> [356] -1.0757084732  1.7181020228  0.7533612244  1.3993027621 -0.3570955905
#> [361] -0.3948264812 -0.4365818797 -0.9924304432  0.8420146207  1.4135056827
#> [366] -0.6757122916  3.3805549076  0.0754499195 -0.3367200865  0.4641085669
#> [371]  0.4174781103 -0.0318807461 -1.6992637998 -0.8790481055 -1.5050684123
#> [376]  1.0206171691 -0.9005541424  1.8722946152  1.8587348075  0.2225106170
#> [381] -0.9644044035  0.0768118227  0.4692328676 -1.0712742049  0.1565004689
#> [386] -1.4237593747  0.5818357676  1.3821693632 -0.7174321628  1.2276404883
#> [391] -1.2119990766 -0.2409267648  0.4659010264 -0.9854198292 -0.4822936395
#> [396] -1.2091195107  0.1365199696 -1.4536686382  2.7031636658  1.2040871514
#> [401]  0.5916033870  0.5995817123  0.0061210875  2.2624384101  0.0587552057
#> [406] -0.6557124526  2.0605332952  1.5057362624  0.2144810497 -2.5123759246
#> [411] -1.7304269850 -2.3793179474 -0.7180772774  1.1184196431  0.7124498472
#> [416]  1.5939748869  1.5209201217 -0.7232212286  0.9312523962  0.4580390160
#> [421]  1.3544257537  0.6160559635 -0.6785545053  3.8118442466 -1.1255997689
#> [426] -0.5238108224  0.8290149113  0.3815430977  1.1044265835  1.1982271349
#> [431]  2.8996725871 -0.8136075420  0.8898298148  0.8001842708  0.6309766860
#> [436]  3.4032383887 -1.6143553564  2.3110405182  3.4573614575 -1.9033622472
#> [441]  2.5110501222 -1.4301368166  1.1179148909  0.2969864813  0.3654867616
#> [446]  0.7843634867 -0.7705090104 -0.4104754304  0.3449095859 -0.0005578803
#> [451] -0.4337409801 -1.7923795378 -1.3657466267  1.3251107090 -1.0812864732
#> [456]  0.6774166484 -0.0980938236 -0.0654323267 -0.3316434181  0.8703017080
#> [461] -1.9114510190  0.8800643914 -0.5641048814  0.9306368094 -0.3894801347
#> [466] -1.7521867849 -2.3139293940 -1.5847479095  1.1570680429  0.0373464924
#> [471] -1.0978023657  0.6035222152 -0.1255777382  0.6387656800  1.3085995627
#> [476]  0.6701643532 -1.8804449994 -0.3922273506 -0.2101890754  0.3138900413
#> [481] -2.6429907746  3.4304255638  0.1293336320  0.5430238256  0.5542894090
#> [486]  1.8546036803 -0.4091280260 -0.8989867804 -0.3665570599  0.7986432448
#> [491]  0.6892764808 -1.7006226820 -0.2142358040 -0.9869753098 -1.4755745854
#> [496]  0.4961533727 -1.5469423218  0.6330250497  3.0126566411  2.1196381137
#> [501] -1.1989121952  1.3692638995  1.2887683893 -0.2727907779 -0.6622467684
#> [506]  1.3754972631 -0.0623980538  0.9491447268 -1.3043163524 -2.9729602464
#> [511]  0.5744312675  0.0705923671 -0.8136075420  0.6088258620  0.9193780134
#> [516] -0.3643409078 -2.2934644451  0.9698612266  2.6288573479 -0.2319036098
#> [521]  0.3242389172  0.0935991794  0.0786205825  2.2558848721 -2.6814963446
#> [526]  1.1076034739  0.7419494371 -0.8866928720  3.2284614455  0.9893224939
#> [531] -1.4382294013 -0.8866928720  0.6953817022  0.3317707542 -1.1733028177
#> [536]  0.5724598915 -0.2166372664 -1.1772792160  1.0807696530 -0.6142798139
#> [541] -1.8169970052  0.2269929252 -0.4607885668  1.4765401532  0.1191478117
#> [546] -0.3733359682  1.0805805653  1.1250044849  0.7596444649  0.5165394897
#> [551] -0.7059923576  0.4597168309  0.2681595146  0.6688430460 -3.0830927001
#> [556] -2.6207682030  2.5006831207  0.4405350291 -0.8420746817  0.0223439101
#> [561] -1.6845354011  1.4439977745 -3.9980349960 -0.7769638586  2.2981289275
#> [566]  0.7652622933 -0.0957138351  0.9077498799  0.0147671850 -1.2304141940
#> [571] -1.2732893615 -1.1552248477  1.3595556120 -1.4985808419 -0.5972245736
#> [576]  0.4898871506  2.5757812590 -0.5847215007  0.2265698744 -1.5791048030
#> [581]  0.0399566277  4.7400192487  0.3831836000  0.1841479805  1.4725861445
#> [586]  1.0297324850  1.4747131717 -0.5021496930  0.7490856320  0.5839962902
#> [591]  1.8555383941  0.3760819718 -0.0793326606  1.0069657210 -1.9662375275
#> [596]  0.1165606125  0.7962862589 -1.4884096243 -0.7562031156 -0.4456983945
#> [601] -0.4982968345  0.2103895218 -2.0412786216 -0.5411987227 -0.3835248474
#> [606]  0.6622251563 -1.1480547866 -2.2027884954 -1.1017240784  1.2282977229
#> [611] -1.6072569529  0.3341982223  1.4791757444 -0.0936422373  0.5897793293
#> [616]  1.1362091060 -0.4843870259  1.0570468952  1.7151545329 -1.4090211044
#> [621] -2.3288252089 -0.0617214925  0.5932077014 -1.5541364809  0.3516913750
#> [626]  1.2012946540 -0.9913567801  0.1927156436 -0.9295268879  2.2602491241
#> [631]  1.0214480011  0.0210571802 -0.0044176046 -1.1480168202  2.9523870249
#> [636]  0.2163478726  3.1897007904  0.8596208087  0.5514722867  0.1679797268
#> [641]  0.7349968315 -0.2248063350 -0.2386906385  0.0290607781  1.3252156372
#> [646] -1.0506258119  0.2795027597  0.1441383470  0.7517855804  0.8093572949
#> [651]  1.7932175096  0.6957881363  0.8811297040 -0.5044555679  1.2367741914
#> [656] -0.6407075492 -0.4988999124 -1.4168998979  1.0057779087 -0.9039531014
#> [661]  1.6793818078  0.5699633633  1.1742700619  0.4109899860 -0.2095158760
#> [666]  0.3394714862  0.5753345856  0.1804495923 -0.2344839340  0.7866511233
#> [671]  2.4731257870 -0.7523168490 -0.3733359682 -0.0199012933  0.6719195850
#> [676]  0.2797182813 -0.5028969866  3.9404934875  0.7946751264 -0.5224567629
#> [681]  1.1357396773  0.9231505259 -1.1518397589  0.4691624533  3.3700490910
#> [686]  4.2799830665  0.0048982575  0.4270913253  0.0514724353  0.3153233226
#> [691] -0.0908388955  1.1345990634 -0.1763354058  0.5198670342  0.2649421032
#> [696] -1.3105622482 -0.3270228611 -0.8885397601  0.9041033467 -2.3244983833
#> [701]  2.4215412279  3.0632545863 -1.0303890631  2.8730258495  0.9043725058
#> [706]  2.7311868970 -0.3368401088 -1.6492376784  0.5693944169  0.4223409670
#> [711] -0.5031664172 -0.6504227389  0.6456299628 -1.0530357982  2.1404312880
#> [716] -0.4693447928  0.2193041186 -1.9994236207  0.3161994359 -2.3288252089
#> [721] -2.2961887745  2.0306591253  2.4894878401 -1.8994569156  0.9690606083
#> [726] -1.2301981777 -0.1925274715 -0.4822936395  1.1233218699  1.0437503350
#> [731] -0.1666960175  0.8636371459  0.1279905369 -1.2116403633 -0.6070652901
#> [736]  2.4661066612  0.3547787355 -0.3468321123  0.3717413625 -0.3737450849
#> [741] -2.1058788758  3.3097165524 -0.7467483637  1.4791493712  0.9547210718
#> [746] -0.4143663154 -0.3189597765  1.3420393739 -1.2594086810  0.0305519032
#> [751]  2.9873742449 -0.1130092808  0.2660232281  1.3750965890 -1.0365537446
#> [756] -0.2015575947 -0.7498205958 -0.2509820301  0.5936935976  0.2459579103
#> [761] -0.4195047602 -0.7173137620  2.1173060473  0.0759366129 -0.3488559780
#> [766]  0.2190089694  0.7685427632  0.2841709259 -2.5422714886 -0.6999117191
#> [771]  1.0547214966 -0.9986273122 -1.1679100734  0.8217562084  0.2054257305
#> [776]  3.3194610343 -0.1147058043 -0.1075114710  0.8980072828  0.5983185785
#> [781] -1.4444749260 -0.8408941479 -0.3570955905  0.4091854843  1.3854175247
#> [786] -0.0823262342  0.0033053625  0.7227914443  2.1237894061  1.0008194769
#> [791] -1.2043076018 -0.4214335166  2.3481794063  0.4789702237 -0.5806151002
#> [796]  0.0095935352  0.5312461040 -0.2161992763 -0.0321136581  0.5691174290
#> [801] -1.8185177477 -1.7729238118  0.0086777138  0.6760042886 -0.5908980195
#> [806] -1.1883668056  0.2299997711 -0.5266637906  0.4529613076 -1.3849639827
#> [811] -3.8494997222 -1.9181144253  3.5139759675 -2.3958056927  2.6556186738
#> [816] -3.3060561647  0.5919350311  0.0696764244  0.2212809971 -0.4574372987
#> [821]  0.8195644973  1.2697263823 -0.6869443272  0.0553436773  0.6607270693
#> [826]  0.4042754336  1.2571283933  0.3533438175  1.6864468732 -0.0716847136
#> [831]  0.8410286460 -0.2613114217 -1.0595884630 -0.8721458396 -0.9744905824
#> [836]  0.8734250937 -1.0913054487  0.3760819718 -0.4877817014 -0.0801535222
#> [841]  0.5073620513  1.3754841800 -1.4955762507 -0.0607232209 -0.1294087257
#> [846]  0.1571206573 -1.7094134649 -0.7266681479  0.3639432785 -2.5314231177
#> [851]  0.3839549237  0.9274407231 -0.8266059321 -0.4540676077 -0.3998456629
#> [856]  1.5894866152  1.2755068495 -2.2325122392  0.0530392535  3.4311503879
#> [861] -0.4843870259 -0.4893118635 -0.7728203841  0.2968775111  0.1464355323
#> [866]  0.5874674600  0.1201329986 -0.7092773263 -0.4793195480  1.6312441373
#> [871] -1.0609871453  0.6494952423 -0.7374493814