Function to perform a series of LQAS analysis on simulated coverage survey data

Function to perform a series of LQAS analysis on simulated coverage survey data

simul_lqas(runs = 50, pop = NULL, n, d.lower, d.upper, p.lower = 0,
  p.upper = 100, fine = 1, progress = TRUE)

Arguments

runs
Number of simulation runs to perform per coverage proportion. Default is 50 runs

pop
Population size from which simulated coverage survey data is to be taken from

n
Sample size of actual or test coverage data

d.lower
A numeric value for the lower classification threshold

d.upper
A numeric value for the upper classification threshold

p.lower
Starting proportion for simulations. Default is 0

p.upper
Ending proportion for simulations. Default is 100

fine
Granularity of simulated proportiongs; Defaul to 1

progress
Logical. Should simulation progress be shown? Default is TRUE

Value

A data.frame of coverage proportions and LQAS outcomes

Examples

simul_lqas(runs = 10, pop = 10000, n = 40, d.lower = 60, d.upper = 90)

#> Running simulations for proportion := 0%
#> Running simulations for proportion := 1%
#> Running simulations for proportion := 2%
#> Running simulations for proportion := 3%
#> Running simulations for proportion := 4%
#> Running simulations for proportion := 5%
#> Running simulations for proportion := 6%
#> Running simulations for proportion := 7%
#> Running simulations for proportion := 8%
#> Running simulations for proportion := 9%
#> Running simulations for proportion := 10%
#> Running simulations for proportion := 11%
#> Running simulations for proportion := 12%
#> Running simulations for proportion := 13%
#> Running simulations for proportion := 14%
#> Running simulations for proportion := 15%
#> Running simulations for proportion := 16%
#> Running simulations for proportion := 17%
#> Running simulations for proportion := 18%
#> Running simulations for proportion := 19%
#> Running simulations for proportion := 20%
#> Running simulations for proportion := 21%
#> Running simulations for proportion := 22%
#> Running simulations for proportion := 23%
#> Running simulations for proportion := 24%
#> Running simulations for proportion := 25%
#> Running simulations for proportion := 26%
#> Running simulations for proportion := 27%
#> Running simulations for proportion := 28%
#> Running simulations for proportion := 29%
#> Running simulations for proportion := 30%
#> Running simulations for proportion := 31%
#> Running simulations for proportion := 32%
#> Running simulations for proportion := 33%
#> Running simulations for proportion := 34%
#> Running simulations for proportion := 35%
#> Running simulations for proportion := 36%
#> Running simulations for proportion := 37%
#> Running simulations for proportion := 38%
#> Running simulations for proportion := 39%
#> Running simulations for proportion := 40%
#> Running simulations for proportion := 41%
#> Running simulations for proportion := 42%
#> Running simulations for proportion := 43%
#> Running simulations for proportion := 44%
#> Running simulations for proportion := 45%
#> Running simulations for proportion := 46%
#> Running simulations for proportion := 47%
#> Running simulations for proportion := 48%
#> Running simulations for proportion := 49%
#> Running simulations for proportion := 50%
#> Running simulations for proportion := 51%
#> Running simulations for proportion := 52%
#> Running simulations for proportion := 53%
#> Running simulations for proportion := 54%
#> Running simulations for proportion := 55%
#> Running simulations for proportion := 56%
#> Running simulations for proportion := 57%
#> Running simulations for proportion := 58%
#> Running simulations for proportion := 59%
#> Running simulations for proportion := 60%
#> Running simulations for proportion := 61%
#> Running simulations for proportion := 62%
#> Running simulations for proportion := 63%
#> Running simulations for proportion := 64%
#> Running simulations for proportion := 65%
#> Running simulations for proportion := 66%
#> Running simulations for proportion := 67%
#> Running simulations for proportion := 68%
#> Running simulations for proportion := 69%
#> Running simulations for proportion := 70%
#> Running simulations for proportion := 71%
#> Running simulations for proportion := 72%
#> Running simulations for proportion := 73%
#> Running simulations for proportion := 74%
#> Running simulations for proportion := 75%
#> Running simulations for proportion := 76%
#> Running simulations for proportion := 77%
#> Running simulations for proportion := 78%
#> Running simulations for proportion := 79%
#> Running simulations for proportion := 80%
#> Running simulations for proportion := 81%
#> Running simulations for proportion := 82%
#> Running simulations for proportion := 83%
#> Running simulations for proportion := 84%
#> Running simulations for proportion := 85%
#> Running simulations for proportion := 86%
#> Running simulations for proportion := 87%
#> Running simulations for proportion := 88%
#> Running simulations for proportion := 89%
#> Running simulations for proportion := 90%
#> Running simulations for proportion := 91%
#> Running simulations for proportion := 92%
#> Running simulations for proportion := 93%
#> Running simulations for proportion := 94%
#> Running simulations for proportion := 95%
#> Running simulations for proportion := 96%
#> Running simulations for proportion := 97%
#> Running simulations for proportion := 98%
#> Running simulations for proportion := 99%
#> Running simulations for proportion := 100%

#>       d outcome proportion
#> 1     0       1          0
#> 2     0       1          0
#> 3     0       1          0
#> 4     0       1          0
#> 5     0       1          0
#> 6     0       1          0
#> 7     0       1          0
#> 8     0       1          0
#> 9     0       1          0
#> 10    0       1          0
#> 11    0       1          1
#> 12    1       1          1
#> 13    0       1          1
#> 14    2       1          1
#> 15    0       1          1
#> 16    0       1          1
#> 17    0       1          1
#> 18    0       1          1
#> 19    0       1          1
#> 20    0       1          1
#> 21    3       1          2
#> 22    1       1          2
#> 23    0       1          2
#> 24    0       1          2
#> 25    0       1          2
#> 26    0       1          2
#> 27    0       1          2
#> 28    2       1          2
#> 29    2       1          2
#> 30    1       1          2
#> 31    0       1          3
#> 32    3       1          3
#> 33    2       1          3
#> 34    2       1          3
#> 35    2       1          3
#> 36    1       1          3
#> 37    2       1          3
#> 38    0       1          3
#> 39    0       1          3
#> 40    0       1          3
#> 41    2       1          4
#> 42    5       1          4
#> 43    2       1          4
#> 44    3       1          4
#> 45    2       1          4
#> 46    0       1          4
#> 47    2       1          4
#> 48    1       1          4
#> 49    6       1          4
#> 50    1       1          4
#> 51    4       1          5
#> 52    3       1          5
#> 53    3       1          5
#> 54    1       1          5
#> 55    2       1          5
#> 56    1       1          5
#> 57    1       1          5
#> 58    3       1          5
#> 59    1       1          5
#> 60    1       1          5
#> 61    5       1          6
#> 62    2       1          6
#> 63    1       1          6
#> 64    4       1          6
#> 65    2       1          6
#> 66    2       1          6
#> 67    2       1          6
#> 68    2       1          6
#> 69    2       1          6
#> 70    1       1          6
#> 71    1       1          7
#> 72    0       1          7
#> 73    4       1          7
#> 74    1       1          7
#> 75    7       1          7
#> 76    2       1          7
#> 77    2       1          7
#> 78    3       1          7
#> 79    2       1          7
#> 80    4       1          7
#> 81    2       1          8
#> 82    2       1          8
#> 83    6       1          8
#> 84    6       1          8
#> 85    4       1          8
#> 86    4       1          8
#> 87    2       1          8
#> 88    3       1          8
#> 89    2       1          8
#> 90    5       1          8
#> 91    6       1          9
#> 92    4       1          9
#> 93    8       1          9
#> 94    3       1          9
#> 95    5       1          9
#> 96    3       1          9
#> 97    1       1          9
#> 98    4       1          9
#> 99    6       1          9
#> 100   4       1          9
#> 101   3       1         10
#> 102   2       1         10
#> 103   8       1         10
#> 104   7       1         10
#> 105   1       1         10
#> 106   4       1         10
#> 107   2       1         10
#> 108   5       1         10
#> 109   6       1         10
#> 110   2       1         10
#> 111   5       1         11
#> 112   2       1         11
#> 113   7       1         11
#> 114   4       1         11
#> 115   2       1         11
#> 116   4       1         11
#> 117   2       1         11
#> 118   4       1         11
#> 119   4       1         11
#> 120   8       1         11
#> 121   8       1         12
#> 122   2       1         12
#> 123   4       1         12
#> 124   5       1         12
#> 125   3       1         12
#> 126   7       1         12
#> 127   6       1         12
#> 128   5       1         12
#> 129   4       1         12
#> 130   6       1         12
#> 131   2       1         13
#> 132   5       1         13
#> 133   6       1         13
#> 134   5       1         13
#> 135   6       1         13
#> 136   5       1         13
#> 137   9       1         13
#> 138   7       1         13
#> 139   3       1         13
#> 140   9       1         13
#> 141   6       1         14
#> 142   7       1         14
#> 143   4       1         14
#> 144   3       1         14
#> 145   7       1         14
#> 146   7       1         14
#> 147   9       1         14
#> 148   3       1         14
#> 149   5       1         14
#> 150   6       1         14
#> 151   4       1         15
#> 152   6       1         15
#> 153   9       1         15
#> 154   6       1         15
#> 155   6       1         15
#> 156  10       1         15
#> 157  10       1         15
#> 158   6       1         15
#> 159   7       1         15
#> 160   7       1         15
#> 161   6       1         16
#> 162   9       1         16
#> 163   7       1         16
#> 164  10       1         16
#> 165   4       1         16
#> 166  11       1         16
#> 167   4       1         16
#> 168   6       1         16
#> 169   5       1         16
#> 170   5       1         16
#> 171  10       1         17
#> 172   9       1         17
#> 173  12       1         17
#> 174   8       1         17
#> 175   4       1         17
#> 176   9       1         17
#> 177   9       1         17
#> 178   9       1         17
#> 179  13       1         17
#> 180   4       1         17
#> 181   6       1         18
#> 182   9       1         18
#> 183  10       1         18
#> 184   1       1         18
#> 185   7       1         18
#> 186   6       1         18
#> 187   7       1         18
#> 188   7       1         18
#> 189   4       1         18
#> 190  10       1         18
#> 191  13       1         19
#> 192   9       1         19
#> 193  12       1         19
#> 194   6       1         19
#> 195   8       1         19
#> 196   7       1         19
#> 197  10       1         19
#> 198  11       1         19
#> 199   9       1         19
#> 200   8       1         19
#> 201   7       1         20
#> 202   5       1         20
#> 203  11       1         20
#> 204   4       1         20
#> 205   8       1         20
#> 206  11       1         20
#> 207   7       1         20
#> 208  13       1         20
#> 209   8       1         20
#> 210  14       1         20
#> 211  13       1         21
#> 212  10       1         21
#> 213   5       1         21
#> 214  12       1         21
#> 215   7       1         21
#> 216   6       1         21
#> 217   6       1         21
#> 218   7       1         21
#> 219   8       1         21
#> 220   7       1         21
#> 221   8       1         22
#> 222   7       1         22
#> 223   9       1         22
#> 224  12       1         22
#> 225   6       1         22
#> 226   8       1         22
#> 227  14       1         22
#> 228   9       1         22
#> 229  12       1         22
#> 230  10       1         22
#> 231   9       1         23
#> 232  15       1         23
#> 233  10       1         23
#> 234   8       1         23
#> 235  13       1         23
#> 236   9       1         23
#> 237   5       1         23
#> 238   3       1         23
#> 239   9       1         23
#> 240  10       1         23
#> 241   7       1         24
#> 242  10       1         24
#> 243   6       1         24
#> 244  13       1         24
#> 245   8       1         24
#> 246   5       1         24
#> 247   9       1         24
#> 248   9       1         24
#> 249  13       1         24
#> 250  11       1         24
#> 251   4       1         25
#> 252  12       1         25
#> 253  16       1         25
#> 254   8       1         25
#> 255  15       1         25
#> 256  11       1         25
#> 257  10       1         25
#> 258   9       1         25
#> 259   9       1         25
#> 260  10       1         25
#> 261  15       1         26
#> 262  11       1         26
#> 263  12       1         26
#> 264  10       1         26
#> 265   8       1         26
#> 266   5       1         26
#> 267  12       1         26
#> 268  10       1         26
#> 269   8       1         26
#> 270  11       1         26
#> 271  11       1         27
#> 272   8       1         27
#> 273  11       1         27
#> 274  12       1         27
#> 275   6       1         27
#> 276  10       1         27
#> 277  10       1         27
#> 278  12       1         27
#> 279  11       1         27
#> 280  10       1         27
#> 281   8       1         28
#> 282   9       1         28
#> 283   8       1         28
#> 284  12       1         28
#> 285   6       1         28
#> 286  15       1         28
#> 287   7       1         28
#> 288  15       1         28
#> 289  14       1         28
#> 290  12       1         28
#> 291  10       1         29
#> 292  17       1         29
#> 293  10       1         29
#> 294  14       1         29
#> 295  13       1         29
#> 296   9       1         29
#> 297   9       1         29
#> 298  14       1         29
#> 299  13       1         29
#> 300  17       1         29
#> 301  10       1         30
#> 302   8       1         30
#> 303  11       1         30
#> 304  11       1         30
#> 305  12       1         30
#> 306  15       1         30
#> 307  11       1         30
#> 308  13       1         30
#> 309  13       1         30
#> 310   6       1         30
#> 311  15       1         31
#> 312  19       1         31
#> 313  15       1         31
#> 314  13       1         31
#> 315  15       1         31
#> 316  12       1         31
#> 317  12       1         31
#> 318   8       1         31
#> 319  14       1         31
#> 320  13       1         31
#> 321  13       1         32
#> 322  16       1         32
#> 323  18       1         32
#> 324  14       1         32
#> 325  15       1         32
#> 326  12       1         32
#> 327  18       1         32
#> 328  16       1         32
#> 329  10       1         32
#> 330  11       1         32
#> 331  13       1         33
#> 332  13       1         33
#> 333  16       1         33
#> 334  14       1         33
#> 335  14       1         33
#> 336  13       1         33
#> 337  15       1         33
#> 338  13       1         33
#> 339  18       1         33
#> 340  15       1         33
#> 341  15       1         34
#> 342  15       1         34
#> 343  12       1         34
#> 344  16       1         34
#> 345  10       1         34
#> 346  11       1         34
#> 347  11       1         34
#> 348  12       1         34
#> 349  14       1         34
#> 350  11       1         34
#> 351  13       1         35
#> 352  17       1         35
#> 353   9       1         35
#> 354  22       1         35
#> 355  16       1         35
#> 356  15       1         35
#> 357  18       1         35
#> 358  10       1         35
#> 359   8       1         35
#> 360  14       1         35
#> 361  14       1         36
#> 362  19       1         36
#> 363  19       1         36
#> 364  14       1         36
#> 365  12       1         36
#> 366  14       1         36
#> 367  19       1         36
#> 368  11       1         36
#> 369  15       1         36
#> 370  12       1         36
#> 371  19       1         37
#> 372  13       1         37
#> 373  16       1         37
#> 374  17       1         37
#> 375  13       1         37
#> 376  14       1         37
#> 377  14       1         37
#> 378  12       1         37
#> 379  14       1         37
#> 380  16       1         37
#> 381  15       1         38
#> 382  17       1         38
#> 383  16       1         38
#> 384  16       1         38
#> 385  22       1         38
#> 386  15       1         38
#> 387  20       1         38
#> 388  11       1         38
#> 389  18       1         38
#> 390  13       1         38
#> 391  14       1         39
#> 392  23       1         39
#> 393  18       1         39
#> 394  16       1         39
#> 395  18       1         39
#> 396  15       1         39
#> 397  16       1         39
#> 398  18       1         39
#> 399  14       1         39
#> 400  18       1         39
#> 401  19       1         40
#> 402  19       1         40
#> 403  17       1         40
#> 404  16       1         40
#> 405   9       1         40
#> 406  11       1         40
#> 407  16       1         40
#> 408  12       1         40
#> 409  16       1         40
#> 410  18       1         40
#> 411  18       1         41
#> 412  16       1         41
#> 413  15       1         41
#> 414  13       1         41
#> 415  14       1         41
#> 416  13       1         41
#> 417   7       1         41
#> 418  14       1         41
#> 419  13       1         41
#> 420  18       1         41
#> 421  16       1         42
#> 422  16       1         42
#> 423  17       1         42
#> 424  14       1         42
#> 425  22       1         42
#> 426  19       1         42
#> 427  22       1         42
#> 428  20       1         42
#> 429  18       1         42
#> 430  14       1         42
#> 431  13       1         43
#> 432  21       1         43
#> 433  14       1         43
#> 434  15       1         43
#> 435  15       1         43
#> 436  14       1         43
#> 437  15       1         43
#> 438  24       1         43
#> 439  19       1         43
#> 440  19       1         43
#> 441  14       1         44
#> 442  22       1         44
#> 443  17       1         44
#> 444  16       1         44
#> 445  20       1         44
#> 446  18       1         44
#> 447  16       1         44
#> 448  17       1         44
#> 449  19       1         44
#> 450  19       1         44
#> 451  18       1         45
#> 452  20       1         45
#> 453  14       1         45
#> 454  12       1         45
#> 455  18       1         45
#> 456  18       1         45
#> 457  15       1         45
#> 458  25       2         45
#> 459  21       1         45
#> 460  19       1         45
#> 461  21       1         46
#> 462  21       1         46
#> 463  17       1         46
#> 464  18       1         46
#> 465  16       1         46
#> 466  20       1         46
#> 467  24       1         46
#> 468  16       1         46
#> 469  22       1         46
#> 470  21       1         46
#> 471  17       1         47
#> 472  16       1         47
#> 473  21       1         47
#> 474  25       2         47
#> 475  16       1         47
#> 476  15       1         47
#> 477  16       1         47
#> 478  18       1         47
#> 479  16       1         47
#> 480  20       1         47
#> 481  19       1         48
#> 482  16       1         48
#> 483  19       1         48
#> 484  17       1         48
#> 485  24       1         48
#> 486  19       1         48
#> 487  17       1         48
#> 488  23       1         48
#> 489  14       1         48
#> 490  22       1         48
#> 491  22       1         49
#> 492  19       1         49
#> 493  20       1         49
#> 494  24       1         49
#> 495  26       2         49
#> 496  20       1         49
#> 497  20       1         49
#> 498  17       1         49
#> 499  18       1         49
#> 500  22       1         49
#> 501  22       1         50
#> 502  25       2         50
#> 503  18       1         50
#> 504  15       1         50
#> 505  21       1         50
#> 506  21       1         50
#> 507  19       1         50
#> 508  20       1         50
#> 509  21       1         50
#> 510  21       1         50
#> 511  15       1         51
#> 512  20       1         51
#> 513  23       1         51
#> 514  17       1         51
#> 515  20       1         51
#> 516  21       1         51
#> 517  22       1         51
#> 518  20       1         51
#> 519  18       1         51
#> 520  19       1         51
#> 521  17       1         52
#> 522  23       1         52
#> 523  22       1         52
#> 524  23       1         52
#> 525  21       1         52
#> 526  21       1         52
#> 527  18       1         52
#> 528  23       1         52
#> 529  21       1         52
#> 530  23       1         52
#> 531  26       2         53
#> 532  22       1         53
#> 533  17       1         53
#> 534  18       1         53
#> 535  24       1         53
#> 536  24       1         53
#> 537  17       1         53
#> 538  20       1         53
#> 539  21       1         53
#> 540  17       1         53
#> 541  18       1         54
#> 542  15       1         54
#> 543  25       2         54
#> 544  19       1         54
#> 545  21       1         54
#> 546  24       1         54
#> 547  25       2         54
#> 548  16       1         54
#> 549  21       1         54
#> 550  16       1         54
#> 551  21       1         55
#> 552  23       1         55
#> 553  20       1         55
#> 554  26       2         55
#> 555  22       1         55
#> 556  22       1         55
#> 557  23       1         55
#> 558  23       1         55
#> 559  23       1         55
#> 560  19       1         55
#> 561  25       2         56
#> 562  20       1         56
#> 563  20       1         56
#> 564  23       1         56
#> 565  27       2         56
#> 566  20       1         56
#> 567  28       2         56
#> 568  29       2         56
#> 569  20       1         56
#> 570  25       2         56
#> 571  25       2         57
#> 572  26       2         57
#> 573  27       2         57
#> 574  23       1         57
#> 575  23       1         57
#> 576  26       2         57
#> 577  23       1         57
#> 578  18       1         57
#> 579  23       1         57
#> 580  26       2         57
#> 581  24       1         58
#> 582  24       1         58
#> 583  21       1         58
#> 584  24       1         58
#> 585  27       2         58
#> 586  22       1         58
#> 587  20       1         58
#> 588  21       1         58
#> 589  24       1         58
#> 590  22       1         58
#> 591  25       2         59
#> 592  27       2         59
#> 593  16       1         59
#> 594  26       2         59
#> 595  28       2         59
#> 596  22       1         59
#> 597  22       1         59
#> 598  21       1         59
#> 599  22       1         59
#> 600  26       2         59
#> 601  28       2         60
#> 602  22       1         60
#> 603  27       2         60
#> 604  25       2         60
#> 605  22       1         60
#> 606  20       1         60
#> 607  26       2         60
#> 608  25       2         60
#> 609  25       2         60
#> 610  24       1         60
#> 611  31       2         61
#> 612  27       2         61
#> 613  25       2         61
#> 614  25       2         61
#> 615  22       1         61
#> 616  26       2         61
#> 617  22       1         61
#> 618  25       2         61
#> 619  27       2         61
#> 620  23       1         61
#> 621  27       2         62
#> 622  25       2         62
#> 623  28       2         62
#> 624  21       1         62
#> 625  24       1         62
#> 626  21       1         62
#> 627  28       2         62
#> 628  24       1         62
#> 629  25       2         62
#> 630  27       2         62
#> 631  21       1         63
#> 632  27       2         63
#> 633  26       2         63
#> 634  28       2         63
#> 635  26       2         63
#> 636  30       2         63
#> 637  22       1         63
#> 638  24       1         63
#> 639  21       1         63
#> 640  26       2         63
#> 641  25       2         64
#> 642  26       2         64
#> 643  27       2         64
#> 644  28       2         64
#> 645  23       1         64
#> 646  25       2         64
#> 647  26       2         64
#> 648  27       2         64
#> 649  28       2         64
#> 650  26       2         64
#> 651  25       2         65
#> 652  24       1         65
#> 653  23       1         65
#> 654  32       2         65
#> 655  26       2         65
#> 656  27       2         65
#> 657  28       2         65
#> 658  25       2         65
#> 659  30       2         65
#> 660  26       2         65
#> 661  25       2         66
#> 662  26       2         66
#> 663  25       2         66
#> 664  28       2         66
#> 665  27       2         66
#> 666  27       2         66
#> 667  25       2         66
#> 668  21       1         66
#> 669  28       2         66
#> 670  21       1         66
#> 671  19       1         67
#> 672  27       2         67
#> 673  25       2         67
#> 674  26       2         67
#> 675  24       1         67
#> 676  30       2         67
#> 677  30       2         67
#> 678  27       2         67
#> 679  29       2         67
#> 680  27       2         67
#> 681  23       1         68
#> 682  30       2         68
#> 683  29       2         68
#> 684  26       2         68
#> 685  29       2         68
#> 686  28       2         68
#> 687  23       1         68
#> 688  27       2         68
#> 689  27       2         68
#> 690  29       2         68
#> 691  27       2         69
#> 692  24       1         69
#> 693  24       1         69
#> 694  29       2         69
#> 695  31       2         69
#> 696  25       2         69
#> 697  30       2         69
#> 698  32       2         69
#> 699  27       2         69
#> 700  27       2         69
#> 701  30       2         70
#> 702  25       2         70
#> 703  33       2         70
#> 704  28       2         70
#> 705  28       2         70
#> 706  26       2         70
#> 707  28       2         70
#> 708  27       2         70
#> 709  30       2         70
#> 710  27       2         70
#> 711  25       2         71
#> 712  28       2         71
#> 713  26       2         71
#> 714  31       2         71
#> 715  23       1         71
#> 716  28       2         71
#> 717  30       2         71
#> 718  26       2         71
#> 719  29       2         71
#> 720  27       2         71
#> 721  31       2         72
#> 722  31       2         72
#> 723  28       2         72
#> 724  33       2         72
#> 725  21       1         72
#> 726  31       2         72
#> 727  28       2         72
#> 728  30       2         72
#> 729  25       2         72
#> 730  22       1         72
#> 731  29       2         73
#> 732  29       2         73
#> 733  28       2         73
#> 734  32       2         73
#> 735  27       2         73
#> 736  32       2         73
#> 737  33       2         73
#> 738  31       2         73
#> 739  29       2         73
#> 740  29       2         73
#> 741  34       2         74
#> 742  31       2         74
#> 743  29       2         74
#> 744  25       2         74
#> 745  29       2         74
#> 746  28       2         74
#> 747  36       2         74
#> 748  30       2         74
#> 749  29       2         74
#> 750  30       2         74
#> 751  32       2         75
#> 752  35       2         75
#> 753  29       2         75
#> 754  31       2         75
#> 755  28       2         75
#> 756  34       2         75
#> 757  29       2         75
#> 758  30       2         75
#> 759  30       2         75
#> 760  29       2         75
#> 761  35       2         76
#> 762  33       2         76
#> 763  30       2         76
#> 764  27       2         76
#> 765  27       2         76
#> 766  32       2         76
#> 767  33       2         76
#> 768  30       2         76
#> 769  32       2         76
#> 770  36       2         76
#> 771  36       2         77
#> 772  33       2         77
#> 773  32       2         77
#> 774  32       2         77
#> 775  32       2         77
#> 776  31       2         77
#> 777  27       2         77
#> 778  29       2         77
#> 779  33       2         77
#> 780  33       2         77
#> 781  32       2         78
#> 782  31       2         78
#> 783  31       2         78
#> 784  29       2         78
#> 785  31       2         78
#> 786  30       2         78
#> 787  30       2         78
#> 788  32       2         78
#> 789  34       2         78
#> 790  28       2         78
#> 791  31       2         79
#> 792  34       2         79
#> 793  30       2         79
#> 794  32       2         79
#> 795  36       2         79
#> 796  32       2         79
#> 797  34       2         79
#> 798  32       2         79
#> 799  31       2         79
#> 800  30       2         79
#> 801  33       2         80
#> 802  32       2         80
#> 803  32       2         80
#> 804  33       2         80
#> 805  29       2         80
#> 806  33       2         80
#> 807  31       2         80
#> 808  31       2         80
#> 809  32       2         80
#> 810  30       2         80
#> 811  33       2         81
#> 812  35       2         81
#> 813  33       2         81
#> 814  35       2         81
#> 815  31       2         81
#> 816  32       2         81
#> 817  30       2         81
#> 818  31       2         81
#> 819  32       2         81
#> 820  28       2         81
#> 821  33       2         82
#> 822  35       2         82
#> 823  33       2         82
#> 824  35       2         82
#> 825  33       2         82
#> 826  34       2         82
#> 827  30       2         82
#> 828  33       2         82
#> 829  33       2         82
#> 830  30       2         82
#> 831  32       2         83
#> 832  34       2         83
#> 833  30       2         83
#> 834  36       2         83
#> 835  33       2         83
#> 836  31       2         83
#> 837  35       2         83
#> 838  34       2         83
#> 839  25       2         83
#> 840  33       2         83
#> 841  33       2         84
#> 842  28       2         84
#> 843  40       3         84
#> 844  34       2         84
#> 845  34       2         84
#> 846  35       2         84
#> 847  35       2         84
#> 848  36       2         84
#> 849  35       2         84
#> 850  35       2         84
#> 851  35       2         85
#> 852  33       2         85
#> 853  36       2         85
#> 854  34       2         85
#> 855  33       2         85
#> 856  30       2         85
#> 857  32       2         85
#> 858  33       2         85
#> 859  36       2         85
#> 860  33       2         85
#> 861  38       3         86
#> 862  35       2         86
#> 863  31       2         86
#> 864  33       2         86
#> 865  35       2         86
#> 866  35       2         86
#> 867  33       2         86
#> 868  35       2         86
#> 869  36       2         86
#> 870  34       2         86
#> 871  36       2         87
#> 872  36       2         87
#> 873  36       2         87
#> 874  36       2         87
#> 875  33       2         87
#> 876  32       2         87
#> 877  36       2         87
#> 878  35       2         87
#> 879  35       2         87
#> 880  34       2         87
#> 881  38       3         88
#> 882  37       3         88
#> 883  36       2         88
#> 884  35       2         88
#> 885  35       2         88
#> 886  33       2         88
#> 887  33       2         88
#> 888  37       3         88
#> 889  38       3         88
#> 890  35       2         88
#> 891  36       2         89
#> 892  35       2         89
#> 893  31       2         89
#> 894  36       2         89
#> 895  37       3         89
#> 896  36       2         89
#> 897  34       2         89
#> 898  39       3         89
#> 899  35       2         89
#> 900  33       2         89
#> 901  37       3         90
#> 902  38       3         90
#> 903  37       3         90
#> 904  39       3         90
#> 905  34       2         90
#> 906  37       3         90
#> 907  35       2         90
#> 908  37       3         90
#> 909  34       2         90
#> 910  38       3         90
#> 911  37       3         91
#> 912  37       3         91
#> 913  37       3         91
#> 914  38       3         91
#> 915  37       3         91
#> 916  37       3         91
#> 917  39       3         91
#> 918  37       3         91
#> 919  37       3         91
#> 920  36       2         91
#> 921  37       3         92
#> 922  36       2         92
#> 923  36       2         92
#> 924  40       3         92
#> 925  38       3         92
#> 926  37       3         92
#> 927  38       3         92
#> 928  35       2         92
#> 929  39       3         92
#> 930  35       2         92
#> 931  36       2         93
#> 932  37       3         93
#> 933  39       3         93
#> 934  37       3         93
#> 935  34       2         93
#> 936  38       3         93
#> 937  35       2         93
#> 938  37       3         93
#> 939  36       2         93
#> 940  38       3         93
#> 941  36       2         94
#> 942  38       3         94
#> 943  39       3         94
#> 944  39       3         94
#> 945  37       3         94
#> 946  40       3         94
#> 947  38       3         94
#> 948  34       2         94
#> 949  38       3         94
#> 950  38       3         94
#> 951  37       3         95
#> 952  39       3         95
#> 953  39       3         95
#> 954  38       3         95
#> 955  36       2         95
#> 956  36       2         95
#> 957  32       2         95
#> 958  39       3         95
#> 959  38       3         95
#> 960  39       3         95
#> 961  37       3         96
#> 962  39       3         96
#> 963  39       3         96
#> 964  38       3         96
#> 965  38       3         96
#> 966  38       3         96
#> 967  40       3         96
#> 968  38       3         96
#> 969  38       3         96
#> 970  39       3         96
#> 971  39       3         97
#> 972  40       3         97
#> 973  38       3         97
#> 974  40       3         97
#> 975  38       3         97
#> 976  39       3         97
#> 977  38       3         97
#> 978  39       3         97
#> 979  40       3         97
#> 980  40       3         97
#> 981  39       3         98
#> 982  39       3         98
#> 983  39       3         98
#> 984  38       3         98
#> 985  39       3         98
#> 986  37       3         98
#> 987  39       3         98
#> 988  39       3         98
#> 989  40       3         98
#> 990  39       3         98
#> 991  40       3         99
#> 992  40       3         99
#> 993  38       3         99
#> 994  39       3         99
#> 995  40       3         99
#> 996  40       3         99
#> 997  38       3         99
#> 998  40       3         99
#> 999  40       3         99
#> 1000 40       3         99
#> 1001 40       3        100
#> 1002 40       3        100
#> 1003 40       3        100
#> 1004 40       3        100
#> 1005 40       3        100
#> 1006 40       3        100
#> 1007 40       3        100
#> 1008 40       3        100
#> 1009 40       3        100
#> 1010 40       3        100