Calculate WHO Growth Reference z-score for given anthropometric measurement/s.

The first function, getWGSR(), is usually called by the addWGSR() function but could be used as a stand-alone calculator for getting z-score for a given anthropometric measurement.

getWGSR(sex, firstPart, secondPart, thirdPart = NA, index = NA, standing = NA)

addWGSR(
  data,
  sex,
  firstPart,
  secondPart,
  thirdPart = NA,
  index = NA,
  standing = NULL,
  output = paste(index, "z", sep = ""),
  digits = 2
)

Arguments

sex
Name of variable specifying the sex of the subject. This must be coded as 1 = male and 2 = female. Give a quoted variable name as in (e.g.) "sex".

firstPart
Name of variable specifying:

Weight (kg) for BMI/A, W/A, W/H, or W/L

Head circumference (cm) for HC/A

Height (cm) for BMI/A for H/A

Length (cm) for L/A

MUAC (cm) for MUAC/A

Sub-scapular skinfold (mm) for SSF/A

Triceps skinfold (mm) for TSF/A

Give a quoted variable name as in (e.g.) "weight". Be careful with units (weight in kg; height, length, head circumference, and MUAC in cm, skinfolds in mm).

secondPart
Name of variable specifying:

Age (days) for H/A, HC/A, L/A, MUAC/A, SSF/A, or TSF/A

Height (cm) BMI/A or W/H

Length (cm) for W/L

Give a quoted variable name as in (e.g.) "age". Be careful with units (age in days; height and length in cm).

thirdPart
Name of variable specifying age (in days) for BMI/A. Give a quoted variable name as in (e.g.) "age". Be careful with units (age in days).

index
The index to be calculated and added to data. One of:

| :--- | :--- | | bfa | BMI for age | | hca | Head circumference for age | | hfa | Height for age | | lfa | Length for age | | mfa | MUAC for age | | ssa | Sub-scapular skinfold for age | | tsa | Triceps skinfold for age | | wfa | Weight for age | | wfh | Weight for height | | wfl | Weight for length |

Give a quoted index name as in (e.g.) "wfh".

standing
Variable specifying how stature was measured. If NULL then age (for "hfa" or "lfa") or height rules (for "wfh" or "wfl") will be applied. This must be coded as 1 = Standing; 2 = Supine; 3 = Unknown. All other values will be recoded to 3 = Unknown. Give a quoted variable name as in (e.g.) "measured" or a single value (e.g. "measured = 1). If no value (or NULL) is specified then height and age rules will be applied.

data
A survey dataset as a data.frame object

output
The name of the column containing the specified index to be added to the dataset. This is an optional parameter. If you do not specify a value for output then the added column will take the name of the specified index with a "z" appended.

digits
The number of decimal places for output. Defaults to 2 d.p.

Value

A data.frame of the survey dataset with the calculated z-scores added.

Details

addWGSR() adds the WHO Growth Reference z-scores to a data frame of anthropometric data for weight, height or length, MUAC, head circumference, sub-scapular skinfold, triceps skinfold, and body mass index (BMI).

Examples

# Given a male child 10 months old with a weight of 5.7 kgs, height of 64.2
# cms, and MUAC of 125 mm:
#
# Calculate weight-for-height
getWGSR(sex = 1,
        firstPart = 5.7,
        secondPart = 64.2,
        index = "wfh",
        standing = 3)

#> [1] -2.725778

# calculate weight-for-age
getWGSR(sex = 1,
        firstPart = 5.7,
        secondPart = 10,
        index = "wfa",
        standing = 3)

#> [1] 3.452888

# calculate height-for-age
getWGSR(sex = 1,
        firstPart = 64.2,
        secondPart = 10,
        index = "hfa",
        standing = 3)

#> [1] 6.584276

# Calculate MUAC-for-age z-score for a girl
getWGSR(sex = 1,
        firstPart = 20,
        secondPart = 62 * (365.25 / 12),
        index = "mfa")

#> [1] 1.992728

# Calculate weight-for-height (wfh) for the anthro3 dataset
addWGSR(data = anthro3,
        sex = "sex",
        firstPart = "weight",
        secondPart = "height",
        index = "wfh")

#> ================================================================================

#>     psu age sex weight height muac oedema  wfhz
#> 1     1  10   1    5.7   64.2  125      2 -2.73
#> 2     1  10   2    5.8   64.4  121      2 -2.04
#> 3     1   9   2    6.5   62.2  139      2  0.13
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2 -3.44
#> 6     1  12   2    6.6   69.4  126      2 -2.26
#> 7     1   9   2    7.0   66.7  136      2 -0.71
#> 8     1   7   1    7.1   63.5  139      2  0.34
#> 9     1   9   2    7.1   66.2  144      2 -0.39
#> 10    1  16   2    7.2   69.0  131      2 -1.12
#> 11    1  13   2    7.4   68.2  137      2 -0.57
#> 12    1  11   2    7.4   70.0  132      2 -1.10
#> 13    1  14   2    7.5   64.8  125      2  0.69
#> 14    2   8   2    7.5   65.7  151      2  0.38
#> 15    2   8   1    7.5   69.1  140      2 -1.13
#> 16    2  13   1    7.6   69.2  131      2 -1.00
#> 17    2  13   2    7.6   70.0  125      2 -0.80
#> 18    2  22   2    7.6   76.2  121      2 -2.43
#> 19    2  16   1    7.7   69.6  136      2 -0.97
#> 20    2   8   2    7.8   63.3  144      2  1.64
#> 21    2  11   2    7.8   71.5  134      2 -0.92
#> 22    2  14   2    7.8   76.6  124      2 -2.21
#> 23    2   7   1    8.0   66.8  139      2  0.48
#> 24    2   8   1    8.0   69.3  147      2 -0.40
#> 25    2  24   2    8.2   70.7  125      2 -0.14
#> 26    2  10   1    8.2   71.1  145      2 -0.68
#> 27    3  15   1    8.3   70.4  133      2 -0.31
#> 28    3  14   2    8.3   71.5  137      2 -0.22
#> 29    3  14   2    8.3   72.9  141      2 -0.59
#> 30    3  17   2    8.3   73.5  136      2 -0.74
#> 31    3  15   1    8.3   76.1  133      2 -1.96
#> 32    3  13   2    8.4   71.8  143      2 -0.17
#> 33    3  20   1    8.4   77.8  132      2 -2.22
#> 34    3  29   2    8.5   75.6  127      2 -0.97
#> 35    3  21   2    8.5   75.6  133      2 -0.97
#> 36    3  21   1    8.5   76.7  123      2 -1.81
#> 37    3  17   1    8.5   77.6  137      2 -2.02
#> 38    3  15   2    8.6   74.9  134      2 -0.67
#> 39    3  24   1    8.6   77.8  137      2 -1.92
#> 40    3  12   1    8.7   72.1  156      2 -0.26
#> 41    4  27   2    8.7   75.8  127      2 -0.75
#> 42    4  25   2    8.7   80.8  129      2 -1.90
#> 43    4  25   1    8.8   81.4  125      2 -2.46
#> 44    4  24   2    8.9   74.3  142      2 -0.15
#> 45    4  30   1    9.0   77.6  132      2 -1.30
#> 46    4  22   1    9.0   82.2  135      2 -2.36
#> 47    4  16   1    9.1   76.4  145      2 -0.88
#> 48    4  24   1    9.1   80.0  126      2 -1.70
#> 49    4  17   2    9.2   74.8  144      2  0.11
#> 50    4  15   2    9.2   75.3  145      2 -0.01
#> 51    4  33   1    9.2   80.1  136      2 -1.59
#> 52    4  25   2    9.2   80.9  142      2 -1.26
#> 53    5  16   1    9.4   75.0  148      2 -0.13
#> 54    5  29   2    9.4   81.9  138      2 -1.26
#> 55    5  16   1    9.5   73.2  151      2  0.48
#> 56    5  17   1    9.6   73.5  146      2  0.52
#> 57    5  29   2    9.6   84.6  134      2 -1.70
#> 58    5  16   1    9.7   74.2  148      2  0.45
#> 59    5  20   1    9.7   80.4  142      2 -0.99
#> 60    5  34   1    9.8   80.2  138      2 -0.82
#> 61    5  21   2    9.8   80.9  145      2 -0.54
#> 62    5  20   1    9.8   82.2  139      2 -1.28
#> 63    5  29   1    9.9   78.6  142      2 -0.35
#> 64    5  25   2    9.9   82.9  138      2 -0.90
#> 65    5  15   1   10.0   75.3  157      2  0.53
#> 66    6  39   1   10.0   81.4  133      2 -0.84
#> 67    6   9   1   10.1   72.6  155      2  1.37
#> 68    6  21   1   10.1   76.7  153      2  0.31
#> 69    6  17   2   10.1   79.1  143      2  0.21
#> 70    6  25   1   10.1   79.5  140      2 -0.30
#> 71    6  13   2   10.2   74.6  157      2  1.27
#> 72    6  25   2   10.3   81.3  141      2 -0.07
#> 73    6  31   2   10.3   85.1  130      2 -1.00
#> 74    6  25   2   10.4   81.3  138      2  0.04
#> 75    6  37   1   10.6   76.2  153      2  1.00
#> 76    6  25   2   10.6   80.7  138      2  0.39
#> 77    6  26   1   10.6   83.1  142      2 -0.52
#> 78    6  36   2   10.6   84.2  135      2 -0.44
#> 79    6  37   2   10.7   84.2  150      2 -0.34
#> 80    7  34   2   10.7   85.7  146      2 -0.71
#> 81    7  47   1   10.7   89.8  131      2 -2.34
#> 82    7  12   1   10.8   82.0  147      2 -0.04
#> 83    7  25   2   10.8   83.6  141      2 -0.08
#> 84    7  31   1   10.8   83.8  158      2 -0.47
#> 85    7  34   1   10.8   84.4  144      2 -0.62
#> 86    7  27   2   10.9   79.0  153      2  1.06
#> 87    7  23   2   10.9   85.9  145      2 -0.55
#> 88    7  19   1   11.0   75.4  168      2  1.62
#> 89    7  26   2   11.0   83.2  145      2  0.22
#> 90    7  54   1   11.0   85.9  139      2 -0.79
#> 91    7  20   1   11.1   79.9  156      2  0.74
#> 92    7  23   1   11.1   84.7  146      2 -0.36
#> 93    7  43   2   11.1   86.0  142      2 -0.37
#> 94    8  31   1   11.1   87.1  143      2 -1.18
#> 95    8  40   2   11.2   95.2  137      2 -2.54
#> 96    8  18   2   11.3   75.5  161      2  2.15
#> 97    8  41   1   11.3   89.4  142      2 -1.53
#> 98    8  40   1   11.3   90.7  148      2 -1.84
#> 99    8  33   2   11.3   91.2  139      2 -1.56
#> 100   8  11   1   11.4   78.4  171      2  1.36
#> 101   8  25   1   11.4   82.2  163      2  0.57
#> 102   8  27   2   11.4   86.8  155      2 -0.26
#> 103   8  38   1   11.4   89.4  141      2 -1.41
#> 104   8  17   2   11.5   80.6  158      2  1.30
#> 105   8  31   2   11.5   85.4  153      2  0.18
#> 106   9  45   1   11.5   86.2  143      2 -0.31
#> 107   9  53   1   11.6   81.4  153      2  0.95
#> 108   9  23   2   11.6   86.3  143      2  0.05
#> 109   9  37   1   11.6   86.3  149      2 -0.23
#> 110   9  38   1   11.6   89.9  146      2 -1.31
#> 111   9  40   1   11.7   93.0  140      2 -1.91
#> 112   9  36   1   11.8   87.1  146      2 -0.40
#> 113   9  29   1   11.9   84.3  166      2  0.59
#> 114   9  26   2   12.0   81.2  164      2  1.62
#> 115   9  33   1   12.0   86.0  154      2  0.26
#> 116   9  36   1   12.0   87.5  150      2 -0.29
#> 117   9  55   2   12.0   96.2  144      2 -1.93
#> 118  10  16   1   12.1   82.3  157      2  1.25
#> 119  10  35   2   12.1   86.6  147      2  0.45
#> 120  10  43   1   12.1   91.7  151      2 -1.18
#> 121  10  38   2   12.2   80.9  175      2  1.86
#> 122  10  37   1   12.2   85.5  162      2  0.59
#> 123  10  41   2   12.2   87.6  155      2  0.14
#> 124  10  45   2   12.2   90.1  149      2 -0.43
#> 125  10  46   2   12.2   99.6  146      2 -2.46
#> 126  10  31   1   12.3   88.0  167      2 -0.11
#> 127  10  44   1   12.3   88.1  151      2 -0.13
#> 128  10  47   2   12.3   94.9  133      2 -1.37
#> 129  10  29   1   12.4   91.8  141      2 -0.89
#> 130  10  46   2   12.4   99.6  140      2 -2.26
#> 131  10  34   1   12.6   87.0  156      2  0.43
#> 132  10  38   2   12.7   85.7  152      2  1.20
#> 133  11  32   1   12.8   87.6  161      2  0.48
#> 134  11  36   1   12.8   90.4  150      2 -0.17
#> 135  11  39   1   12.8   92.7  152      2 -0.69
#> 136  11  42   1   12.8   94.0  146      2 -0.97
#> 137  11  38   2   12.9   83.4  174      2  1.90
#> 138  11  26   2   12.9   86.2  159      2  1.25
#> 139  11  35   1   12.9   90.9  154      2 -0.19
#> 140  11  41   1   12.9   91.2  160      2 -0.26
#> 141  11  32   1   12.9   91.4  148      2 -0.30
#> 142  11  49   1   13.0   93.1  152      2 -0.58
#> 143  11  30   1   13.1   84.3  161      2  1.72
#> 144  11  32   1   13.1   85.1  136      2  1.53
#> 145  11  35   1   13.1   87.5  158      2  0.78
#> 146  11  34   1   13.1   88.6  162      2  0.52
#> 147  11  45   2   13.1   91.6  150      2  0.05
#> 148  12  29   1   13.2   90.8  158      2  0.12
#> 149  12  47   2   13.2   91.4  146      2  0.18
#> 150  12  37   2   13.2   93.8  155      2 -0.33
#> 151  12  35   1   13.3   96.5  142      2 -1.02
#> 152  12  15   2   13.4   82.4  157      2  2.51
#> 153  12  38   2   13.4   93.0  152      2  0.01
#> 154  12  37   2   13.4   94.1  149      2 -0.22
#> 155  12  53   2   13.4  100.3  145      2 -1.50
#> 156  12  37   2   13.5   84.4  168      2  2.14
#> 157  12  43   1   13.5   96.1  151      2 -0.74
#> 158  12  52   1   13.5   97.0  144      2 -0.94
#> 159  12  45   2   13.5   98.0  148      2 -0.94
#> 160  12  39   2   13.5   99.1  145      2 -1.16
#> 161  12  47   2   13.6   97.0  146      2 -0.65
#> 162  13  26   2   13.7   85.7  171      2  2.00
#> 163  13  24   1   13.8   80.9  172      2  3.02
#> 164  13  38   2   13.8   93.3  153      2  0.27
#> 165  13  57   1   13.8   94.0  149      2 -0.02
#> 166  13  42   2   13.8  103.4  144      2 -1.84
#> 167  13  42   2   13.9   92.3  175      2  0.56
#> 168  13  52   1   13.9   95.5  151      2 -0.25
#> 169  13  52   1   13.9   95.7  156      2 -0.29
#> 170  13  57   2   13.9   97.2  153      2 -0.44
#> 171  13  52   2   13.9   97.8  141      2 -0.56
#> 172  13  52   2   13.9   97.9  155      2 -0.58
#> 173  13  49   2   13.9   98.2  150      2 -0.64
#> 174  13  58   2   14.0   96.6  154      2 -0.24
#> 175  13  50   1   14.0   97.9  158      2 -0.68
#> 176  13  39   2   14.1   93.7  155      2  0.43
#> 177  14  49   2   14.1   96.4  140      2 -0.12
#> 178  14  50   2   14.1   97.1  148      2 -0.26
#> 179  14  59   2   14.1   97.3  152      2 -0.30
#> 180  14  52   2   14.1  102.4  156      2 -1.37
#> 181  14  41   1   14.2   96.7  146      2 -0.24
#> 182  14  53   1   14.2   97.7  151      2 -0.46
#> 183  14  49   2   14.2   98.1  156      2 -0.38
#> 184  14  35   2   14.3   89.8  165      2  1.39
#> 185  14  50   1   14.3   93.7  152      2  0.48
#> 186  14  41   1   14.4   94.8  158      2  0.34
#> 187  14  42   1   14.5   95.1  161      2  0.36
#> 188  14  58   1   14.5  104.9  130      2 -1.79
#> 189  14  49   1   14.7   96.6  152      2  0.21
#> 190  14  52   2   14.7   96.7  154      2  0.28
#> 191  14  51   2   14.7   99.2  155      2 -0.22
#> 192  14  49   1   14.7   99.6  164      2 -0.45
#> 193  15  42   2   15.0   93.5  172      2  1.14
#> 194  15  28   2   15.1   91.5  164      2  1.62
#> 195  15  41   2   15.2   95.9  165      2  0.81
#> 196  15  53   1   15.3   95.0  166      2  1.03
#> 197  15  58   1   15.3  100.7  155      2 -0.20
#> 198  15  48   1   15.3  103.5  143      2 -0.82
#> 199  15  54   1   15.5   97.9  159      2  0.57
#> 200  15  58   1   15.5  101.3  166      2 -0.18
#> 201  15  52   2   15.6  100.0  164      2  0.27
#> 202  15  50   1   15.7   99.6  153      2  0.36
#> 203  15  53   2   15.7  103.1  158      2 -0.32
#> 204  15  43   1   15.8   91.1  179      2  2.22
#> 205  15  57   1   15.9  103.8  161      2 -0.42
#> 206  15  55   2   16.3   95.4  169      2  1.65
#> 207  16  54   2   15.3  102.0  156      2 -0.37
#> 208  16  44   1   16.3   96.2  173      2  1.54
#> 209  16  47   1   16.3  102.0  166      2  0.27
#> 210  16  57   1   16.3  108.4  138      2 -1.13
#> 211  16  52   1   16.4  103.9  152      2 -0.07
#> 212  16  52   2   16.6   97.8  144      2  1.37
#> 213  16  56   1   16.6  103.9  148      2  0.07
#> 214  16  55   1   16.7  106.3  154      2 -0.39
#> 215  16  52   1   17.0  101.3  163      2  0.93
#> 216  16  50   1   17.3  101.8  168      2  1.02
#> 217  16  53   1   17.5  102.2  168      2  1.06
#> 218  16  42   1   17.7  100.9  145      2  1.48
#> 219  16  48   2   17.8  111.3  176      2 -0.77
#> 220  16  53   1   17.9   98.7  171      2  2.10
#> 221  16  50   1   18.1  106.4  166      2  0.53

# Calculate weight-for-age (wfa) for the anthro3 dataset
addWGSR(data = anthro3,
        sex = "sex",
        firstPart = "weight",
        secondPart = "age",
        index = "wfa")

#> ================================================================================

#>     psu age sex weight height muac oedema  wfaz
#> 1     1  10   1    5.7   64.2  125      2  3.45
#> 2     1  10   2    5.8   64.4  121      2  3.95
#> 3     1   9   2    6.5   62.2  139      2  5.12
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2  3.82
#> 6     1  12   2    6.6   69.4  126      2  5.01
#> 7     1   9   2    7.0   66.7  136      2  5.91
#> 8     1   7   1    7.1   63.5  139      2  5.84
#> 9     1   9   2    7.1   66.2  144      2  6.06
#> 10    1  16   2    7.2   69.0  131      2  5.53
#> 11    1  13   2    7.4   68.2  137      2  6.13
#> 12    1  11   2    7.4   70.0  132      2  6.33
#> 13    1  14   2    7.5   64.8  125      2  6.18
#> 14    2   8   2    7.5   65.7  151      2  6.79
#> 15    2   8   1    7.5   69.1  140      2  6.36
#> 16    2  13   1    7.6   69.2  131      2  6.02
#> 17    2  13   2    7.6   70.0  125      2  6.43
#> 18    2  22   2    7.6   76.2  121      2  5.52
#> 19    2  16   1    7.7   69.6  136      2  5.84
#> 20    2   8   2    7.8   63.3  144      2  7.26
#> 21    2  11   2    7.8   71.5  134      2  6.95
#> 22    2  14   2    7.8   76.6  124      2  6.63
#> 23    2   7   1    8.0   66.8  139      2  7.23
#> 24    2   8   1    8.0   69.3  147      2  7.13
#> 25    2  24   2    8.2   70.7  125      2  6.17
#> 26    2  10   1    8.2   71.1  145      2  7.23
#> 27    3  15   1    8.3   70.4  133      2  6.82
#> 28    3  14   2    8.3   71.5  137      2  7.38
#> 29    3  14   2    8.3   72.9  141      2  7.38
#> 30    3  17   2    8.3   73.5  136      2  7.04
#> 31    3  15   1    8.3   76.1  133      2  6.82
#> 32    3  13   2    8.4   71.8  143      2  7.64
#> 33    3  20   1    8.4   77.8  132      2  6.41
#> 34    3  29   2    8.5   75.6  127      2  6.12
#> 35    3  21   2    8.5   75.6  133      2  6.89
#> 36    3  21   1    8.5   76.7  123      2  6.44
#> 37    3  17   1    8.5   77.6  137      2  6.88
#> 38    3  15   2    8.6   74.9  134      2  7.71
#> 39    3  24   1    8.6   77.8  137      2  6.26
#> 40    3  12   1    8.7   72.1  156      2  7.76
#> 41    4  27   2    8.7   75.8  127      2  6.57
#> 42    4  25   2    8.7   80.8  129      2  6.76
#> 43    4  25   1    8.8   81.4  125      2  6.43
#> 44    4  24   2    8.9   74.3  142      2  7.14
#> 45    4  30   1    9.0   77.6  132      2  6.21
#> 46    4  22   1    9.0   82.2  135      2  7.02
#> 47    4  16   1    9.1   76.4  145      2  7.86
#> 48    4  24   1    9.1   80.0  126      2  6.94
#> 49    4  17   2    9.2   74.8  144      2  8.35
#> 50    4  15   2    9.2   75.3  145      2  8.60
#> 51    4  33   1    9.2   80.1  136      2  6.20
#> 52    4  25   2    9.2   80.9  142      2  7.45
#> 53    5  16   1    9.4   75.0  148      2  8.30
#> 54    5  29   2    9.4   81.9  138      2  7.32
#> 55    5  16   1    9.5   73.2  151      2  8.44
#> 56    5  17   1    9.6   73.5  146      2  8.46
#> 57    5  29   2    9.6   84.6  134      2  7.58
#> 58    5  16   1    9.7   74.2  148      2  8.73
#> 59    5  20   1    9.7   80.4  142      2  8.22
#> 60    5  34   1    9.8   80.2  138      2  6.88
#> 61    5  21   2    9.8   80.9  145      2  8.73
#> 62    5  20   1    9.8   82.2  139      2  8.36
#> 63    5  29   1    9.9   78.6  142      2  7.48
#> 64    5  25   2    9.9   82.9  138      2  8.41
#> 65    5  15   1   10.0   75.3  157      2  9.30
#> 66    6  39   1   10.0   81.4  133      2  6.71
#> 67    6   9   1   10.1   72.6  155      2 10.22
#> 68    6  21   1   10.1   76.7  153      2  8.66
#> 69    6  17   2   10.1   79.1  143      2  9.67
#> 70    6  25   1   10.1   79.5  140      2  8.18
#> 71    6  13   2   10.2   74.6  157      2 10.36
#> 72    6  25   2   10.3   81.3  141      2  8.95
#> 73    6  31   2   10.3   85.1  130      2  8.31
#> 74    6  25   2   10.4   81.3  138      2  9.09
#> 75    6  37   1   10.6   76.2  153      2  7.62
#> 76    6  25   2   10.6   80.7  138      2  9.37
#> 77    6  26   1   10.6   83.1  142      2  8.74
#> 78    6  36   2   10.6   84.2  135      2  8.22
#> 79    6  37   2   10.7   84.2  150      2  8.26
#> 80    7  34   2   10.7   85.7  146      2  8.54
#> 81    7  47   1   10.7   89.8  131      2  6.96
#> 82    7  12   1   10.8   82.0  147      2 10.88
#> 83    7  25   2   10.8   83.6  141      2  9.64
#> 84    7  31   1   10.8   83.8  158      2  8.45
#> 85    7  34   1   10.8   84.4  144      2  8.15
#> 86    7  27   2   10.9   79.0  153      2  9.54
#> 87    7  23   2   10.9   85.9  145      2 10.02
#> 88    7  19   1   11.0   75.4  168      2 10.18
#> 89    7  26   2   11.0   83.2  145      2  9.79
#> 90    7  54   1   11.0   85.9  139      2  6.85
#> 91    7  20   1   11.1   79.9  156      2 10.18
#> 92    7  23   1   11.1   84.7  146      2  9.78
#> 93    7  43   2   11.1   86.0  142      2  8.24
#> 94    8  31   1   11.1   87.1  143      2  8.84
#> 95    8  40   2   11.2   95.2  137      2  8.62
#> 96    8  18   2   11.3   75.5  161      2 11.27
#> 97    8  41   1   11.3   89.4  142      2  8.15
#> 98    8  40   1   11.3   90.7  148      2  8.23
#> 99    8  33   2   11.3   91.2  139      2  9.41
#> 100   8  11   1   11.4   78.4  171      2 11.92
#> 101   8  25   1   11.4   82.2  163      2  9.94
#> 102   8  27   2   11.4   86.8  155      2 10.22
#> 103   8  38   1   11.4   89.4  141      2  8.53
#> 104   8  17   2   11.5   80.6  158      2 11.71
#> 105   8  31   2   11.5   85.4  153      2  9.89
#> 106   9  45   1   11.5   86.2  143      2  8.07
#> 107   9  53   1   11.6   81.4  153      2  7.61
#> 108   9  23   2   11.6   86.3  143      2 11.00
#> 109   9  37   1   11.6   86.3  149      2  8.88
#> 110   9  38   1   11.6   89.9  146      2  8.78
#> 111   9  40   1   11.7   93.0  140      2  8.73
#> 112   9  36   1   11.8   87.1  146      2  9.22
#> 113   9  29   1   11.9   84.3  166      2 10.11
#> 114   9  26   2   12.0   81.2  164      2 11.15
#> 115   9  33   1   12.0   86.0  154      2  9.79
#> 116   9  36   1   12.0   87.5  150      2  9.48
#> 117   9  55   2   12.0   96.2  144      2  8.41
#> 118  10  16   1   12.1   82.3  157      2 12.19
#> 119  10  35   2   12.1   86.6  147      2 10.24
#> 120  10  43   1   12.1   91.7  151      2  8.96
#> 121  10  38   2   12.2   80.9  175      2 10.06
#> 122  10  37   1   12.2   85.5  162      2  9.63
#> 123  10  41   2   12.2   87.6  155      2  9.77
#> 124  10  45   2   12.2   90.1  149      2  9.42
#> 125  10  46   2   12.2   99.6  146      2  9.33
#> 126  10  31   1   12.3   88.0  167      2 10.40
#> 127  10  44   1   12.3   88.1  151      2  9.12
#> 128  10  47   2   12.3   94.9  133      2  9.37
#> 129  10  29   1   12.4   91.8  141      2 10.77
#> 130  10  46   2   12.4   99.6  140      2  9.57
#> 131  10  34   1   12.6   87.0  156      2 10.45
#> 132  10  38   2   12.7   85.7  152      2 10.69
#> 133  11  32   1   12.8   87.6  161      2 10.93
#> 134  11  36   1   12.8   90.4  150      2 10.49
#> 135  11  39   1   12.8   92.7  152      2 10.18
#> 136  11  42   1   12.8   94.0  146      2  9.90
#> 137  11  38   2   12.9   83.4  174      2 10.94
#> 138  11  26   2   12.9   86.2  159      2 12.38
#> 139  11  35   1   12.9   90.9  154      2 10.72
#> 140  11  41   1   12.9   91.2  160      2 10.12
#> 141  11  32   1   12.9   91.4  148      2 11.06
#> 142  11  49   1   13.0   93.1  152      2  9.55
#> 143  11  30   1   13.1   84.3  161      2 11.56
#> 144  11  32   1   13.1   85.1  136      2 11.32
#> 145  11  35   1   13.1   87.5  158      2 10.97
#> 146  11  34   1   13.1   88.6  162      2 11.08
#> 147  11  45   2   13.1   91.6  150      2 10.51
#> 148  12  29   1   13.2   90.8  158      2 11.82
#> 149  12  47   2   13.2   91.4  146      2 10.45
#> 150  12  37   2   13.2   93.8  155      2 11.42
#> 151  12  35   1   13.3   96.5  142      2 11.23
#> 152  12  15   2   13.4   82.4  157      2 14.84
#> 153  12  38   2   13.4   93.0  152      2 11.57
#> 154  12  37   2   13.4   94.1  149      2 11.68
#> 155  12  53   2   13.4  100.3  145      2 10.18
#> 156  12  37   2   13.5   84.4  168      2 11.80
#> 157  12  43   1   13.5   96.1  151      2 10.66
#> 158  12  52   1   13.5   97.0  144      2  9.91
#> 159  12  45   2   13.5   98.0  148      2 10.99
#> 160  12  39   2   13.5   99.1  145      2 11.59
#> 161  12  47   2   13.6   97.0  146      2 10.93
#> 162  13  26   2   13.7   85.7  171      2 13.47
#> 163  13  24   1   13.8   80.9  172      2 13.32
#> 164  13  38   2   13.8   93.3  153      2 12.07
#> 165  13  57   1   13.8   94.0  149      2  9.90
#> 166  13  42   2   13.8  103.4  144      2 11.65
#> 167  13  42   2   13.9   92.3  175      2 11.77
#> 168  13  52   1   13.9   95.5  151      2 10.38
#> 169  13  52   1   13.9   95.7  156      2 10.38
#> 170  13  57   2   13.9   97.2  153      2 10.45
#> 171  13  52   2   13.9   97.8  141      2 10.85
#> 172  13  52   2   13.9   97.9  155      2 10.85
#> 173  13  49   2   13.9   98.2  150      2 11.10
#> 174  13  58   2   14.0   96.6  154      2 10.49
#> 175  13  50   1   14.0   97.9  158      2 10.66
#> 176  13  39   2   14.1   93.7  155      2 12.34
#> 177  14  49   2   14.1   96.4  140      2 11.34
#> 178  14  50   2   14.1   97.1  148      2 11.25
#> 179  14  59   2   14.1   97.3  152      2 10.52
#> 180  14  52   2   14.1  102.4  156      2 11.08
#> 181  14  41   1   14.2   96.7  146      2 11.71
#> 182  14  53   1   14.2   97.7  151      2 10.66
#> 183  14  49   2   14.2   98.1  156      2 11.46
#> 184  14  35   2   14.3   89.8  165      2 13.05
#> 185  14  50   1   14.3   93.7  152      2 11.01
#> 186  14  41   1   14.4   94.8  158      2 11.96
#> 187  14  42   1   14.5   95.1  161      2 11.98
#> 188  14  58   1   14.5  104.9  130      2 10.64
#> 189  14  49   1   14.7   96.6  152      2 11.57
#> 190  14  52   2   14.7   96.7  154      2 11.78
#> 191  14  51   2   14.7   99.2  155      2 11.87
#> 192  14  49   1   14.7   99.6  164      2 11.57
#> 193  15  42   2   15.0   93.5  172      2 13.12
#> 194  15  28   2   15.1   91.5  164      2 15.06
#> 195  15  41   2   15.2   95.9  165      2 13.48
#> 196  15  53   1   15.3   95.0  166      2 11.94
#> 197  15  58   1   15.3  100.7  155      2 11.56
#> 198  15  48   1   15.3  103.5  143      2 12.37
#> 199  15  54   1   15.5   97.9  159      2 12.10
#> 200  15  58   1   15.5  101.3  166      2 11.79
#> 201  15  52   2   15.6  100.0  164      2 12.83
#> 202  15  50   1   15.7   99.6  153      2 12.67
#> 203  15  53   2   15.7  103.1  158      2 12.86
#> 204  15  43   1   15.8   91.1  179      2 13.47
#> 205  15  57   1   15.9  103.8  161      2 12.32
#> 206  15  55   2   16.3   95.4  169      2 13.37
#> 207  16  54   2   15.3  102.0  156      2 12.30
#> 208  16  44   1   16.3   96.2  173      2 13.97
#> 209  16  47   1   16.3  102.0  166      2 13.66
#> 210  16  57   1   16.3  108.4  138      2 12.78
#> 211  16  52   1   16.4  103.9  152      2 13.32
#> 212  16  52   2   16.6   97.8  144      2 14.00
#> 213  16  56   1   16.6  103.9  148      2 13.21
#> 214  16  55   1   16.7  106.3  154      2 13.41
#> 215  16  52   1   17.0  101.3  163      2 14.02
#> 216  16  50   1   17.3  101.8  168      2 14.56
#> 217  16  53   1   17.5  102.2  168      2 14.52
#> 218  16  42   1   17.7  100.9  145      2 15.90
#> 219  16  48   2   17.8  111.3  176      2 15.84
#> 220  16  53   1   17.9   98.7  171      2 14.98
#> 221  16  50   1   18.1  106.4  166      2 15.51

# Calculate height-for-age (hfa) for the anthro3 dataset
addWGSR(data = anthro3,
        sex = "sex",
        firstPart = "height",
        secondPart = "age",
        index = "hfa")

#> ================================================================================

#>     psu age sex weight height muac oedema  hfaz
#> 1     1  10   1    5.7   64.2  125      2  6.58
#> 2     1  10   2    5.8   64.4  121      2  7.17
#> 3     1   9   2    6.5   62.2  139      2  6.11
#> 4     1  11   9    6.5   64.9  129      2    NA
#> 5     1  24   2    6.5   72.9  120      2 10.35
#> 6     1  12   2    6.6   69.4  126      2  9.60
#> 7     1   9   2    7.0   66.7  136      2  8.49
#> 8     1   7   1    7.1   63.5  139      2  6.51
#> 9     1   9   2    7.1   66.2  144      2  8.22
#> 10    1  16   2    7.2   69.0  131      2  9.01
#> 11    1  13   2    7.4   68.2  137      2  8.87
#> 12    1  11   2    7.4   70.0  132      2 10.02
#> 13    1  14   2    7.5   64.8  125      2  6.98
#> 14    2   8   2    7.5   65.7  151      2  8.06
#> 15    2   8   1    7.5   69.1  140      2  9.35
#> 16    2  13   1    7.6   69.2  131      2  8.91
#> 17    2  13   2    7.6   70.0  125      2  9.82
#> 18    2  22   2    7.6   76.2  121      2 12.24
#> 19    2  16   1    7.7   69.6  136      2  8.85
#> 20    2   8   2    7.8   63.3  144      2  6.79
#> 21    2  11   2    7.8   71.5  134      2 10.82
#> 22    2  14   2    7.8   76.6  124      2 13.18
#> 23    2   7   1    8.0   66.8  139      2  8.24
#> 24    2   8   1    8.0   69.3  147      2  9.46
#> 25    2  24   2    8.2   70.7  125      2  9.22
#> 26    2  10   1    8.2   71.1  145      2 10.20
#> 27    3  15   1    8.3   70.4  133      2  9.36
#> 28    3  14   2    8.3   71.5  137      2 10.50
#> 29    3  14   2    8.3   72.9  141      2 11.24
#> 30    3  17   2    8.3   73.5  136      2 11.28
#> 31    3  15   1    8.3   76.1  133      2 12.33
#> 32    3  13   2    8.4   71.8  143      2 10.77
#> 33    3  20   1    8.4   77.8  132      2 12.76
#> 34    3  29   2    8.5   75.6  127      2 11.33
#> 35    3  21   2    8.5   75.6  133      2 12.02
#> 36    3  21   1    8.5   76.7  123      2 12.10
#> 37    3  17   1    8.5   77.6  137      2 12.93
#> 38    3  15   2    8.6   74.9  134      2 12.20
#> 39    3  24   1    8.6   77.8  137      2 12.40
#> 40    3  12   1    8.7   72.1  156      2 10.53
#> 41    4  27   2    8.7   75.8  127      2 11.60
#> 42    4  25   2    8.7   80.8  129      2 14.35
#> 43    4  25   1    8.8   81.4  125      2 14.18
#> 44    4  24   2    8.9   74.3  142      2 11.08
#> 45    4  30   1    9.0   77.6  132      2 11.79
#> 46    4  22   1    9.0   82.2  135      2 14.86
#> 47    4  16   1    9.1   76.4  145      2 12.40
#> 48    4  24   1    9.1   80.0  126      2 13.54
#> 49    4  17   2    9.2   74.8  144      2 11.96
#> 50    4  15   2    9.2   75.3  145      2 12.41
#> 51    4  33   1    9.2   80.1  136      2 12.82
#> 52    4  25   2    9.2   80.9  142      2 14.40
#> 53    5  16   1    9.4   75.0  148      2 11.67
#> 54    5  29   2    9.4   81.9  138      2 14.56
#> 55    5  16   1    9.5   73.2  151      2 10.73
#> 56    5  17   1    9.6   73.5  146      2 10.79
#> 57    5  29   2    9.6   84.6  134      2 15.95
#> 58    5  16   1    9.7   74.2  148      2 11.25
#> 59    5  20   1    9.7   80.4  142      2 14.11
#> 60    5  34   1    9.8   80.2  138      2 12.79
#> 61    5  21   2    9.8   80.9  145      2 14.77
#> 62    5  20   1    9.8   82.2  139      2 15.04
#> 63    5  29   1    9.9   78.6  142      2 12.38
#> 64    5  25   2    9.9   82.9  138      2 15.43
#> 65    5  15   1   10.0   75.3  157      2 11.92
#> 66    6  39   1   10.0   81.4  133      2 13.01
#> 67    6   9   1   10.1   72.6  155      2 11.09
#> 68    6  21   1   10.1   76.7  153      2 12.10
#> 69    6  17   2   10.1   79.1  143      2 14.21
#> 70    6  25   1   10.1   79.5  140      2 13.19
#> 71    6  13   2   10.2   74.6  157      2 12.24
#> 72    6  25   2   10.3   81.3  141      2 14.61
#> 73    6  31   2   10.3   85.1  130      2 16.03
#> 74    6  25   2   10.4   81.3  138      2 14.61
#> 75    6  37   1   10.6   76.2  153      2 10.51
#> 76    6  25   2   10.6   80.7  138      2 14.30
#> 77    6  26   1   10.6   83.1  142      2 14.96
#> 78    6  36   2   10.6   84.2  135      2 15.14
#> 79    6  37   2   10.7   84.2  150      2 15.06
#> 80    7  34   2   10.7   85.7  146      2 16.07
#> 81    7  47   1   10.7   89.8  131      2 16.66
#> 82    7  12   1   10.8   82.0  147      2 15.72
#> 83    7  25   2   10.8   83.6  141      2 15.80
#> 84    7  31   1   10.8   83.8  158      2 14.89
#> 85    7  34   1   10.8   84.4  144      2 14.94
#> 86    7  27   2   10.9   79.0  153      2 13.25
#> 87    7  23   2   10.9   85.9  145      2 17.17
#> 88    7  19   1   11.0   75.4  168      2 11.60
#> 89    7  26   2   11.0   83.2  145      2 15.50
#> 90    7  54   1   11.0   85.9  139      2 14.19
#> 91    7  20   1   11.1   79.9  156      2 13.85
#> 92    7  23   1   11.1   84.7  146      2 16.06
#> 93    7  43   2   11.1   86.0  142      2 15.49
#> 94    8  31   1   11.1   87.1  143      2 16.58
#> 95    8  40   2   11.2   95.2  137      2 20.37
#> 96    8  18   2   11.3   75.5  161      2 12.23
#> 97    8  41   1   11.3   89.4  142      2 16.92
#> 98    8  40   1   11.3   90.7  148      2 17.67
#> 99    8  33   2   11.3   91.2  139      2 18.96
#> 100   8  11   1   11.4   78.4  171      2 13.93
#> 101   8  25   1   11.4   82.2  163      2 14.59
#> 102   8  27   2   11.4   86.8  155      2 17.26
#> 103   8  38   1   11.4   89.4  141      2 17.16
#> 104   8  17   2   11.5   80.6  158      2 14.99
#> 105   8  31   2   11.5   85.4  153      2 16.18
#> 106   9  45   1   11.5   86.2  143      2 14.99
#> 107   9  53   1   11.6   81.4  153      2 12.00
#> 108   9  23   2   11.6   86.3  143      2 17.38
#> 109   9  37   1   11.6   86.3  149      2 15.67
#> 110   9  38   1   11.6   89.9  146      2 17.42
#> 111   9  40   1   11.7   93.0  140      2 18.84
#> 112   9  36   1   11.8   87.1  146      2 16.16
#> 113   9  29   1   11.9   84.3  166      2 15.32
#> 114   9  26   2   12.0   81.2  164      2 14.47
#> 115   9  33   1   12.0   86.0  154      2 15.85
#> 116   9  36   1   12.0   87.5  150      2 16.36
#> 117   9  55   2   12.0   96.2  144      2 19.65
#> 118  10  16   1   12.1   82.3  157      2 15.48
#> 119  10  35   2   12.1   86.6  147      2 16.44
#> 120  10  43   1   12.1   91.7  151      2 17.94
#> 121  10  38   2   12.2   80.9  175      2 13.31
#> 122  10  37   1   12.2   85.5  162      2 15.26
#> 123  10  41   2   12.2   87.6  155      2 16.45
#> 124  10  45   2   12.2   90.1  149      2 17.39
#> 125  10  46   2   12.2   99.6  146      2 22.06
#> 126  10  31   1   12.3   88.0  167      2 17.05
#> 127  10  44   1   12.3   88.1  151      2 16.03
#> 128  10  47   2   12.3   94.9  133      2 19.63
#> 129  10  29   1   12.4   91.8  141      2 19.17
#> 130  10  46   2   12.4   99.6  140      2 22.06
#> 131  10  34   1   12.6   87.0  156      2 16.27
#> 132  10  38   2   12.7   85.7  152      2 15.74
#> 133  11  32   1   12.8   87.6  161      2 16.75
#> 134  11  36   1   12.8   90.4  150      2 17.84
#> 135  11  39   1   12.8   92.7  152      2 18.77
#> 136  11  42   1   12.8   94.0  146      2 19.18
#> 137  11  38   2   12.9   83.4  174      2 14.57
#> 138  11  26   2   12.9   86.2  159      2 17.04
#> 139  11  35   1   12.9   90.9  154      2 18.18
#> 140  11  41   1   12.9   91.2  160      2 17.84
#> 141  11  32   1   12.9   91.4  148      2 18.70
#> 142  11  49   1   13.0   93.1  152      2 18.18
#> 143  11  30   1   13.1   84.3  161      2 15.23
#> 144  11  32   1   13.1   85.1  136      2 15.47
#> 145  11  35   1   13.1   87.5  158      2 16.45
#> 146  11  34   1   13.1   88.6  162      2 17.09
#> 147  11  45   2   13.1   91.6  150      2 18.14
#> 148  12  29   1   13.2   90.8  158      2 18.66
#> 149  12  47   2   13.2   91.4  146      2 17.88
#> 150  12  37   2   13.2   93.8  155      2 19.92
#> 151  12  35   1   13.3   96.5  142      2 21.05
#> 152  12  15   2   13.4   82.4  157      2 16.13
#> 153  12  38   2   13.4   93.0  152      2 19.43
#> 154  12  37   2   13.4   94.1  149      2 20.07
#> 155  12  53   2   13.4  100.3  145      2 21.84
#> 156  12  37   2   13.5   84.4  168      2 15.16
#> 157  12  43   1   13.5   96.1  151      2 20.17
#> 158  12  52   1   13.5   97.0  144      2 19.92
#> 159  12  45   2   13.5   98.0  148      2 21.35
#> 160  12  39   2   13.5   99.1  145      2 22.42
#> 161  12  47   2   13.6   97.0  146      2 20.68
#> 162  13  26   2   13.7   85.7  171      2 16.79
#> 163  13  24   1   13.8   80.9  172      2 14.00
#> 164  13  38   2   13.8   93.3  153      2 19.58
#> 165  13  57   1   13.8   94.0  149      2 18.05
#> 166  13  42   2   13.8  103.4  144      2 24.32
#> 167  13  42   2   13.9   92.3  175      2 18.74
#> 168  13  52   1   13.9   95.5  151      2 19.17
#> 169  13  52   1   13.9   95.7  156      2 19.27
#> 170  13  57   2   13.9   97.2  153      2 19.99
#> 171  13  52   2   13.9   97.8  141      2 20.68
#> 172  13  52   2   13.9   97.9  155      2 20.73
#> 173  13  49   2   13.9   98.2  150      2 21.11
#> 174  13  58   2   14.0   96.6  154      2 19.62
#> 175  13  50   1   14.0   97.9  158      2 20.53
#> 176  13  39   2   14.1   93.7  155      2 19.70
#> 177  14  49   2   14.1   96.4  140      2 20.22
#> 178  14  50   2   14.1   97.1  148      2 20.48
#> 179  14  59   2   14.1   97.3  152      2 19.89
#> 180  14  52   2   14.1  102.4  156      2 22.96
#> 181  14  41   1   14.2   96.7  146      2 20.63
#> 182  14  53   1   14.2   97.7  151      2 20.20
#> 183  14  49   2   14.2   98.1  156      2 21.06
#> 184  14  35   2   14.3   89.8  165      2 18.07
#> 185  14  50   1   14.3   93.7  152      2 18.41
#> 186  14  41   1   14.4   94.8  158      2 19.67
#> 187  14  42   1   14.5   95.1  161      2 19.74
#> 188  14  58   1   14.5  104.9  130      2 23.45
#> 189  14  49   1   14.7   96.6  152      2 19.95
#> 190  14  52   2   14.7   96.7  154      2 20.13
#> 191  14  51   2   14.7   99.2  155      2 21.45
#> 192  14  49   1   14.7   99.6  164      2 21.46
#> 193  15  42   2   15.0   93.5  172      2 19.34
#> 194  15  28   2   15.1   91.5  164      2 19.58
#> 195  15  41   2   15.2   95.9  165      2 20.63
#> 196  15  53   1   15.3   95.0  166      2 18.84
#> 197  15  58   1   15.3  100.7  155      2 21.34
#> 198  15  48   1   15.3  103.5  143      2 23.51
#> 199  15  54   1   15.5   97.9  159      2 20.23
#> 200  15  58   1   15.5  101.3  166      2 21.64
#> 201  15  52   2   15.6  100.0  164      2 21.77
#> 202  15  50   1   15.7   99.6  153      2 21.38
#> 203  15  53   2   15.7  103.1  158      2 23.22
#> 204  15  43   1   15.8   91.1  179      2 17.63
#> 205  15  57   1   15.9  103.8  161      2 22.97
#> 206  15  55   2   16.3   95.4  169      2 19.25
#> 207  16  54   2   15.3  102.0  156      2 22.60
#> 208  16  44   1   16.3   96.2  173      2 20.14
#> 209  16  47   1   16.3  102.0  166      2 22.83
#> 210  16  57   1   16.3  108.4  138      2 25.28
#> 211  16  52   1   16.4  103.9  152      2 23.40
#> 212  16  52   2   16.6   97.8  144      2 20.68
#> 213  16  56   1   16.6  103.9  148      2 23.09
#> 214  16  55   1   16.7  106.3  154      2 24.37
#> 215  16  52   1   17.0  101.3  163      2 22.09
#> 216  16  50   1   17.3  101.8  168      2 22.49
#> 217  16  53   1   17.5  102.2  168      2 22.46
#> 218  16  42   1   17.7  100.9  145      2 22.69
#> 219  16  48   2   17.8  111.3  176      2 27.73
#> 220  16  53   1   17.9   98.7  171      2 20.70
#> 221  16  50   1   18.1  106.4  166      2 24.81

# Calculate MUAC-for-age (mfa) for the anthro4 dataset

## Convert age in anthro4 from months to days
testData <- anthro4
testData$age <- testData$agemons * (365.25 / 12)

addWGSR(data = testData,
        sex = "sex",
        firstPart = "muac",
        secondPart = "age",
        index = "mfa")

#> ================================================================================

#>     pk_serial muac   agemons sex  age  mfaz
#> 1       76220 16.0  73.23203   2 2229 -0.88
#> 2       76290 18.4  93.63450   1 2850  0.07
#> 3       76310 14.4  75.69610   1 2304 -2.20
#> 4       76460 15.4  87.16222   2 2653 -1.63
#> 5       76570 18.0 139.69611   1 4252 -1.76
#> 6       76580 18.0 119.81931   2 3647 -1.04
#> 7       76620 14.6 126.52156   2 3851 -3.39
#> 8       76840 12.6  91.95893   1 2799 -4.72
#> 9       76860 14.5  74.54620   1 2269 -2.06
#> 10      77060 16.0 115.84394   1 3526 -2.28
#> 11      77250 16.5  82.89117   1 2523 -0.74
#> 12      77350 13.4  62.12731   2 1891 -2.63
#> 13      77510 13.5  64.82136   2 1973 -2.60
#> 14      77570 14.9  82.82546   2 2521 -1.88
#> 15      77760 16.0 118.14374   1 3596 -2.36
#> 16      77860 11.5 103.55647   2 3152 -4.94
#> 17      77870 13.6  76.68173   1 2334 -3.06
#> 18      77990 14.0  70.63655   2 2150 -2.30
#> 19      78040 19.0 136.24641   2 4147 -1.10
#> 20      78210 17.2 135.75359   1 4132 -2.12
#> 21      78240 17.0  72.34497   2 2202 -0.23
#> 22      78340 14.7  84.17248   2 2562 -2.06
#> 23      78370 17.0  96.19713   1 2928 -0.86
#> 24      78410 17.0 155.95894   1 4747 -3.21
#> 25      78570 17.5 115.41684   1 3513 -1.15
#> 26      78780 15.3 122.48049   1 3728 -3.14
#> 27      78800 16.0  72.31212   2 2201 -0.86
#> 28      79090 17.6 137.95482   2 4199 -1.89
#> 29      79290 19.0 139.00616   1 4231 -1.15
#> 30      79300 17.0 124.64887   1 3794 -1.82
#> 31      79500 15.0 133.22382   1 4055 -3.73
#> 32      79540 13.7  60.35318   1 1837 -2.26
#> 33      79570 16.7  73.10062   1 2225 -0.25
#> 34      79610 16.0  63.57290   1 1935 -0.46
#> 35      79760 13.5  77.04312   1 2345 -3.18
#> 36      79770 14.5  84.04107   1 2558 -2.48
#> 37      79840 17.6  91.99178   1 2800 -0.34
#> 38      79850 13.0  78.88296   2 2401 -3.33
#> 39      79860 17.0  70.86653   2 2157 -0.19
#> 40      79910 17.0  71.03080   1 2162  0.02
#> 41      79970 15.6  69.61807   1 2119 -0.94
#> 42      80010 16.0  78.39014   1 2386 -0.94
#> 43      80200 16.0  86.37372   1 2629 -1.25
#> 44      80290 16.0 134.37372   1 4090 -3.00
#> 45      80380 17.0  87.55647   1 2665 -0.57
#> 46      80490 15.0  66.03696   2 2010 -1.41
#> 47      80610 17.1 103.29363   2 3144 -1.01
#> 48      80630 14.0  93.33881   2 2841 -2.83
#> 49      80640 15.4  96.65708   2 2942 -1.87
#> 50      80660 16.0  70.43942   2 2144 -0.81
#> 51      80680 15.2 109.73306   2 3340 -2.39
#> 52      80940 17.5  66.62833   2 2028  0.23
#> 53      81060 16.0 126.45585   1 3849 -2.67
#> 54      81080 14.0  71.45791   1 2175 -2.40
#> 55      81100 15.3  87.35934   1 2659 -1.87
#> 56      81260 18.1 142.75154   1 4345 -1.83
#> 57      81540 15.5  89.59343   2 2727 -1.62
#> 58      81590 15.6  77.73306   1 2366 -1.23
#> 59      81630 17.9 126.68583   1 3856 -1.29
#> 60      81880 15.5 116.89528   1 3558 -2.76
#> 61      81890 17.0 117.65092   1 3581 -1.57
#> 62      81900 15.5  69.58521   1 2118 -1.01
#> 63      82010 13.0  64.13142   1 1952 -3.14
#> 64      82080 15.0 132.13963   2 4022 -3.33
#> 65      82190 14.0  82.46407   2 2510 -2.57
#> 66      82320 16.0  93.27310   2 2839 -1.39
#> 67      82370 15.2  99.05544   2 3015 -2.07
#> 68      82560 16.3 154.54620   2 4704 -3.44
#> 69      82670 15.2  77.40452   2 2356 -1.53
#> 70      82690 15.9  68.53388   1 2086 -0.67
#> 71      83720 13.6  60.15606   2 1831 -2.34
#> 72      84080 15.3  65.77412   2 2002 -1.18
#> 73      84220 12.6 100.23819   1 3051 -4.98
#> 74      84430 16.8  63.47433   1 1932  0.11
#> 75      84750 15.0  72.08214   1 2194 -1.51
#> 76      84990 14.0  61.79877   1 1881 -2.06
#> 77      85130 16.0  82.92403   1 2524 -1.12
#> 78      85170 16.0  80.65708   2 2455 -1.06
#> 79      86090 16.6 124.55031   1 3791 -2.11
#> 80      86300 16.3  70.37372   1 2142 -0.44
#> 81      86400 16.4 120.08214   1 3655 -2.10
#> 82      86460 15.0  95.44147   2 2905 -2.12
#> 83      86720 15.0  80.68994   1 2456 -1.86
#> 84      86970 14.6  73.75770   1 2245 -1.93
#> 85      87050 15.1 109.93018   2 3346 -2.46
#> 86      87140 15.0 110.81725   1 3373 -3.03
#> 87      87260 17.0  92.28748   1 2809 -0.73
#> 88      87450 17.2  90.67762   2 2760 -0.60
#> 89      87460 12.8  62.22587   1 1894 -3.28
#> 90      88360 17.5  91.43327   1 2783 -0.38
#> 91      88520 15.2 101.58521   1 3092 -2.50
#> 92      88690 14.3  61.53593   1 1873 -1.77
#> 93      88990 15.5  83.54826   1 2543 -1.54
#> 94      89090 17.0 112.95277   2 3438 -1.35
#> 95      89410 12.3 100.27105   1 3052 -5.27
#> 96      89710 15.0  74.97330   2 2282 -1.62
#> 97      89900 19.0  96.45996   1 2936  0.29
#> 98      90730 16.0  65.67557   1 1999 -0.51
#> 99      90740 13.0  68.36961   2 2081 -3.12
#> 100     90850 16.0 118.20944   2 3598 -2.12
#> 101     91280 16.7 115.64682   1 3520 -1.71
#> 102     91360 15.5 110.12731   1 3352 -2.52
#> 103     91610 14.7 109.73306   1 3340 -3.27
#> 104     91740 16.0 124.61602   1 3793 -2.60
#> 105     91830 12.3  63.01437   2 1918 -3.62
#> 106     91980 15.0  94.35729   2 2872 -2.09
#> 107     92360 16.6 105.75770   1 3219 -1.46
#> 108     92730 17.0 121.46201   1 3697 -1.70
#> 109     94170 16.0  83.15401   2 2531 -1.13
#> 110     94280 12.0  82.46407   1 2510 -4.93
#> 111     94760 14.0  83.87679   2 2553 -2.60
#> 112     95390 17.0 110.06160   1 3350 -1.31
#> 113     95440 16.5  67.18686   1 2045 -0.20
#> 114     95490 16.0  73.23203   2 2229 -0.88
#> 115     95530 15.7  64.52567   2 1964 -0.87
#> 116     95910 15.0  75.79466   1 2307 -1.66
#> 117     95970 18.5 128.91992   1 3924 -1.02
#> 118     96150 16.0 104.47639   2 3180 -1.70
#> 119     96400 15.0  87.95072   1 2677 -2.16
#> 120     96490 16.0  64.59138   2 1966 -0.66
#> 121     96730 19.6 149.81520   2 4560 -1.31
#> 122     96870 15.5  98.03696   1 2984 -2.10
#> 123     96970 14.2  68.27105   1 2078 -2.09
#> 124     97070 16.0  60.68173   1 1847 -0.39
#> 125     97080 17.0 108.68172   1 3308 -1.26
#> 126     97190 15.0  65.54415   1 1995 -1.29
#> 127     97450 16.5 117.88091   1 3588 -1.94
#> 128     97740 14.0  68.89528   1 2097 -2.30
#> 129     97760 18.5  83.81109   2 2551  0.28
#> 130     98160 17.8 141.37166   2 4303 -1.92
#> 131     98190 14.0  71.88501   2 2188 -2.33
#> 132     98360 14.5  72.47639   1 2206 -1.97
#> 133     98420 14.2  75.17043   1 2288 -2.37
#> 134     98640 18.0  66.52978   1 2025  0.78
#> 135     98720 13.7  82.82546   2 2521 -2.83
#> 136     98830 14.7  80.49281   2 2450 -1.97
#> 137     98900 16.0  61.79877   1 1881 -0.41
#> 138     98980 14.8  63.67146   1 1938 -1.41
#> 139     99410 15.0  69.09240   2 2103 -1.48
#> 140     99440 16.6  97.93840   2 2981 -1.14
#> 141     99610 15.8  98.46407   1 2997 -1.85
#> 142     99670 13.2  73.33060   1 2232 -3.31
#> 143     99780 17.0  93.17454   2 2836 -0.78
#> 144     99800 15.5  63.70431   1 1939 -0.84
#> 145     99920 17.5 147.31827   2 4484 -2.35
#> 146    100050 17.5  86.30801   1 2627 -0.20
#> 147    100060 14.0  70.63655   2 2150 -2.30
#> 148    100100 16.2 105.52772   1 3212 -1.76
#> 149    100120 15.4 101.71663   2 3096 -2.01
#> 150    100440 16.0 109.20739   2 3324 -1.84
#> 151    100460 15.0 110.91581   2 3376 -2.57
#> 152    100610 15.0  94.32443   1 2871 -2.42
#> 153    100820 19.0 104.80493   1 3190  0.04
#> 154    101430 15.0  78.62013   2 2393 -1.71
#> 155    101460 13.2 131.48254   2 4002 -4.49
#> 156    101710 16.7 129.77412   1 3950 -2.24
#> 157    101780 17.2 133.35524   2 4059 -1.94
#> 158    101790 15.0  98.36550   2 2994 -2.20
#> 159    101850 16.6 108.05750   2 3289 -1.43
#> 160    101920 16.7 107.63039   1 3276 -1.44
#> 161    101970 15.2  64.39425   1 1960 -1.10
#> 162    102220 19.6 143.14580   1 4357 -1.01
#> 163    102380 16.2 123.89323   1 3771 -2.41
#> 164    102390 15.2 113.21561   1 3446 -2.92
#> 165    102550 14.0  86.43942   1 2631 -3.11
#> 166    102630 13.0  67.38398   1 2051 -3.25
#> 167    102930 17.5 135.62218   1 4128 -1.91
#> 168    103060 19.0 149.02669   1 4536 -1.59
#> 169    103090 14.5  65.41273   2 1991 -1.78
#> 170    103300 16.0 113.41273   1 3452 -2.19
#> 171    103580 13.8  61.17454   2 1862 -2.21
#> 172    103590 16.0 101.94661   2 3103 -1.62
#> 173    103610 16.4 152.47638   2 4641 -3.29
#> 174    104070 14.0  66.69405   2 2030 -2.21
#> 175    104190 17.5 133.55237   2 4065 -1.78
#> 176    104360 18.0 107.23614   2 3264 -0.65
#> 177    104600 15.7  61.47023   2 1871 -0.80
#> 178    104710 16.0  62.32444   1 1897 -0.43
#> 179    104980 16.0 123.66325   2 3764 -2.31
#> 180    105360 16.5 113.24846   1 3447 -1.78
#> 181    105470 13.5 108.41889   1 3300 -4.34
#> 182    105870 15.0  69.81519   1 2125 -1.43
#> 183    105880 17.0 104.31212   2 3175 -1.09
#> 184    106400 14.0  67.84394   1 2065 -2.26
#> 185    106530 16.0 129.08418   2 3929 -2.52
#> 186    106640 13.0  61.83162   1 1882 -3.05
#> 187    106700 17.0  65.37987   2 1990 -0.04
#> 188    106950 14.5 101.25668   2 3082 -2.65
#> 189    107000 16.8  83.31827   1 2536 -0.55
#> 190    107610 14.5  65.21561   1 1985 -1.71
#> 191    107780 14.0  71.12936   2 2165 -2.31
#> 192    107810 18.3 133.61807   1 4067 -1.32
#> 193    107890 15.0  97.08419   1 2955 -2.53
#> 194    108100 15.8  85.12526   2 2591 -1.31
#> 195    108220 14.5  88.27926   2 2687 -2.31
#> 196    108310 13.0  83.05544   2 2528 -3.41
#> 197    108680 16.3 118.24230   1 3599 -2.12
#> 198    108810 17.0 123.17043   2 3749 -1.68
#> 199    109160 12.9  66.26694   1 2017 -3.31
#> 200    109180 22.0 119.06365   1 3624  0.83
#> 201    109190 13.4  61.96304   1 1886 -2.66
#> 202    109230 16.0  73.33060   2 2232 -0.88
#> 203    109280 17.0 150.14374   2 4570 -2.80
#> 204    109440 17.0  64.19713   1 1954  0.22
#> 205    109460 14.1  94.39014   2 2873 -2.78
#> 206    109480 16.2  91.53183   2 2786 -1.21
#> 207    109510 14.5  94.02875   1 2862 -2.92
#> 208    110100 13.8 101.55236   2 3091 -3.20
#> 209    110170 16.0  61.14169   1 1861 -0.40
#> 210    110190 15.5  82.72690   1 2518 -1.51
#> 211    110210 14.0  66.75975   2 2032 -2.22
#> 212    110270 17.5 125.99590   2 3835 -1.50
#> 213    110350 15.4 115.81109   2 3525 -2.44
#> 214    110370 14.0  70.37372   1 2142 -2.36
#> 215    110450 16.0  94.29158   1 2870 -1.54
#> 216    110460 15.4  66.36550   1 2020 -0.99
#> 217    110500 11.5  80.36140   2 2446 -4.59
#> 218    110610 16.0  60.61602   1 1845 -0.39
#> 219    110660 15.0  90.18481   1 2745 -2.25
#> 220    110680 14.0  79.90144   2 2432 -2.51
#> 221    110790 15.0  79.93429   1 2433 -1.83
#> 222    110840 16.5  94.25873   2 2869 -1.10
#> 223    110860 15.3  76.02464   2 2314 -1.43
#> 224    110920 17.1 112.06571   1 3411 -1.31
#> 225    110950 15.5  90.84189   1 2765 -1.83
#> 226    111120 16.0  83.41684   1 2539 -1.13
#> 227    111210 15.0  64.00000   1 1948 -1.25
#> 228    111500 18.0 135.12936   1 4113 -1.56
#> 229    111530 15.0  70.04517   2 2132 -1.51
#> 230    111790 17.9  98.85832   1 3009 -0.38
#> 231    112610 16.0 129.31416   2 3936 -2.53
#> 232    112820 16.5 106.71047   1 3248 -1.56
#> 233    112830 15.5  97.74127   2 2975 -1.83
#> 234    112950 16.0  64.85421   1 1974 -0.49
#> 235    112960 16.0  60.05750   1 1828 -0.37
#> 236    113170 12.7  64.68994   2 1969 -3.31
#> 237    113300 14.5  69.45380   1 2114 -1.85
#> 238    113360 16.5  91.00616   1 2770 -1.04
#> 239    113860 16.0  71.72074   1 2183 -0.70
#> 240    113920 14.0  83.84394   1 2552 -2.99
#> 241    114000 18.0 121.95483   1 3712 -1.06
#> 242    114210 16.4  66.29980   2 2018 -0.44
#> 243    114220 16.0  80.22998   1 2442 -1.01
#> 244    114730 17.2 104.60780   1 3184 -1.00
#> 245    114770 14.0  74.54620   1 2269 -2.54
#> 246    114880 16.0  70.76797   2 2154 -0.82
#> 247    115110 12.0  89.06776   2 2711 -4.32
#> 248    115220 14.5  98.20123   2 2989 -2.56
#> 249    115340 16.2 140.45175   2 4275 -2.88
#> 250    115600 19.0  97.64271   2 2972  0.11
#> 251    115690 14.5  66.82546   1 2034 -1.76
#> 252    115830 13.5  73.49487   1 2237 -3.01
#> 253    115920 14.0  66.23408   1 2016 -2.20
#> 254    115930 15.0 105.06776   2 3198 -2.39
#> 255    115990 20.5 130.33264   1 3967 -0.09
#> 256    116310 14.6  89.16632   2 2714 -2.25
#> 257    116320 18.0 120.57495   2 3670 -1.06