wgsData
Source: R/zscorer.R

World Health Organization (WHO) Growth Reference (2006) data

wgsData

Format

A data frame with 6 columns and 1746 rows.

indicator

One of weight-for-age (waz), height-for-age (haz), or weight-for-height (whz) anthropometric indicators

sex

Sex of child (1 = Male; 2 = Female)

given

Variable to which standardisation is to be made. For waz and haz, given is age in months. For whz, given is height in cm

l

L component of the LMS method for normalising growth centile standards. L is the trend in the optimal power to obtain normality

m

M component of the LMS method for normalising growth centile standards. M is the trend in the mean

s

S component of the LMS method for normalising growth centile standards. S is the trend in the coefficient of variation

Source

World Health Organization. WHO Child Growth Standards: Length/Height-for-age, Weight-for-age, Weight-for-length, Weight-for-height, and Body Mass Index-for age: Methods and Development. 1st ed. World Health Organization; 2006.